Module Interfaces

Rimu.Interfaces — Module

module Interfaces

This module contains interfaces that can be used to extend and modify the algorithms and behaviours of Rimu.

Interfaces

Follow the links for the definitions of the interfaces!

AbstractHamiltonian for defining Hamiltonians

AbstractOperator for defining observable operators

AbstractObservable for defining observables

AbstractOperatorColumn for defining operator columns

AbstractDVec for defining data structures for Rimu as in DictVectors

StochasticStyle for controlling the stochastic algorithms used by ProjectorMonteCarloProblem as implemented in StochasticStyles

AbstractFockAddress for defining Fock states, see also BitStringAddresses.

Additional exports

Interface functions forAbstractHamiltonians:

starting_address

operator_column

parent_operator

diagonal_element

random_offdiagonal

offdiagonals.

num_offdiagonals

get_offdiagonal

LOStructure

allows_address_type

undo_transform

has_random_offdiagonal

has_iterable_offdiagonals

maximum_mode_occupation

working with AbstractDVecs and StochasticStyle

deposit!

default_style

CompressionStrategy

The interface from VectorInterface.jl.

Functions Rimu.jl uses to do FCIQMC:

apply_column!

apply_operator!

step_stats

Functions for retrieving information from DataFrames:

num_replicas

num_spectral_states

num_overlaps

Functions for working with AbstractFockAddresss:

num_particles

num_modes

num_components

maximum_mode_occupation