API

Rimu

Rimu.Rimu — Module

Rimu

Random Integrator for Many-Body Quantum Systems

Rimu.ConstantTimeStep — Type

Keep dτ constant.

Rimu.DelayedLogUpdate — Type

DelayedLogUpdate(ζ = 0.3, a = 10) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and delay of a steps.

\[S^{n+a} = S^n -\frac{ζ}{a dτ}\ln\left(\frac{\|Ψ\|_1^{n+a}}{\|Ψ\|_1^n}\right)\]

Rimu.DelayedLogUpdateAfterTargetWalkers — Type

DelayedLogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.3, a = 10) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and delay of a steps. See DelayedLogUpdate.

Rimu.DeltaMemory — Type

DeltaMemory(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = (pnorm - tnorm)/(dτ*pnorm) + shift
r = r̃ - <r̃>

Rimu.DontUpdate — Type

DontUpdate() <: ShiftStrategy Don't update the shift.

Rimu.DoubleLogUpdate — Type

DoubleLogUpdate(targetwalkers, ζ = 0.3, ξ = 0.0225) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^\text{target}}\right)\]

Rimu.DoubleLogUpdateAfterTargetWalkers — Type

LogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.3, ξ = 0.0225) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and ξ. See DoubleLogUpdate.

Rimu.DoubleLogUpdateAfterTargetWalkersSwitch — Type

LogUpdateAfterTargetWalkersSwitch(targetwalkers, ζ = 0.3, ξ = 0.0225) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and ξ. After a steps the strategy swiches to LogUpdate. See DoubleLogUpdate.

Rimu.EveryKthStep — Type

EveryKthStep(;k = 10)

Report every kth step.

Rimu.EveryTimeStep — Type

Report every time step.

Rimu.FciqmcRunStrategy — Type

Abstract type representing the strategy for running and terminating fciqmc!(). Implemented strategies:

RunTillLastStep

Rimu.HistoryLogUpdate — Type

HistoryLogUpdate(df::DataFrame; d = 100, k=1, ζ= 0.3)

Strategy for updating the shift according to log formula but with walker numbers accumulated from k samples of the history with delay d. A recent history has to be passed with the data frame df for initialisation.

\[N_w^{n} = \sum_{i=0}^{k-1} \|Ψ\|_1^{n-i} \\ S^{n+a} = S^n -\frac{ζ}{a dτ}\ln\left(\frac{N_w^{n+1}}{N_w^n}\right)\]

Rimu.IsStochasticWithThreshold — Type

IsStochasticWithThreshold(threshold::Float16)

Trait for generalised vector of configurations indicating stochastic propagation with real walker numbers and cutoff threshold.

> StochasticStyle(V) = IsStochasticWithThreshold(threshold)

During stochastic propoagation, walker numbers small than threshold will be stochastically projected to either zero or threshold.

Rimu.LogUpdate — Type

LogUpdate(ζ = 0.3) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)\]

Rimu.LogUpdateAfterTargetWalkers — Type

LogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.3) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ. See LogUpdate.

Rimu.MPIData — Type

Simple wrapper used for signaling that this data is part of a distributed data structure and communication should happen with MPI.

Rimu.MPIDefault — Type

MPIDefault(nprocs, my_rank, comm)

Strategy for point-to-point MPI communication.

Rimu.MPIOSWin — Type

MPIOSWin(nprocs, myrank, comm, ::Type{T}, capacity)

Communication buffer for use with MPI one-sided communication (remote memory access). Up to capacity elements of type T can be exchanged between MPI ranks via put. It is important that isbitstype(T) == true. Objects of type MPIOSWin have to be freed manually with a (blocking) call to free().

Rimu.MemoryStrategy — Type

Abstract type for defining the strategy for injectimg memory noise. Implemented strategies:

NoMemory

DeltaMemory

ShiftMemory

Rimu.NoMemory — Type

NoMemory <: MemoryStrategy

Default strategy for MemoryStrategy indicating that no memory noise will be used.

Rimu.NoProjection — Type

Do not project the walker amplitudes. See norm_project.

Rimu.OvershootControl — Type

Slow down/Speed up dτ to control the psips overshoot.

Rimu.ProjectStrategy — Type

Abstract type for defining the stategy of projection for fciqmc with floating point walker number with norm_project. Implemented stategies:

NoProjection

ThresholdProject

ScaledThresholdProject

Rimu.ReportDFAndInfo — Type

ReportDFAndInfo(; k=10, i=100, io=stdout, writeinfo=true)

Report every kth step in DataFrame and write info message to io every ith step (unless writeinfo == false). The flag writeinfo is useful for controlling info messages in MPI codes.

Rimu.ReportingStrategy — Type

Abstract type for strategies for reporting data in a DataFrame with report!(). Implemented strategies:

EveryTimeStep

EveryKthStep

ReportDFAndInfo

Rimu.RunTillLastStep — Type

RunTillLastStep(step::Int = 0 # number of current/starting timestep
             laststep::Int = 50 # number of final timestep
             shiftMode::Bool = false # whether to adjust shift
             shift::Float64 = 0.0 # starting/current value of shift
             dτ::Float64 = 0.01 # current value of time step
) <: FciqmcRunStrategy

Parameters for running fciqmc!() for a fixed number of time steps. For alternative strategies, see FciqmcRunStrategy.

Rimu.ScaledThresholdProject — Type

ScaledThresholdProject(threshold = 1.0) <: ProjectStrategy

Project stochastically for walker amplitudes below threshold and scale configuration array as to keep the norm constant. As a consequence, the final configuration amplitudes may be smaller than threshold. See norm_project.

Rimu.ShiftMemory — Type

ShiftMemory(Δ::Int) <: MemoryStrategy

Effectively replaces the fluctuating shift update procedure for the coefficient vector by an averaged shift over Δ timesteps, where Δ = 1 means no averaging.

Rimu.StochasticStyle — Type

StochasticStyle(v)
StochasticStyle(typeof(v))

StochasticStyle specifies the native style of the generalised vector v that determines how simulations are to proceed. This can be fully stochastic (with IsStochastic), fully deterministic (with IsDeterministic), or semistochastic (with IsSemistochastic).

Rimu.ThresholdProject — Type

ThresholdProject(threshold = 1.0) <: ProjectStrategy

Project stochastically for walker amplitudes below threshold. See norm_project.

Rimu.TimeStepStrategy — Type

Abstract type for strategies for updating the time step with update_dτ(). Implemented strategies:

ConstantTimeStep

Rimu.fciqmc! — Function

fciqmc!(v, pa::FciqmcRunStrategy, [df,]
         ham, s_strat::ShiftStrategy,
         [r_strat::ReportingStrategy, τ_strat::TimeStepStrategy, w])
-> df

Perform the FCIQMC algorithm for determining the lowest eigenvalue of ham. v can be a single starting vector of type :<AbstractDVec or a vector of such structures. In the latter case, independent replicas are constructed. Returns a DataFrame df with statistics about the run, or a tuple of DataFrames for a replica run. Strategies can be given for updating the shift (see ShiftStrategy) and (optionally), for reporting (see ReportingStrategy), and for updating the time step dτ (see TimeStepStrategy).

A pre-allocated data structure w for working memory can be passed as argument, and controls multi-threading behaviour. By default multi-threading is turned on. To turn multi-threading off, pass similar(localpart(v)) for w.

This function mutates v, the parameter struct pa as well as df, and w.

Rimu.free — Method

free(obj::MPIOSWin)

De-reference the object, call finalizer and the garbage collector immediately. This is a syncronizing MPI call. Make sure that the object is not used later.

Rimu.greet — Method

brief greeting

Rimu.mpi_default — Function

mpi_default(data, comm = MPI.COMM_WORLD, root = 0)

Declare data as mpi-distributed and set communication strategy to default. Sets up the MPIData structure with MPIDefault strategy.

Rimu.mpi_no_exchange — Function

mpi_no_exchange(data, comm = MPI.COMM_WORLD, root = 0)

Declare data as mpi-distributed and set communication strategy to MPINoWalkerExchange. Sets up the MPIData structure with MPINoWalkerExchange strategy.

Rimu.mpi_one_sided — Function

mpi_one_sided(data, comm = MPI.COMM_WORLD, root = 0)

Declare data as mpi-distributed and set communication strategy to one-sided with remote memory access (RMA). Sets up the MPIData structure with MPIOSWin strategy.

Rimu.put — Method

put(buf::Vector{T}, [len,] targetrank, s::MPIOSWin{T})
put(obj::T, targetrank, s::MPIOSWin{T})

Deposit a single obj or vector buf into the MPI window s on rank targetrank. If len is given, only the first len elements are transmitted.

Rimu.targetrank — Function

targetrank(key, np, hash = hash(key))

Compute the rank where the key belongs.

Rimu.DeltaMemory2 — Type

DeltaMemory2(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = pnorm - tnorm + shift*dτ*pnorm
r = (r̃ - <r̃>)/(dτ*pnorm)

The long-term average of r is not guaranteed to be zero.

Rimu.IsSemistochastic — Type

IsSemistochastic(threshold::Float16, d_space)

Trait for generalised vector of configurations indicating semistochastic propagation. Set with setSemistochastic!.

> StochasticStyle(V) = IsSemistochastic(threshold, d_space)

where d_space is a vector of addresses defining the the stochastic subspace.

Rimu.MPINoWalkerExchange — Type

MPINoWalkerExchange(nprocs, my_rank, comm)

Strategy for for not exchanging walkers between ranks. Consequently there will be no cross-rank annihilations.

Rimu.ShiftStrategy — Type

Abstract type for defining the strategy for updating the shift with update_shift(). Implemented strategies:

DontUpdate

LogUpdate

DelayedLogUpdate

LogUpdateAfterTargetWalkers

DelayedLogUpdateAfterTargetWalkers

HistoryLogUpdate

Base.length — Method

length(md::MPIData)

Compute the length of the distributed data on every MPI rank with MPI.Allreduce.

LinearAlgebra.norm — Function

norm(md::MPIData, p=2)

Compute the norm of the distributed data on every MPI rank with MPI.Allreduce.

Rimu.applyMemoryNoise! — Method

applyMemoryNoise!(w, v, shift, dτ, pnorm, m_strat::MemoryStrategy)

Apply memory noise to w according to the strategy m_strat. Note that the strategy needs to be compatible with StochasticStyle(w). The default is to not add memory noise. See MemoryStrategy.

w is the walker array after fciqmc step, v the previous one, pnorm the norm of v.

Rimu.clearDFlags! — Method

clearDFlags!(dv)

Clear all flags in dv of the deterministic bit (rightmost bit).

Rimu.fciqmc_col! — Method

fciqmc_col!(w, ham, add, num, shift, dτ)
fciqmc_col!(::Type{T}, args...)
-> spawns, deaths, clones, antiparticles, annihilations

Spawning and diagonal step of FCIQMC for single column of ham. In essence it computes

w .+= (1 .+ dτ.*(shift .- ham[:,add])).*num.

Depending on T ==StochasticStyle(w), a stochastic or deterministic algorithm will be chosen. The possible values for T are:

IsDeterministic() deteministic algorithm

IsStochastic() stochastic version where the changes added to w are purely integer, according to the FCIQMC algorithm

IsStochasticNonlinear(c) stochastic algorithm with nonlinear diagonal

IsSemistochastic() semistochastic version: TODO

Rimu.fciqmc_step! — Method

fciqmc_step!(Ĥ, v, shift, dτ, pnorm, w;
                      m_strat::MemoryStrategy = NoMemory()) -> ṽ, w̃, stats

Perform a single matrix(/operator)-vector multiplication:

\[\tilde{v} = [1 - dτ(\hat{H} - S)]⋅v ,\]

where Ĥ == ham and S == shift. Whether the operation is performed in stochastic, semistochastic, or determistic way is controlled by the trait StochasticStyle(w). See StochasticStyle. w is a local data structure with the same size and type as v and used for working. Both v and w are modified.

Returns the result ṽ, a (possibly changed) reference to working memory w̃, and the array stats = [spawns, deaths, clones, antiparticles, annihilations]. Stats will contain zeros when running in deterministic mode.

Rimu.next_mpiID — Function

next_mpiID()

Produce a new ID number for MPI distributed objects. Uses an internal counter.

Rimu.norm_project! — Method

norm_project!(w, p_strat::ProjectStrategy) -> norm

Computes the 1-norm of w. Project all elements of w to s.threshold preserving the sign if StochasticStyle(w) requires projection according to p_strat. See ProjectStrategy.

Rimu.report! — Method

report!(df::DataFrame, t::Tuple, s<:ReportingStrategy)

Record results in df and write informational messages according to strategy s. See ReportingStrategy.

Rimu.setSemistochastic! — Method

setSemistochastic!(dv, threshold::Float16, d_space)

Set the deterministic space for dv with threshold threshold, where d_space is a vector of addresses defining the the stochastic subspace.

Rimu.sort_into_targets! — Method

sort_into_targets!(target, source, stats) -> agg, wm, agg_stats

Aggregate coefficients from source to agg and from stats to agg_stats according to thread- or MPI-level parallelism. wm passes back a reference to working memory.

Rimu.sort_into_targets! — Method

sort_into_targets!(target::MPIData, source)

Distribute the entries of source to the target data structure such that all entries in the target dictionaries are on the process with the correct rank as controlled by targetrank().

Rimu.update_dτ — Method

update_dτ(s<:TimeStepStrategy, dτ, args...) -> new dτ

Update the time step according to the strategy s.

Rimu.update_shift — Method

update_shift(s <: ShiftStrategy, shift, shiftMode, tnorm, pnorm, dτ, step, df)

Update the shift according to strategy s. See ShiftStrategy.

Reexported Submodules

Hamiltonians

Link to Module Hamiltionians.jl

BitStringAddresses

Rimu.BitStringAddresses.BSAdd128 — Type

BSAdd128 <: BitStringAddressType

Address type that encodes a bistring address in a UInt128.

Rimu.BitStringAddresses.BSAdd64 — Type

BSAdd64 <: BitStringAddressType

Address type that encodes a bistring address in a UInt64.

Rimu.BitStringAddresses.BStringAdd — Type

BStringAdd <: BitStringAddressType

Address type that encodes a bistring address in a one-dim. BitArray.

Rimu.BitStringAddresses.BitAdd — Type

BitAdd{I,B} <: BitStringAddressType
BitAdd(address::Integer, B)
BitAdd(chunks::T, B) where T<:Union{Tuple,SVector}
BitAdd{B}(address)

Address type that encodes a bistring address with B bits. The bits are stored efficiently as SVector of I chunks of type UInt64. The two-argument constructor is preferred due to safety (consistency checks). In hot loops there maybe gain from the (unsafe) parametric constructor. If an integer address is passed, its bit representation is used to initialize BitAdd. For large bit numbers, BigInt is convenient. BitAdd[i] will return bit i (as Bool), counting from right to left.

BitAdd{B}() creates a BitAdd with all ones.

zero(BitAdd{B}) creates a BitAdd with all zeros.

Note that no checking for ghost bits occurs when constructing BitAdd from SVector or Tuple. See bitadd(), check_consistency(), and remove_ghost_bits() methods!

Rimu.BitStringAddresses.BitStringAddressType — Type

BitStringAddressType

Abstract type for configuration addresses with the nature of a bitstring. A number of methods need to be implemented, in particular Base.isless(a,b), numBits(), numChunks().

Rimu.BitStringAddresses.BoseFS — Type

BoseFS{N,M,A} <: BosonicFockStateAddress <: BitStringAddressType
BoseFS(bs::A) where A <: BitAdd
BoseFS(bs::A, b)

Address type that represents a Fock state of N spinless bosons in M orbitals by wrapping a bitstring of type A. Orbitals are stored in reverse order, i.e. the first orbital in a BoseFS is stored rightmost in the bitstring bs. If the number of significant bits b is not encoded in A it must be passed as an argument (e.g. for BSAdd64 and BSAdd128).

Rimu.BitStringAddresses.BoseFS — Method

BoseFS(onr::T) where T<:Union{AbstractVector,Tuple}
BoseFS{BST}(onr::T)

Create BoseFS address from an occupation number representation, specifying the occupation number of each orbital. If a type BST is given it will define the underlying bit string type. Otherwise, the bit string type is chosen to fit the onr.

Rimu.BitStringAddresses.bitaddr — Method

bitaddr(onr, Type)

Calculate a bitstring address from an occupation number representation the type of the bitstring address is passed as the second argument.

Rimu.BitStringAddresses.nearUniform — Method

nearUniform(N, M) -> onr::SVector{M,Int}

Create occupation number representation onr distributing N particles in M modes in a close-to-uniform fashion with each orbital filled with at least N ÷ M particles and at most with N ÷ M + 1 particles.

Rimu.BitStringAddresses.nearUniform — Method

nearUniform(BoseFS{N,M})
nearUniform(BoseFS{N,M,A}) -> bfs::BoseFS{N,M,A}

Create bosonic Fock state with near uniform occupation number of M modes with a total of N particles. Specifying the bit address type A is optional.

Examples

julia> nearUniform(BoseFS{7,5,BitAdd})
BoseFS{BitAdd}((2,2,1,1,1))

julia> nearUniform(BoseFS{7,5})
BoseFS{BSAdd64}((2,2,1,1,1))

Rimu.BitStringAddresses.numBits — Method

numBits(a)

Number of bit chunks representing a.

Rimu.BitStringAddresses.numChunks — Method

numChunks(a)

Number of 64-bit chunks representing a.

Rimu.BitStringAddresses.numModes — Method

numModes(a)

Number of modes represented by a.

Rimu.BitStringAddresses.numParticles — Method

numParticles(a)

Number of particles represented by a.

Rimu.BitStringAddresses.onr — Method

onr(address, m)

Compute and return the occupation number representation as an array of Int corresponding to the given address.

Rimu.BitStringAddresses.onr — Method

onr(bs)

Compute and return the occupation number representation of the bit string address bs as an SVector{M,Int}, where M is the number of orbitals.

Rimu.BitStringAddresses.BosonicFockStateAddress — Type

BosonicFockStateAddress <: BitStringAddressType

Supertype representing a bosonic Fock state. Implement numModes() and numParticles().

Base.:<< — Method

<<(b::BitAdd,n::Integer)

Bitshift b to the left by n bits and fill from the right with zeros.

Base.:>>> — Method

>>>(b::BitAdd,n::Integer)

Bitshift b to the right by n bits and fill from the left with zeros.

Rimu.BitStringAddresses.lbshr — Method

lbshr(c,k)

Apply logical bit shift to the right by k bits to c.

Rimu.BitStringAddresses.remove_ghost_bits — Method

remove_ghost_bits(bs)

Remove set bits outside data field if any are present.

DictVectors

Rimu.DictVectors — Module

Module that provides data structures that behave similar to sparse vectors, but are indexed by arbitrary types (could be non-integers) similarly to dictionaries. The idea is to do linear algebra with data structures that are neither subtyped to AbstractVector nor to AbstractDict and are suitable for use with KrylovKit.jl. For this, the abstract type and interface AbstractDVec is provided, with the following concrete subtypes:

DVec

DFVec

FastDVec

Rimu.DictVectors.AbstractDVec — Type

DictVectors.AbstractDVec{K,V}

Abstract type for sparse vectors with valtype() V based on dictionary-like structures. They behave like vectors except that indexing is performed with an arbitrary keytype() K (no order required). getindex() returns zero(V) for any unknown key and zeroed elements should be deleted with delete!(). Thus, iteration typically returns only non-zero entries. In variance to AbstractVector, length() will return the number of non-zero elements, while capacity() returns the holding capacity (assumed fixed) of the object. Even though AbstractDVec is similar to both AbstractDict and AbstractVector, it is explicitly not subtyped to either. The aim is to create data structures suitable for FCIQMC and compatible with KrylovKit.jl.

Interface

The interface is similar to the AbstractDict interface. Implement what would be needed for the AbstractDict interface (setindex!, getindex, delete!, length, haskey, empty!, isempty) and, in addition:

capacity(dv): holding capacity

similar(dv [,Type])

iterate(): should return values of type V

pairs(): should return an iterator over key::K => content pairs. If content ≠ value::V the provide values() iterator as well!

Rimu.DictVectors.DFVec — Type

DFVec{K,V,F}(capacity) <: AbstractDVec{K,V <: Number}
DFVec(d::Dict [, capacity])
DFVec(v::Vector{V} [, capacity])

Construct a wrapped dictionary with minimum capacity capacity to represent a vector-like object with valtype(dv) == V. The value of the Dict are of type Tuple{V,F}, which allows for storing a flag of type F for each entry. Indexing is done with an arbitrary (in general non-integer) keytype(dv) == K. When constructed from a Vector{V}, the keys will be integers ∈ [1, length(v)] and the flag zero(UInt16). See AbstractDVec. The method capacity() is defined but not a strict upper limit as Dict objects can expand.

Rimu.DictVectors.DVec — Type

DVec{K,T}(capacity) <: AbstractDVec{K,T}
DVec(d::Dict [, capacity])
DVec(v::Vector{T} [, capacity])

Construct a wrapped dictionary with minimum capacity capacity to represent a vector-like object with eltype(dv) == T, which corresponds to the values of the Dict. Indexing is done with an arbitrary (in general non-integer) keytype(dv) == K. When constructed from a Vector, the keys will be integers ∈ [0, length(v)]. See AbstractDVec. The method capacity() is defined but not a strict upper limit as Dict objects can expand.

Rimu.DictVectors.FastDVec — Type

DictVectors.FastDVec{K,V}(capacity)

Create a dictionary-like array indexed by keys of type K and values of type V with a maximum capacity of capacity.

FastDVec(a::AbstractArray, [capacity = length(a)])
FastDVec(d::AbstractDict{K,V}, [capacity = length(d)])
FastDVec(d::AbstractDVec{K,V}, [capacity = length(d)])

Construct a FastDVec object from an existing array or dictionary.

Rimu.DictVectors.NormProjector — Type

NormProjector()

Results in computing the one-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,1) # with type valtype(x)

NormProjector() thus represents the vector sign.(x).

Rimu.DictVectors.UniformProjector — Type

UniformProjector()

Represents a vector with all elements 1. To be used with dot(). Minimizes memory allocations.

UniformProjector()⋅v == sum(v)
dot(UniformProjector(), LO, v) == sum(LO*v)

Rimu.DictVectors.add! — Method

add!(x::AbstractDVec,y::AbstactDVec)

Inplace add x+y and store result in x.

Rimu.DictVectors.capacity — Function

DictVectors.capacity(dv::AbstractDVec, [s = :effective])
capacity(dvs::Tuple, [s = :effective])

gives the effective holding capacity of dv. If a tuple of dvs is given it aggregates the capacities.

Optional argument s:

:effective the number of elements that

can be stored savely (default)

:allocated actual internal memory allocation

Rimu.DictVectors.flags — Method

flags(dv::DFVec)

An iterator that yields flags stored in the DFVec dv.

Rimu.DictVectors.gettuple — Method

gettuple(dv::DFVec, key)

Retrieve the tuple (v,f) stored at the given key. If the key is not found, return a tuple of zeros.

Rimu.DictVectors.kvpairs — Method

kvpairs(collection)

Return an iterator over key => value pairs ignoring any flags. If no flags are present, eg. for generic AbstractDVec, this falls back to Base.pairs.

Rimu.DictVectors.kvpairs — Method

kvpairs(dv::DFVec)

An iterator that yields key => value pairs stored in the DFVec dv ignoring any flags. In contrast, Base.pairs() will return pairs key => (value, flag).

Rimu.DictVectors.pairtype — Method

pairtype(dv)

Returns the type of stored data, as returned by the pairs() iterator.

Rimu.DictVectors.tuples — Method

tuples(dv::DFVec)

An iterator that yields tuples (v, f) with value v and flag f stored in the DFVec dv.

Rimu.DictVectors.zero! — Method

zero!(v::AbstractDVec)

Replace v by a zero vector as an inplace operation. For AbstractDVec types it means removing all non-zero elements.

Rimu.DictVectors.ADVPairsIterator — Type

ADVPairsIterator

Iterator type for pairs from a AbstractDVec.

Rimu.DictVectors.FlagsIterator — Type

FlagsIterator

Iterator type returned by flags().

Rimu.DictVectors.KVPairsIterator — Type

KVPairsIterator

Iterator type returned by kvpairs().

Base.fill! — Method

fill!(da::AbstractDVec, x)

Empties da if x==zero(valtype(da)) and throws an error otherwise.

Base.get — Method

get(dv::DFVec, key, deftup::Tuple)

Return tuple containing value and flag if key exists and otherwise deftup.

Base.get — Method

get(dv::DFVec, key, defnum::Number)

Return value only if key exists and otherwise defnum

Base.isequal — Method

isequal(l::AbstractDVec, r::AbstractDVec)

Returns true if all non-zero entries have the same value. Equality of flags is not tested unless both l and r support flags.

Base.isequal — Method

isequal(l::DFVec, r::DFVec)

Returns true if all non-zero entries have the same value and the same flag.

LinearAlgebra.axpy! — Method

axpy!(α::Number, X::AbstractDVec, Y::AbstractDVec)
axpy!(α::Number, X::AbstractDVec, Ys::Tuple, batchsize)

Overwrite Y with α*X + Y where α is scalar for AbstractDVecs. If a tuple Ys is passed with Threads.nthreads() AbstractDVecs, then perform the operation in parallel over threads with batchsize elements at a time.

LinearAlgebra.norm — Method

norm(x::AbstractDVec, p)

Computes the p-norm of the DVec x. Implemented for p ∈ {1, 2, Inf}. Returns zero if x is empty.

LinearAlgebra.norm — Method

norm(x::AbstractDVec{K,V})

Computes the 2-norm of the DVec x.

Rimu.DictVectors.normInf — Method

normInf(x::AbstractDVec)

Infinity norm: largest absolute value of entries.

Rimu.DictVectors.norm_sqr — Method

norm_sqr(x::AbstractDVec)

Fast calculation of the square of the 2-norm of x.

Blocking

Rimu.Blocking — Module

Blocking

Module that contains functions performing the Flyvbjerg-Petersen blocking analysis for evaluating the standard error on a correlated data set.

Rimu.Blocking.blocker — Method

Reblock the data by successively taking the mean of adjacent data points

Rimu.Blocking.blocking — Method

perform a blocking analysis for two data sets

Rimu.Blocking.blocking — Method

perform a blocking analysis for single data set

Rimu.Blocking.blockingErrorEstimation — Method

estimating stnadard error from blocking analysis based on the overlapping of error bars, if all the error bars (or more than 3 on a roll) behind current one are overlapping with it, return the current standard error with error bar.

Rimu.Blocking.mtest — Method

The "M test" based on Jonsson, M. Physical Review E, 98(4), 043304, (2018). If the blocking analysis (BA) has passed the M test, an error estimation will be given based on the smallest k (i.e. meaningful results at the k-th data point on a BA plot).

Rimu.Blocking.combination_division — Method

find the standard error on standard errors on two datasets

Rimu.Blocking.covariance — Method

Calculate the covariance between the two data sets vi and vj.

Rimu.Blocking.sd — Method

Calculate the standard deviation of the dataset v

Rimu.Blocking.se — Method

Calculate the standard error of the dataset v

Rimu.Blocking.variance — Method

Calculate the variance of the dataset v

Unexported Submodules

FastBufs

Rimu.FastBufs — Module

Provides the FastBuf data structure.

Rimu.FastBufs.FastBuf — Type

FastBuf{T}(n)

Create a first-in-last-out buffer of maximum capacity n. The element type is T. Fill a FastBuf with push!(), retrieve with pop!(). Accessing elements with getindex() and setindex!() as well as iteration are also supported. Check the capacity of a FastBuf with capacity().

Rimu.FastBufs.capacity — Method

Gives the capacity of a FastBuf object.

ConsistentRNG

Link to Module ConsistentRNG.jl

Index

Rimu.Blocking

Rimu.ConsistentRNG

Rimu.DictVectors

Rimu.FastBufs

Rimu.Hamiltonians

Rimu.Rimu

Rimu.BitStringAddresses.BSAdd128

Rimu.BitStringAddresses.BSAdd64

Rimu.BitStringAddresses.BStringAdd

Rimu.BitStringAddresses.BitAdd

Rimu.BitStringAddresses.BitStringAddressType

Rimu.BitStringAddresses.BoseFS

Rimu.BitStringAddresses.BoseFS

Rimu.BitStringAddresses.BosonicFockStateAddress

Rimu.ConsistentRNG.CRNG

Rimu.ConstantTimeStep

Rimu.DelayedLogUpdate

Rimu.DelayedLogUpdateAfterTargetWalkers

Rimu.DeltaMemory

Rimu.DeltaMemory2

Rimu.DictVectors.ADVPairsIterator

Rimu.DictVectors.AbstractDVec

Rimu.DictVectors.DFVec

Rimu.DictVectors.DVec

Rimu.DictVectors.FastDVec

Rimu.DictVectors.FlagsIterator

Rimu.DictVectors.KVPairsIterator

Rimu.DictVectors.NormProjector

Rimu.DictVectors.UniformProjector

Rimu.DontUpdate

Rimu.DoubleLogUpdate

Rimu.DoubleLogUpdateAfterTargetWalkers

Rimu.DoubleLogUpdateAfterTargetWalkersSwitch

Rimu.EveryKthStep

Rimu.EveryTimeStep

Rimu.FastBufs.FastBuf

Rimu.FciqmcRunStrategy

Rimu.Hamiltonians.BoseHubbardReal1D

Rimu.Hamiltonians.BosonicHamiltonian

Rimu.Hamiltonians.ExtendedBHReal1D

Rimu.Hamiltonians.Hops

Rimu.Hamiltonians.LinearOperator

Rimu.HistoryLogUpdate

Rimu.IsSemistochastic

Rimu.IsStochasticWithThreshold

Rimu.LogUpdate

Rimu.LogUpdateAfterTargetWalkers

Rimu.MPIData

Rimu.MPIDefault

Rimu.MPINoWalkerExchange

Rimu.MPIOSWin

Rimu.MemoryStrategy

Rimu.NoMemory

Rimu.NoProjection

Rimu.OvershootControl

Rimu.ProjectStrategy

Rimu.ReportDFAndInfo

Rimu.ReportingStrategy

Rimu.RunTillLastStep

Rimu.ScaledThresholdProject

Rimu.ShiftMemory

Rimu.ShiftStrategy

Rimu.StochasticStyle

Rimu.ThresholdProject

Rimu.TimeStepStrategy

Base.:<<

Base.:>>>

Base.fill!

Base.get

Base.get

Base.isequal

Base.isequal

Base.length

LinearAlgebra.axpy!

LinearAlgebra.norm

LinearAlgebra.norm

LinearAlgebra.norm

Rimu.BitStringAddresses.bitaddr

Rimu.BitStringAddresses.lbshr

Rimu.BitStringAddresses.nearUniform

Rimu.BitStringAddresses.nearUniform

Rimu.BitStringAddresses.nearUniform

Rimu.BitStringAddresses.numBits

Rimu.BitStringAddresses.numChunks

Rimu.BitStringAddresses.numModes

Rimu.BitStringAddresses.numParticles

Rimu.BitStringAddresses.onr

Rimu.BitStringAddresses.onr

Rimu.BitStringAddresses.remove_ghost_bits

Rimu.Blocking.blocker

Rimu.Blocking.blocking

Rimu.Blocking.blocking

Rimu.Blocking.blockingErrorEstimation

Rimu.Blocking.combination_division

Rimu.Blocking.covariance

Rimu.Blocking.mtest

Rimu.Blocking.sd

Rimu.Blocking.se

Rimu.Blocking.variance

Rimu.ConsistentRNG.cRand

Rimu.ConsistentRNG.newChildRNG

Rimu.ConsistentRNG.seedCRNG!

Rimu.ConsistentRNG.trng

Rimu.DictVectors.add!

Rimu.DictVectors.capacity

Rimu.DictVectors.flags

Rimu.DictVectors.gettuple

Rimu.DictVectors.kvpairs

Rimu.DictVectors.kvpairs

Rimu.DictVectors.normInf

Rimu.DictVectors.norm_sqr

Rimu.DictVectors.pairtype

Rimu.DictVectors.tuples

Rimu.DictVectors.zero!

Rimu.FastBufs.capacity

Rimu.Hamiltonians.bit_String_Length

Rimu.Hamiltonians.diagME

Rimu.Hamiltonians.dimensionLO

Rimu.Hamiltonians.fDimensionLO

Rimu.Hamiltonians.generateRandHop

Rimu.Hamiltonians.hasIntDimension

Rimu.Hamiltonians.hop

Rimu.Hamiltonians.numOfHops

Rimu.applyMemoryNoise!

Rimu.clearDFlags!

Rimu.fciqmc!

Rimu.fciqmc_col!

Rimu.fciqmc_step!

Rimu.free

Rimu.greet

Rimu.mpi_default

Rimu.mpi_no_exchange

Rimu.mpi_one_sided

Rimu.next_mpiID

Rimu.norm_project!

Rimu.put

Rimu.report!

Rimu.setSemistochastic!

Rimu.sort_into_targets!

Rimu.sort_into_targets!

Rimu.targetrank

Rimu.update_dτ

Rimu.update_shift