API

Rimu

Rimu.Rimu — Module

Rimu

Random Integrator for Many-Body Quantum Systems

Rimu.AllOverlaps — Type

AllOverlaps(n=2, operator=nothing) <: ReplicaStrategy{n}

Run n replicas and report overlaps between all pairs of replica vectors. If operator is not nothing, the overlap dot(c1, operator, c2) is reported as well. If operator is a tuple of operators, the overlaps are computed for all operators.

Column names in the report are of the form c{i}dotc{j} for vector-vector overlaps, and c{i}Op{k}c{j} for operator overlaps.

See ReplicaStrategy and AbstractHamiltonian (for an interface for implementing operators).

Rimu.ConstantTimeStep — Type

ConstantTimeStep <: TimeStepStrategy

Keep dτ constant.

Rimu.DeltaMemory — Type

DeltaMemory(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = (pnorm - tnorm)/(dτ*pnorm) + shift
r = r̃ - <r̃>

Rimu.DeltaMemory2 — Type

DeltaMemory2(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = pnorm - tnorm + shift*dτ*pnorm
r = (r̃ - <r̃>)/(dτ*pnorm)

The long-term average of r is not guaranteed to be zero.

Rimu.DeltaMemory3 — Type

DeltaMemory3(Δ::Int, level::Float64) <: MemoryStrategy

Before updating the shift, apply multiplicative memory noise with a memory length of Δ at level level, where Δ = 1 means no memory noise.

r̃ = (pnorm - tnorm)/pnorm + dτ*shift
r = r̃ - <r̃>
w .*= 1 + level*r

Rimu.DontUpdate — Type

DontUpdate(; targetwalkers = 1_000_000) <: ShiftStrategy

Don't update the shift. Return when targetwalkers is reached.

Rimu.DoubleLogProjected — Type

DoubleLogProjected(; target, projector, ζ = 0.08, ξ = ζ^2/4) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ after projecting onto projector.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{P⋅Ψ^{(n+1)}}{P⋅Ψ^{(n)}}\right)-\frac{ξ}{dτ}\ln\left(\frac{P⋅Ψ^{(n+1)}}{\text{target}}\right)\]

Rimu.DoubleLogSumUpdate — Type

DoubleLogSumUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, α = 1/2) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameters ζ and ξ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^n}\right) - \frac{ξ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^\text{target}}\right),\]

where $N_\mathrm{w} =$ (1-α)*walkernumber() + α*UniformProjector()⋅ψ computed with walkernumber() and UniformProjector(). When ξ = ζ^2/4 this corresponds to critical damping with a damping time scale T = 2/ζ.

Rimu.DoubleLogUpdate — Type

DoubleLogUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)-\frac{ξ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^\text{target}}\right)\]

When ξ = ζ^2/4 this corresponds to critical damping with a damping time scale T = 2/ζ.

Rimu.DoubleLogUpdateAfterTargetWalkers — Type

DoubleLogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08, ξ = 0.0016) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and ξ. See DoubleLogUpdate.

Rimu.FciqmcRunStrategy — Type

 FciqmcRunStrategy{T}

Abstract type representing the strategy for running and terminating fciqmc!(). The type parameter T is relevant for reporting the shift and the norm.

Implemented strategies:

RunTillLastStep

Rimu.LogUpdate — Type

LogUpdate(ζ = 0.08) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)\]

Rimu.LogUpdateAfterTargetWalkers — Type

LogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ. See LogUpdate.

Rimu.MemoryStrategy — Type

Abstract type for defining the strategy for injectimg memory noise. Implemented strategies:

NoMemory

DeltaMemory

ShiftMemory

Rimu.MultiScalar — Type

MultiScalar

Wrapper over a tuple that supports +, -, min, and max. Used with MPI communication because SVectors are treated as arrays by MPI.Allreduce and Tuples do not support scalar operations.

Example

Suppose you want to compute the sum of a vector dv and also get the number of positive elements it has in a single pass. You can use MultiScalar:

julia> dv = DVec(:a => 1, :b => -2, :c => 1);

julia> s, p = mapreduce(+, values(dv)) do v
    Rimu.MultiScalar(v, Int(sign(v) == 1))
end;

julia> s, p
(0, 2)

This will work with MPIData.

Note that only MultiScalars with the same types can be operated on. This is a feature, as it forces type stability.

Rimu.NoMemory — Type

NoMemory <: MemoryStrategy

Default strategy for MemoryStrategy indicating that no memory noise will be used.

Rimu.NoStats — Type

NoStats(N=1) <: ReplicaStrategy{N}

The default ReplicaStrategy. N replicas are run, but no statistics are collected.

Rimu.PostStepStrategy — Type

PostStepStrategy

Subtypes of PostStepStrategy can be used to perform arbitrary computation on a replica after an FCIQMC step is finished and report the results.

Implemented strategies:

ProjectedEnergy

Projector

SignCoherence

WalkerLoneliness

Note: A tuple of multiple strategies can be passed to lomc!. In that case, all reported column names must be distinct.

Interface:

A subtype of this type must implement post_step(::PostStepStrategy, ::ReplicaState).

Rimu.ProjectedEnergy — Type

ProjectedEnergy(hamiltonian, projector; hproj=:vproj, vproj=:vproj) <: PostStepStrategy

After every step, compute hproj = dot(projector, hamiltonian, dv) and vproj = dot(projector, dv), where dv is the instantaneous coefficient vector. projector can be an AbstractDVec, or an AbstractProjector.

Reports to columns hproj and vproj, which can be used to compute projective energy, e.g. with Rimu.StatsTools.ratio_of_means. The keyword arguments hproj and vproj can be used to change the names of these columns. This can be used to make the names unique when computing projected energies with different projectors in the same run.

Rimu.ProjectedMemory — Type

ProjectedMemory(Δ::Int, projector, pp::Number) <: MemoryStrategy
ProjectedMemory(Δ::Int, projector, v::AbstractDVec)

Before updating the shift, apply memory noise to minimize the fluctuations of the overlap of the coefficient vector with projector. Averaging over Δ time steps is applied, where Δ = 1 means no memory noise is applied. Use pp to initialise the value of the projection or pass v in order to initialise the projection with pp = projector.v.

r̃ = (projector⋅v - projector⋅w)/projector⋅v + dτ*shift
r = r̃ - <r̃>

where v is the coefficient vector before and w after applying a regular FCIQMC step.

Rimu.Projector — Type

Projector(name=projector) <: PostStepStrategy

After each step, compute dot(projector, dv) and report it in the DataFrame under name. projector can be an AbstractDVec, or an AbstractProjector.

Rimu.PurgeNegatives — Type

PurgeNegatives <: MemoryStrategy

Purge all negative sign walkers.

Rimu.QMCState — Type

QMCState

Holds all information needed to run FCIQMC, except the data frame. Holds a NTuple of ReplicaStates and various strategies that control the algorithm.

Rimu.ReplicaState — Type

ReplicaState(v, w, pnorm, params, id)

Struct that holds all information needed for an independent run of the algorithm.

Can be advanced a step forward with advance!.

Fields

hamiltonian: the model Hamiltonian.

v: vector.

w: working memory.

pnorm: previous walker number (see walkernumber).

params: the FCIQMCRunStrategy.

id: appended to reported columns.

Rimu.ReplicaStrategy — Type

ReplicaStrategy{N}

An abstract type that controles how lomc! uses replicas. A subtype of ReplicaStrategy{N} operates on N replicas and must implement the following function:

replica_stats(::ReplicaStrategy{N}, ::NTuple{N,ReplicaState}) - return a tuple of Strings or Symbols of replica statistic names and a tuple of the values. These will be reported to the DataFrame returned by lomc!

Concrete implementations:

NoStats: run (possibly one) replica(s), but don't report any additional info.

AllOverlaps: report overlaps between all pairs of replica vectors.

Rimu.Report — Type

Report

Internal structure that holds the temporary reported values. See report!.

Rimu.ReportDFAndInfo — Type

ReportDFAndInfo(; k=1, i=100, io=stdout, writeinfo=false) <: ReportingStrategy

The default ReportingStrategy. Report every kth step to a DataFrame and write info message to io every ith step (unless writeinfo == false). The flag writeinfo is useful for controlling info messages in MPI codes, e.g. by setting writeinfo=is_mpi_root().

Rimu.ReportToFile — Type

ReportToFile(; kwargs...) <: ReportingStrategy

Reporting strategy that writes the report directly to a file. Useful when dealing with long jobs or large numbers of replicas, when the report can incur a significant memory cost.

Keyword arguments

filename: the file to report to. If the file already exists, a new file is created.

chunk_size = 1000: the size of each chunk that is written to the file. A DataFrame of this size is collected in memory and written to disk. When saving, an info message is also printed to io.

save_if = true: if this value is true, save the report, otherwise ignore it. Use save_if=is_mpi_root() when running MPI jobs.

return_df: if this value is true, read the file and return the data frame at the end of computation. Otherwise, an empty DataFrame is returned.

io=stdout: The IO to print messages to. Set to devnull if you don't want to see messages printed out.

Rimu.ReportingStrategy — Type

ReportingStrategy

Abstract type for strategies for reporting data in a DataFrame with report!().

Implemented strategies:

ReportDFAndInfo

ReportToFile

Interface:

A ReportingStrategy can define any of the following:

refine_r_strat

report!

report_after_step

finalize_report!

Rimu.RunTillLastStep — Type

RunTillLastStep(step::Int = 0 # number of current/starting timestep
             laststep::Int = 100 # number of final timestep
             shiftMode::Bool = false # whether to adjust shift
             shift = 0.0 # starting/current value of shift
             dτ::Float64 = 0.01 # current value of time step
) <: FciqmcRunStrategy

Parameters for running fciqmc!() for a fixed number of time steps. For alternative strategies, see FciqmcRunStrategy.

Rimu.ShiftMemory — Type

ShiftMemory(Δ::Int) <: MemoryStrategy

Effectively replaces the fluctuating shift update procedure for the coefficient vector by an averaged shift over Δ timesteps, where Δ = 1 means no averaging.

Rimu.ShiftStrategy — Type

Abstract type for defining the strategy for updating the shift with update_shift(). Implemented strategies:

DontUpdate

DoubleLogUpdate - default in lomc!()

LogUpdate

LogUpdateAfterTargetWalkers - FCIQMC standard

DoubleLogUpdateAfterTargetWalkers

Rimu.SignCoherence — Type

SignCoherence(reference[; name=:coherence]) <: PostStepStrategy

After each step, compute the proportion of configurations that have the same sign as they do in the reference_dvec. Reports to a column named name, which defaults to coherence.

Rimu.TimeStepStrategy — Type

TimeStepStrategy

Abstract type for strategies for updating the time step with update_dτ(). Implemented strategies:

ConstantTimeStep

Rimu.TripleLogUpdate — Type

TripleLogUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, η = 0.01) <: ShiftStrategy

Strategy for updating the shift according to the extended log formula with damping parameters ζ, ξ, and η.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^n}\right) - \frac{ξ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^\text{target}}\right) - \frac{η}{dτ}\ln\left(\frac{\|ℜ(Ψ^{n+1})\|_1^2 + \|ℑ(Ψ^{n+1})\|_1^2} {\|ℜ(Ψ^{n})\|_1^2 + \|ℑ(Ψ^{n})\|_1^2}\right),\]

where $N_\mathrm{w}$ is the walkernumber(). When ξ = ζ^2/4 this corresponds to critical damping with a damping time scale T = 2/ζ.

Rimu.WalkerLoneliness — Type

WalkerLoneliness(threshold=1) <: PostStepStrategy

After each step, compute the proportion of configurations that are occupied by at most threshold walkers. Reports to a column named loneliness.

Rimu._n_walkers — Method

_n_walkers(v, s_strat)

Returns an estimate of the expected number of walkers as an integer.

Rimu.advance! — Method

advance!(report::Report, state::QMCState, replica::ReplicaState)

Advance the replica by one step. The state is used only to access the various strategies involved. Steps, stats, and computed quantities are written to the report.

Returns true if the step was successful and calculation should proceed, false when it should terminate.

Rimu.all_overlaps — Method

all_overlaps(operators, vectors)

Get all overlaps between vectors and operators. This function is overlpaded for MPIData.

Rimu.apply_memory_noise! — Method

r = apply_memory_noise!(w, v, shift, dτ, pnorm, m_strat::MemoryStrategy)

Apply memory noise to w, i.e. w .+= r.*v, computing the noise r according to m_strat. Note that m_strat needs to be compatible with StochasticStyle(w). Otherwise, an error exception is thrown. See MemoryStrategy.

w is the walker array after fciqmc step, v the previous one, pnorm the norm of v, and r the instantaneously applied noise.

Rimu.fciqmc_col! — Method

fciqmc_col!(w, ham, add, num, shift, dτ)
fciqmc_col!(::Type{T}, args...)
-> spawns, deaths, clones, antiparticles, annihilations

Spawning and diagonal step of FCIQMC for single column of ham. In essence it computes

w .+= (1 .+ dτ.*(shift .- ham[:,add])).*num.

Depending on T ==StochasticStyle(w), a stochastic or deterministic algorithm will be chosen. The possible values for T are:

IsDeterministic() deteministic algorithm

IsStochasticInteger() stochastic version where the changes added to w are purely integer, according to the FCIQMC algorithm

IsStochasticWithThreshold(c) stochastic algorithm with floating point walkers.

Rimu.fciqmc_step! — Function

fciqmc_step!(Ĥ, v, shift, dτ, pnorm, w;
                      m_strat::MemoryStrategy = NoMemory()) -> ṽ, w̃, stats

Perform a single matrix(/operator)-vector multiplication:

\[\tilde{v} = [1 - dτ(\hat{H} - S)]⋅v ,\]

where Ĥ == ham and S == shift. Whether the operation is performed in stochastic, semistochastic, or determistic way is controlled by the trait StochasticStyle(w). See StochasticStyle. w is a local data structure with the same size and type as v and used for working. Both v and w are modified.

Returns the result ṽ, a (possibly changed) reference to working memory w̃, and the array stats = [spawns, deaths, clones, antiparticles, annihilations]. Stats will contain zeros when running in deterministic mode.

Rimu.finalize_report! — Method

finalize_report!(::ReportingStrategy, report)

Finalize the report. This function is called after all steps in lomc! have finished.

Rimu.lomc! — Method

lomc!(ham::AbstractHamiltonian, [v]; kwargs...) -> df, state
lomc!(state::QMCState, [df]; kwargs...) -> df, state

Linear operator Monte Carlo: Perform a projector quantum Monte Carlo simulation for determining the lowest eigenvalue of ham. v can be a single starting vector. The default choice is

v = DVec(starting_address(ham) => 10; style=IsStochasticInteger())

and triggers the integer walker FCIQMC algorithm. See DVec and StochasticStyle.

Keyword arguments, defaults, and precedence:

params::FciqmcRunStrategy = RunTillLastStep(laststep = 100, dτ = 0.01, shift = diagonal_element(ham, starting_address(ham))) - basic parameters of simulation state, see FciqmcRunStrategy; is mutated

laststep - can be used to override information otherwise contained in params

s_strat::ShiftStrategy = DoubleLogUpdate(targetwalkers = 100, ζ = 0.08, ξ = ζ^2/4) - how to update the shift, see ShiftStrategy

maxlength = 2 * s_strat.targetwalkers + 100 - upper limit on the length of v; when reached, lomc! will abort

style = IsStochasticInteger() - set StochasticStyle for default v; unused if v is specified.

post_step::NTuple{N,<:PostStepStrategy} = () - extract observables (e.g. ProjectedEnergy), see PostStepStrategy.

replica::ReplicaStrategy = NoStats(1) - run several synchronised simulation, see ReplicaStrategy.

r_strat::ReportingStrategy = ReportDFAndInfo() - how and when to report results, see ReportingStrategy

τ_strat::TimeStepStrategy = ConstantTimeStep() - adjust time step dynamically, see TimeStepStrategy

m_strat::MemoryStrategy = NoMemory() - experimental: inject memory noise, see MemoryStrategy

threading = :auto - can be used to control the use of multithreading (overridden by wm)

:auto - use multithreading if s_strat.targetwalkers ≥ 500

true - use multithreading if available (set shell variable JULIA_NUM_THREADS!)

false - run on single thread

wm - working memory; if set, it controls the use of multithreading and overrides threading; is mutated

df = DataFrame() - when called with AbstractHamiltonian argument, a DataFrame can be passed into lomc! that will be pushed into.

Return values

lomc! returns a named tuple with the following fields:

df: a DataFrame with all statistics being reported.

state: a QMCState that can be used for continuations.

Example

julia> add = BoseFS((1,2,3));

julia> hamiltonian = HubbardReal1D(add);

julia> df1, state = lomc!(hamiltonian);

julia> df2, _ = lomc!(state, df1; laststep=200); # Continuation run

julia> size(df1)
(100, 12)

julia> size(df2)
(200, 12)

Rimu.post_step — Function

post_step(::PostStepStrategy, ::ReplicaState) -> kvpairs

Compute statistics after FCIQMC step. Should return a tuple of :key => value pairs. See also PostStepStrategy.

Rimu.refine_r_strat — Method

refine_r_strat(r_strat::ReportingStrategy) -> r_strat

Initialize the reporting strategy. This can be used to set up filenames or other attributes that need to be unique for a run of FCIQMC.

Rimu.replica_stats — Function

replica_stats(::ReplicaStrategy{N}, replicas::NTuple{N,ReplicaState}) -> (names, values)

Return the names and values of statistics reported by ReplicaStrategy. names should be a tuple of Symbols or Strings and values should be a tuple of the same length.

Rimu.report! — Method

 report!(::ReportingStrategy, step, report::Report, keys, values, id="")
 report!(::ReportingStrategy, step, report::Report, nt, id="")

Report keys and values to report, which will be converted to a DataFrame before lomc! exits. Alternatively, a nt::NamedTuple can be passed in place of keys and values. If id is specified, it is appended to all keys. This is used to differentiate between values reported by different replicas.

To overload this function for a new ReportingStrategy, overload report!(::ReportingStrategy, step, args...) and apply the report by calling report!(args...).

Rimu.report! — Method

report!(report, keys, values, id="")
report!(report, pairs, id="")

Write keys, values pairs to report that will be converted to a DataFrame later. Alternatively, a named tuple or a collection of pairs can be passed instead of keys and values.

The value of id is appended to the name of the column, e.g. report!(report, :key, value, :_1) will report value to a column named :key_1.

Rimu.report_after_step — Method

report_after_step(::ReportingStrategy, step, report, state)

This function is called exactly once at the very end of a step. For example, it can be used to print some information to stdout.

Rimu.sort_into_targets! — Method

sort_into_targets!(target, source, stats) -> agg, wm, agg_stats

Aggregate coefficients from source to agg and from stats to agg_stats according to thread- or MPI-level parallelism. wm passes back a reference to working memory.

Rimu.step_stats — Method

step_stats(::StochasticStyle)

Return a tuple of names (Symbol or String) and a zeros of values of the same length. These will be reported as columns in the DataFrame returned by lomc!.

Rimu.threshold_projected_deposit! — Method

threshold_projected_deposit!

This function performs threshold projection before spawning, but only for IsDynamicSemistochastic with the project_later parameter set to false.

Rimu.update_dvec! — Method

update_dvec!([::StochasticStyle,] dvec) -> dvec, nt

Perform an arbitrary transformation on dvec after the spawning step is completed and report statistics to the DataFrame.

Returns the new dvec and a NamedTuple nt of statistics to be reported.

When extending this function for a custom StochasticStyle, define a method for the two-argument call signature!

Rimu.update_dτ — Method

update_dτ(s<:TimeStepStrategy, dτ, tnorm) -> new dτ

Update the time step according to the strategy s.

Rimu.update_shift — Function

update_shift(s <: ShiftStrategy, shift, shiftMode, tnorm, pnorm, dτ, step, df, v_new, v_old)

Update the shift according to strategy s. See ShiftStrategy.

Reexported Submodules

Hamiltonians

Link to Module Hamiltionians.jl

BitStringAddresses

Rimu.BitStringAddresses.AbstractFockAddress — Type

AbstractFockAddress

Supertype representing a Fock state.

Rimu.BitStringAddresses.BitString — Type

BitString{B,N,T<:Unsigned}

Type for storing bitstrings of static size. Holds B bits in N chunks, where each chunk is of type T.

N is chosen automatically to accommodate B bits as efficiently as possible.

Constructors

BitString{B,N,T}(::SVector{N,T}): unsafe constructor. Does not check for ghost bits.

BitString{B,N,T}(i::T): as above, but sets i as the rightmost chunk.

BitString{B}(::Integer): Convert integer to BitString. Integer is truncated to the correct number of bits.

Rimu.BitStringAddresses.BoseFS — Type

BoseFS{N,M,S} <: SingleComponentFockAddress

Address type that represents a Fock state of N spinless bosons in M modes by wrapping a bitstring of type S <: BitString.

Constructors

BoseFS{N,M}(bs::BitString): Unsafe constructor. Does not check whether the number of particles in bs is equal to N.

BoseFS(::BitString): Automatically determine N and M. This constructor is not type stable!

BoseFS{[N,M,S]}(onr): Create BoseFS{N,M} from onr representation. This is efficient as long as at least N is provided.

See also: SingleComponentFockAddress, FermiFS, BitString.

Rimu.BitStringAddresses.BoseFS2C — Type

BoseFS2C{NA,NB,M,AA,AB} <: AbstractFockAddress

Address type that constructed with two BoseFS{N,M,S}. It represents a Fock state with two components, e.g. two different species of bosons with particle number NA from species S and particle number NB from species B. The number of modes M is expected to be the same for both components.

Rimu.BitStringAddresses.BoseFSIndex — Type

BoseFSIndex

Convenience struct for indexing into a BoseFS.

Fields:

occnum: the occupation number.

mode: the index of the mode.

offset: the bit offset of the mode.

Rimu.BitStringAddresses.CompositeFS — Type

CompositeFS(addresses::SingleComponentFockAddress...) <: AbstractFockAddress

Used to encode addresses for multi-component models. All component addresses are expected have the same number of modes M.

See also: BoseFS, FermiFS, SingleComponentFockAddress, num_modes

Rimu.BitStringAddresses.FermiFS — Type

FermiFS{N,M,S} <: SingleComponentFockAddress

Address type that represents a Fock state of N fermions of the same spin in M modes by wrapping a bitstring of type S <: BitString.

Constructors

FermiFS{N,M}(bs::BitString): Unsafe constructor. Does not check whether the number of particles in bs is equal to N.

FermiFS(::BitString): Automatically determine N and M. This constructor is not type stable!

FermiFS{[N,M,S]}(onr): Create FermiFS{N,M} from onr representation. This is efficient as long as at least N is provided.

See also: SingleComponentFockAddress, BoseFS, BitString.

Rimu.BitStringAddresses.FermiOccupiedModes — Type

FermiOccupiedModes

Iterator over occupied modes in address. See occupied_modes.

Rimu.BitStringAddresses.SingleComponentFockAddress — Type

SingleComponentFockAddress{M}

A type representing a single component Fock state with M modes.

Implemented subtypes: BoseFS, FermiFS.

Supported functionality

find_mode

find_occupied_mode

num_occupied_modes

occupied_modes

move_particle

Rimu.BitStringAddresses.check_bitstring_typeparams — Method

check_bitstring_typeparams(::Val{B}, ::Val{N})

Check if number of bits B is consistent with number of chunks N. Throw an error if not.

Rimu.BitStringAddresses.chunk_bits — Method

chunks_bits(::Type{<:BitString}, i)
chunks_bits(s, i)

Number of bits in the i-th chunk of s.

Rimu.BitStringAddresses.chunk_type — Method

chunk_type(::Type{<:BitString})
chunk_type(s::BitString)

Type of unsigned integer used to store the chunks.

Rimu.BitStringAddresses.chunks — Method

chunks(s::BitString)

SVector that stores the chunks of s.

Rimu.BitStringAddresses.find_mode — Function

find_mode(::SingleComponentFockAddress, i)

Find the i-th mode in address. Returns BoseFSIndex for BoseFS, and an integer for FermiFS.

julia> find_mode(BoseFS((1, 0, 2)), 2)
3-element Rimu.BitStringAddresses.BoseFSIndex with indices SOneTo(3):
 0
 2
 2

julia> find_mode(FermiFS((1, 1, 1, 0)), 2)
2

Rimu.BitStringAddresses.find_occupied_mode — Function

find_occupied_mode(::SingleComponentFockAddress, k)

Find the k-th occupied mode in address. Returns BoseFSIndex for BoseFS, and an integer for FermiFS.

Example

julia> find_occupied_mode(BoseFS((1, 0, 2)), 2)
3-element Rimu.BitStringAddresses.BoseFSIndex with indices SOneTo(3):
 2
 3
 3

julia> find_occupied_mode(FermiFS((1, 1, 1, 0)), 2)
2

Rimu.BitStringAddresses.has_ghost_bits — Method

has_ghost_bits(s::BitString)

Check for bits outside data field.

See also: remove_ghost_bits.

Rimu.BitStringAddresses.m_onr — Method

m_onr(bs)

Compute and return the occupation number representation of the bit string address bs as an MVector{M,Int32}, where M is the number of modes.

Rimu.BitStringAddresses.move_particle — Function

move_particle(add::SingleComponentFockAddress, i, j) -> nadd, α

Move particle from mode i to mode j. Equivalent to $a^{\dagger}_i a_j |add\rangle$. Returns the new Fock state address nadd and integer α. For FermiFS the factor α is the correct (signed) amplitude, whereas for BoseFS the correct amplitude is $\sqrt{α}$.

Note that the modes in BoseFS are indexed by BoseFSIndex, while the ones in FermiFS are indexed by integers (see example below).

Example

julia> b = BoseFS((1, 1, 3, 0))
BoseFS{5,4}((1, 1, 3, 0))

julia> i = find_occupied_mode(b, 2)
3-element Rimu.BitStringAddresses.BoseFSIndex with indices SOneTo(3):
 1
 2
 2

julia> j = find_mode(b, i.mode + 1)
3-element Rimu.BitStringAddresses.BoseFSIndex with indices SOneTo(3):
 3
 3
 4

julia> move_particle(b, i, j)
(BoseFS{5,4}((1, 0, 4, 0)), 2.0)

julia> move_particle(b, j, j)
(BoseFS{5,4}((1, 1, 3, 0)), 3.0)

julia> f = FermiFS((1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0))
FermiFS{7,12}((1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0))

julia> i = find_occupied_mode(f, 2)
6

julia> move_particle(f, i, i + 1)
(FermiFS{7,12}((1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0)), 0.0)

julia> move_particle(f, i, i - 1)
(FermiFS{7,12}((1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0)), 1.0)

julia> move_particle(f, i, 12)
(FermiFS{7,12}((1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1)), -1.0)

Rimu.BitStringAddresses.near_uniform — Method

near_uniform(BoseFS{N,M})
near_uniform(BoseFS{N,M,S}) -> bfs::BoseFS{N,M,S}

Create bosonic Fock state with near uniform occupation number of M modes with a total of N particles. Specifying the bit address type S is optional.

Examples

julia> near_uniform(BoseFS{7,5,BitString{14}})
BoseFS((2,2,1,1,1))

julia> near_uniform(BoseFS{7,5})
BoseFS((2,2,1,1,1))

Rimu.BitStringAddresses.near_uniform_onr — Method

near_uniform_onr(N, M) -> onr::SVector{M,Int}

Create occupation number representation onr distributing N particles in M modes in a close-to-uniform fashion with each mode filled with at least N ÷ M particles and at most with N ÷ M + 1 particles.

Rimu.BitStringAddresses.num_bits — Method

num_bits(::Type{<:BitString})
num_bits(s::BitString)

Total number of bits stored in bitstring.

Rimu.BitStringAddresses.num_chunks — Method

num_chunks(::Type{<:BitString})
num_chunks(s::BitString)

Number of chunks in bitstring. Equivalent to length(chunks(s)).

Rimu.BitStringAddresses.num_chunks — Method

num_chunks(::Val{B})

Determine the number and type of chunks needed to store B bits.

Rimu.BitStringAddresses.num_components — Method

num_components(::Type{<:AbstractFockAddress})
num_components(::AbstractFockAddress)

Number of components in address.

Rimu.BitStringAddresses.num_modes — Method

num_modes(::Type{<:AbstractFockAddress})
num_modes(::AbstractFockAddress)

Number of modes represented by address.

Rimu.BitStringAddresses.num_occupied_modes — Function

num_occupied_modes(::SingleComponentFockAddress)

Get the number of occupied modes in address. Equivalent to length(occupied_modes(address)), or the number of non-zeros in its ONR representation.

Example

julia> num_occupied_modes(BoseFS((1, 0, 2)))
2
julia> num_occupied_modes(FermiFS((1, 1, 1, 0)))
3

Rimu.BitStringAddresses.num_particles — Method

num_particles(::Type{<:AbstractFockAddress})
num_particles(::AbstractFockAddress)

Number of particles represented by address.

Rimu.BitStringAddresses.occupied_modes — Function

occupied_modes(::SingleComponentFockAddress)

Iterate over all occupied modes in an address. Iterates BoseFSIndex for BoseFS, and an integers for FermiFS.

Example

julia> b = BoseFS((1,5,0,4))
julia> for (n, i, m) in occupied_modes(b)
    @show n, i, m
end
(n, i, m) = (1, 1, 0)
(n, i, m) = (5, 2, 2)
(n, i, m) = (4, 4, 9)

julia> f = FermiFS((1,1,0,1,0,0,1))
julia> for i in occupied_modes(f)
    @show i
end
i = 1
i = 2
i = 4
i = 7

Rimu.BitStringAddresses.onr — Method

onr(bs)

Compute and return the occupation number representation of the bit string address bs as an SVector{M,Int32}, where M is the number of modes.

Rimu.BitStringAddresses.partial_left_shift — Method

partial_left_shift(bs::BitString, i, j)

Shift a part of the bitstring left by one place with boundaries i < j. In a BoseFS bitstring, it moves a particle at offset i to the position at offset j.

See also: move_particle, partial_right_shift.

Rimu.BitStringAddresses.partial_right_shift — Method

partial_right_shift(bs::BitString, i, j)

Shift a part of the bitstring right by one place with boundaries i < j. In a BoseFS bitstring, it moves a particle at offset j to the position at offset i.

See also: partial_left_shift, move_particle.

Rimu.BitStringAddresses.remove_ghost_bits — Method

remove_ghost_bits(s::BitString)

Remove set bits outside data field if any are present.

See also: has_ghost_bits.

Rimu.BitStringAddresses.top_chunk_bits — Method

top_chunk_bits(::Type{<:BitString})
top_chunk_bits(s::BitString)

Number of bits stored in top chunk. Equivalent to chunk_bits(s, 1).

Rimu.BitStringAddresses.update_component — Method

update_component(c::CompositeFS, new, ::Val{i})

Replace the i-th component in c with new. Used for updating a single component in the address.

DictVectors

Rimu.DictVectors — Module

Module that provides data structures that behave similar to sparse vectors, but are indexed by arbitrary types (could be non-integers) similarly to dictionaries. The idea is to do linear algebra with data structures that are neither subtyped to AbstractVector nor to AbstractDict and are suitable for use with KrylovKit.jl. For this, the abstract type and interface AbstractDVec is provided, with the concrete implementation of DVec

Rimu.DictVectors.AbstractDVec — Type

DictVectors.AbstractDVec{K,V}

Abstract type for sparse vectors with valtype V based on dictionary-like structures. The vectors are designed to work well with FCIQMC and KrylovKit.

They lie somewhere between AbstractDicts and sparse AbstractVectors, generally behaving like a dictionary, while supportting various linear algebra functionality. Indexing with a value not stored in the dictionary returns zero(V). Setting a stored value to 0 or below eps(V::AbstractFloat) removes the value from the dictionary. Their length signals the number of stored elements, not the size of the vector space.

They have a StochasticStyle which selects the spawning algorithm in FCIQMC.

To iterate over an AbstractDVec, use pairs or values.

Interface

The interface is similar to the AbstractDict interface.

Implement what would be needed for the AbstractDict interface (pairs, keys, values, setindex!, getindex, delete!, length, haskey, empty!, isempty) and, in addition:

StochasticStyle

storage(dv) returns an AbstractDict storing the raw data with possibly different valtype than V.

Rimu.DictVectors.AbstractProjector — Type

Abstract supertype for projectors to be used in in lieu of DVecs or Vectors.

Rimu.DictVectors.CoherentInitiator — Type

CoherentInitiator(threshold) <: InitiatorRule

Initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators can spawn to initiators.

Multiple non-initiators can spawn to a single non-initiator if their contributions add up to a value greater than the initiator threshold.

See InitiatorRule.

Rimu.DictVectors.DVec — Type

DVec{K,V,D<:AbstractDict{K,V},S}

Dictionary-based vector-like data structure for use with FCIQMC and KrylovKit. While mostly behaving like a Dict, it supports various linear algebra operations such as norm and dot. It has a StochasticStyle that is used to select an appropriate spawning strategy in the FCIQMC algorithm.

See also: AbstractDVec.

Constructors

DVec(dict::AbstractDict[; style, capacity]): create a DVec with dict for storage. Note that the data may or may not be copied.

DVec(args...[; style, capacity]): args... are passed to the Dict constructor. The Dict is used for storage.

DVec{K,V}([; style, capacity]): create an empty DVec{K,V}.

DVec(dv::AbstractDVec[; style, capacity]): create a DVec with the same contents as adv. The style is inherited from dv by default.

The default style is selected based on the DVec's valtype (see default_style). If a style is given and the valtype does not match the style's eltype, the values are converted to an appropriate type.

The capacity argument is optional and sets the initial size of the DVec via sizehint!.

Examples

julia> dv = DVec(:a => 1)
DVec{Symbol,Int64} with 1 entries, style = IsStochasticInteger{Int64}()
  :a => 1

julia> dv = DVec(:a => 2, :b => 3; style=IsDynamicSemistochastic())
DVec{Symbol,Float32} with 2 entries, style = IsDynamicSemistochastic{Float64, true}(1.0, Inf, 1.0)
  :a => 2.0
  :b => 3.0

Rimu.DictVectors.FrozenDVec — Type

FrozenDVec

See: freeze.

Rimu.DictVectors.Initiator — Type

Initiator(threshold) <: InitiatorRule

Initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators can spawn to initiators.

See InitiatorRule.

Rimu.DictVectors.InitiatorDVec — Type

InitiatorDVec{K,V} <: AbstractDVec{K,V}

Dictionary-based vector-like data structure for use with lomc! and KrylovKit.jl. See AbstractDVec. Functionally identical to DVec, but contains InitiatorValues internally in order to facilitate initiator methods. How the initiators are handled is controlled by the initiator keyword argument (see below).

Constructors

InitiatorDVec(dict::AbstractDict[; style, initiator, capacity]): create an InitiatorDVec with dict for storage. Note that the data may or may not be copied.

InitiatorDVec(args...[; style, initiator, capacity]): args... are passed to the Dict constructor. The Dict is used for storage.

InitiatorDVec{K,V}([; style, initiator, capacity]): create an empty InitiatorDVec{K,V}.

InitiatorDVec(dv::AbstractDVec[; style, initiator, capacity]): create an InitiatorDVec with the same contents as dv. The style is inherited from dv by default.

Keyword arguments

style: A valid StochasticStyle. The default is selected based on the InitiatorDVec's valtype (see default_style). If a style is given and the valtype does not match the style's eltype, the values are converted to an appropriate type.

initiator = Initiator(1): A valid InitiatorRule. See Initiator.

capacity: Indicative size as Int. Optional. Sets the initial size of the InitiatorDVec via sizehint!.

Rimu.DictVectors.InitiatorIterator — Type

InitiatorIterator

Iterator over pairs or values of an InitiatorDVec. Supports the SplittablesBase interface.

Rimu.DictVectors.InitiatorRule — Type

InitiatorRule{V}

Abstract type for defining initiator rules for InitiatorDVec. Concrete implementations:

Initiator

SimpleInitiator

CoherentInitiator

When defining a new InitiatorRule, also define a corresponding method for value!

Rimu.DictVectors.InitiatorValue — Type

InitiatorValue{V}(; safe::V, unsafe::V, initiator::V) where V

Composite "walker" with three fields. For use with InitiatorDVecs.

Rimu.DictVectors.IsDeterministic — Type

IsDeterministic{T=Float64}() <: StochasticStyle{T}

Trait for generalised vector of configuration indicating deterministic propagation of walkers.

See also StochasticStyle.

Rimu.DictVectors.IsDynamicSemistochastic — Type

IsDynamicSemistochastic{T=Float64}(rel_threshold=1, abs_threshold=Inf, proj_threshold=1) <: StochasticStyle{T}

QMC propagation with non-integer walker numbers and reduced noise. All possible spawns are performed deterministically when number of walkers in a configuration is high. Stochastic vector compression with threshold proj_threshold is applied after spawning and diagonal death steps.

Unlike with IsStochasticWithThreshold, when late_projection is set to true, walker annihilation is done before the stochastic vector compression.

Parameters:

late_projection = true: If set to true, threshold projection is done after all spawns are collected, otherwise, values are projected as they are being spawned.

rel_threshold = 1.0: If the walker number on a configuration times this threshold is greater than the number of offdiagonals, spawning is done deterministically. Should be set to 1 or more for best performance.

abs_threshold = Inf: If the walker number on a configuration is greater than this value, spawning is done deterministically. Can be set to e.g abs_threshold = 0.1 * target_walkers.

proj_threshold = 1.0: Values below this number are stochastically projected to this value or zero. See also IsStochasticWithThreshold.

See also StochasticStyle.

Rimu.DictVectors.IsStochastic2Pop — Type

IsStochastic2Pop{T=Complex{Int}}() <: StochasticStyle{T}

Trait for generalised vector of configurations indicating stochastic propagation with complex walker numbers representing two populations of integer walkers.

See also StochasticStyle.

Rimu.DictVectors.IsStochasticInteger — Type

IsStochasticInteger{T=Int}() <: StochasticStyle{T}

Trait for generalised vector of configurations indicating stochastic propagation as seen in the original FCIQMC algorithm.

See also StochasticStyle.

Rimu.DictVectors.IsStochasticWithThreshold — Type

IsStochasticWithThreshold(threshold=1.0) <: StochasticStyle

Trait for generalised vector of configurations indicating stochastic propagation with real walker numbers and cutoff threshold.

During stochastic propagation, walker numbers small than threshold will be stochastically projected to either zero or threshold.

See also StochasticStyle.

Rimu.DictVectors.Norm1ProjectorPPop — Type

Norm1ProjectorPPop()

Results in computing the one-norm per population when used in dot(). E.g.

dot(Norm1ProjectorPPop(),x)
-> norm(real.(x),1) + im*norm(imag.(x),1)

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.Norm2Projector — Type

Norm2Projector()

Results in computing the two-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,2) # with type Float64

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.NormProjector — Type

NormProjector()

Results in computing the one-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,1)

NormProjector() thus represents the vector sign.(x).

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.PopsProjector — Type

PopsProjector()

Results in computing the projection of one population on the other when used in dot(). E.g.

dot(PopsProjector(),x)
-> real(x) ⋅ imag(x)

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.SimpleInitiator — Type

SimpleInitiator(threshold) <: InitiatorRule

Simplified initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators cannot spawn.

See InitiatorRule.

Rimu.DictVectors.StochasticStyle — Type

StochasticStyle(v)

Abstract type. When called as a function it returns the native style of the generalised vector v that determines how simulations are to proceed.

Implemented styles

IsStochasticInteger - integer walker FCIQMC

IsDeterministic - perform deterministic variant of power method

IsStochasticWithThreshold - floating point walker FCIQMC

IsDynamicSemistochastic

Usage

Concrete StochasticStyles can be used for the style keyword argument of lomc! and DVec.

Interface

When defining a new StochasticStyle, subtype it as MyStyle<:StochasticStyle{T} where T is the concrete value type the style is designed to work with.

For it to work with lomc!, a StochasticStyle must define the following:

fciqmc_col!(::StochasticStyle, w, H, address, value, shift, dτ)

step_stats(::StochasticStyle)

Optionally, it can also define update_dvec!, which can be used to perform arbitrary transformations on the generalised vector after the spawning step is complete.

Rimu.DictVectors.StyleUnknown — Type

StyleUnknown{T}() <: StochasticStyle

Trait for value types not (currently) compatible with FCIQMC. This style makes it possible to construct dict vectors with unsupported valtypes.

See also StochasticStyle.

Rimu.DictVectors.UniformProjector — Type

UniformProjector()

Represents a vector with all elements 1. To be used with dot(). Minimizes memory allocations.

UniformProjector()⋅v == sum(v)
dot(UniformProjector(), LO, v) == sum(LO*v)

See also ReportingStrategy for use of projectors in FCIQMC.

OrderedCollections.freeze — Method

freeze(dv)

Create a "frozen" version of dv which can no longer be modified or used in the conventional manner, but supports faster dot products.

If dv is an MPIData, synchronize its contents among the ranks first.

Rimu.DictVectors.add! — Method

add!(x::AbstractDVec,y::AbstactDVec)

Inplace add x+y and store result in x.

Rimu.DictVectors.default_style — Method

default_style(::Type)

Pick a StochasticStyle based on the value type. Throws an error if no known default style is known.

Rimu.DictVectors.deposit! — Method

deposit!(w::AbstractDVec, add, val, parent::Pair)

Add val into w at address add, taking into account initiator rules if applicable. parent contains the address => value pair from which the pair add => val was created. InitiatorDVec can intercept this and add its own functionality.

Rimu.DictVectors.deposit! — Method

deposit!(w::InitiatorDVec, add, val, p_add=>p_val)

Add val into w at address add as an InitiatorValue.

Rimu.DictVectors.localpart — Method

localpart(dv) -> AbstractDVec

Get the part of dv that is located on this MPI rank. Returns dv itself for DictVectors.

Rimu.DictVectors.storage — Function

storage(dvec) -> AbstractDict

Return the raw storage associated with dvec as an AbstractDict. Used in MPI communication.

Rimu.DictVectors.value — Function

value(i::InitiatorRule, v::InitiatorValue)

Convert the InitiatorValue v into a scalar value according to the InitiatorRule i.

Internal function that implements functionality of InitiatorDVec.

Rimu.DictVectors.walkernumber — Method

walkernumber(w)

Compute the number of walkers in w. It is used for updating the shift. Overload this function for modifying population control.

In most cases walkernumber(w) is identical to norm(w,1). For AbstractDVecs with complex coefficients it reports the one norm separately for the real and the imaginary part as a ComplexF64. See Norm1ProjectorPPop.

Rimu.DictVectors.zero! — Method

zero!(v)

Replace v by a zero vector as an inplace operation. For AbstractDVec types it means removing all non-zero elements. For AbstractArrays, it sets all of the values to zero.

ConsistentRNG

Link to Module ConsistentRNG.jl

Unexported Submodules

StatsTools

Link to Module Rimu/StatsTools

Blocking

Modules = [Blocking]

Index

Rimu.ConsistentRNG

Rimu.DictVectors

Rimu.Hamiltonians

Rimu.Rimu

Rimu.StatsTools

Rimu.ConsistentRNG.CRNGs

MonteCarloMeasurements.Particles

Rimu.AllOverlaps

Rimu.BitStringAddresses.AbstractFockAddress

Rimu.BitStringAddresses.BitString

Rimu.BitStringAddresses.BoseFS

Rimu.BitStringAddresses.BoseFS2C

Rimu.BitStringAddresses.BoseFSIndex

Rimu.BitStringAddresses.CompositeFS

Rimu.BitStringAddresses.FermiFS

Rimu.BitStringAddresses.FermiOccupiedModes

Rimu.BitStringAddresses.SingleComponentFockAddress

Rimu.ConsistentRNG.CRNG

Rimu.ConstantTimeStep

Rimu.DeltaMemory

Rimu.DeltaMemory2

Rimu.DeltaMemory3

Rimu.DictVectors.AbstractDVec

Rimu.DictVectors.AbstractProjector

Rimu.DictVectors.CoherentInitiator

Rimu.DictVectors.DVec

Rimu.DictVectors.FrozenDVec

Rimu.DictVectors.Initiator

Rimu.DictVectors.InitiatorDVec

Rimu.DictVectors.InitiatorIterator

Rimu.DictVectors.InitiatorRule

Rimu.DictVectors.InitiatorValue

Rimu.DictVectors.IsDeterministic

Rimu.DictVectors.IsDynamicSemistochastic

Rimu.DictVectors.IsStochastic2Pop

Rimu.DictVectors.IsStochasticInteger

Rimu.DictVectors.IsStochasticWithThreshold

Rimu.DictVectors.Norm1ProjectorPPop

Rimu.DictVectors.Norm2Projector

Rimu.DictVectors.NormProjector

Rimu.DictVectors.PopsProjector

Rimu.DictVectors.SimpleInitiator

Rimu.DictVectors.StochasticStyle

Rimu.DictVectors.StyleUnknown

Rimu.DictVectors.UniformProjector

Rimu.DontUpdate

Rimu.DoubleLogProjected

Rimu.DoubleLogSumUpdate

Rimu.DoubleLogUpdate

Rimu.DoubleLogUpdateAfterTargetWalkers

Rimu.FciqmcRunStrategy

Rimu.Hamiltonians.AbstractHamiltonian

Rimu.Hamiltonians.BoseHubbardMom1D2C

Rimu.Hamiltonians.BoseHubbardReal1D2C

Rimu.Hamiltonians.ExtendedHubbardReal1D

Rimu.Hamiltonians.HardwallBoundaries

Rimu.Hamiltonians.HubbardMom1D

Rimu.Hamiltonians.HubbardReal1D

Rimu.Hamiltonians.HubbardRealSpace

Rimu.Hamiltonians.LOStructure

Rimu.Hamiltonians.LadderBoundaries

Rimu.Hamiltonians.MatrixHamiltonian

Rimu.Hamiltonians.PeriodicBoundaries

Rimu.LogUpdate

Rimu.LogUpdateAfterTargetWalkers

Rimu.MemoryStrategy

Rimu.MultiScalar

Rimu.NoMemory

Rimu.NoStats

Rimu.PostStepStrategy

Rimu.ProjectedEnergy

Rimu.ProjectedMemory

Rimu.Projector

Rimu.PurgeNegatives

Rimu.QMCState

Rimu.ReplicaState

Rimu.ReplicaStrategy

Rimu.Report

Rimu.ReportDFAndInfo

Rimu.ReportToFile

Rimu.ReportingStrategy

Rimu.RunTillLastStep

Rimu.ShiftMemory

Rimu.ShiftStrategy

Rimu.SignCoherence

Rimu.StatsTools.BlockingResult

Rimu.StatsTools.RatioBlockingResult

Rimu.TimeStepStrategy

Rimu.TripleLogUpdate

Rimu.WalkerLoneliness

Measurements.measurement

OrderedCollections.freeze

Rimu.BitStringAddresses.check_bitstring_typeparams

Rimu.BitStringAddresses.chunk_bits

Rimu.BitStringAddresses.chunk_type

Rimu.BitStringAddresses.chunks

Rimu.BitStringAddresses.find_mode

Rimu.BitStringAddresses.find_occupied_mode

Rimu.BitStringAddresses.has_ghost_bits

Rimu.BitStringAddresses.m_onr

Rimu.BitStringAddresses.move_particle

Rimu.BitStringAddresses.near_uniform

Rimu.BitStringAddresses.near_uniform_onr

Rimu.BitStringAddresses.num_bits

Rimu.BitStringAddresses.num_chunks

Rimu.BitStringAddresses.num_chunks

Rimu.BitStringAddresses.num_components

Rimu.BitStringAddresses.num_modes

Rimu.BitStringAddresses.num_occupied_modes

Rimu.BitStringAddresses.num_particles

Rimu.BitStringAddresses.occupied_modes

Rimu.BitStringAddresses.onr

Rimu.BitStringAddresses.partial_left_shift

Rimu.BitStringAddresses.partial_right_shift

Rimu.BitStringAddresses.remove_ghost_bits

Rimu.BitStringAddresses.top_chunk_bits

Rimu.BitStringAddresses.update_component

Rimu.ConsistentRNG.cRand

Rimu.ConsistentRNG.cRandn

Rimu.ConsistentRNG.check_crng_independence

Rimu.ConsistentRNG.newChildRNG

Rimu.ConsistentRNG.seedCRNG!

Rimu.ConsistentRNG.sync_cRandn

Rimu.ConsistentRNG.trng

Rimu.DictVectors.add!

Rimu.DictVectors.default_style

Rimu.DictVectors.deposit!

Rimu.DictVectors.deposit!

Rimu.DictVectors.localpart

Rimu.DictVectors.storage

Rimu.DictVectors.value

Rimu.DictVectors.walkernumber

Rimu.DictVectors.zero!

Rimu.Hamiltonians.diagonal_element

Rimu.Hamiltonians.dimension

Rimu.Hamiltonians.get_offdiagonal

Rimu.Hamiltonians.neighbour_site

Rimu.Hamiltonians.num_neighbours

Rimu.Hamiltonians.num_offdiagonals

Rimu.Hamiltonians.offdiagonals

Rimu.Hamiltonians.random_offdiagonal

Rimu.Hamiltonians.starting_address

Rimu.StatsTools.autocovariance

Rimu.StatsTools.blocker

Rimu.StatsTools.blocking_analysis

Rimu.StatsTools.blocks_with_m

Rimu.StatsTools.growth_estimator

Rimu.StatsTools.growth_witness

Rimu.StatsTools.mean_and_se

Rimu.StatsTools.med_and_errs

Rimu.StatsTools.mixed_estimator

Rimu.StatsTools.mtest

Rimu.StatsTools.pseudo_cov

Rimu.StatsTools.ratio_estimators

Rimu.StatsTools.ratio_of_means

Rimu.StatsTools.ratio_with_errs

Rimu.StatsTools.replica_fidelity

Rimu.StatsTools.smoothen

Rimu.StatsTools.to_measurement

Rimu.StatsTools.w_exp

Rimu.StatsTools.w_lin

Rimu.StatsTools.x_by_y_linear

Rimu._n_walkers

Rimu.advance!

Rimu.all_overlaps

Rimu.apply_memory_noise!

Rimu.fciqmc_col!

Rimu.fciqmc_step!

Rimu.finalize_report!

Rimu.lomc!

Rimu.post_step

Rimu.refine_r_strat

Rimu.replica_stats

Rimu.report!

Rimu.report!

Rimu.report_after_step

Rimu.sort_into_targets!

Rimu.step_stats

Rimu.threshold_projected_deposit!

Rimu.update_dvec!

Rimu.update_dτ

Rimu.update_shift

Statistics.cov