API

Rimu

Rimu.Rimu — Module

Rimu

Random Integrator for Many-Body Quantum Systems

Rimu.ConstantTimeStep — Type

Keep dτ constant.

Rimu.DelayedDoubleLogUpdate — Type

DelayedDoubleLogUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, A=10) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ and delay of A steps. See DoubleLogUpdate.

\[S^{n+A} = S^n -\frac{ζ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^n}\right)-\frac{ξ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^\text{target}}\right)\]

Rimu.DelayedDoubleLogUpdateAfterTW — Type

DelayedDoubleLogUpdateAfterTW(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, A=10) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ and delay of A steps after the number of target walkers is reached. See DoubleLogUpdate.

\[S^{n+A} = S^n -\frac{ζ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^n}\right)-\frac{ξ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^\text{target}}\right)\]

Rimu.DelayedLogUpdate — Type

DelayedLogUpdate(ζ = 0.08, a = 10) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and delay of a steps.

\[S^{n+a} = S^n -\frac{ζ}{a dτ}\ln\left(\frac{\|Ψ\|_1^{n+a}}{\|Ψ\|_1^n}\right)\]

Rimu.DelayedLogUpdateAfterTargetWalkers — Type

DelayedLogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08, a = 10) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and delay of a steps. See DelayedLogUpdate.

Rimu.DeltaMemory — Type

DeltaMemory(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = (pnorm - tnorm)/(dτ*pnorm) + shift
r = r̃ - <r̃>

Rimu.DeltaMemory2 — Type

DeltaMemory2(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = pnorm - tnorm + shift*dτ*pnorm
r = (r̃ - <r̃>)/(dτ*pnorm)

The long-term average of r is not guaranteed to be zero.

Rimu.DeltaMemory3 — Type

DeltaMemory3(Δ::Int, level::Float64) <: MemoryStrategy

Before updating the shift, apply multiplicative memory noise with a memory length of Δ at level level, where Δ = 1 means no memory noise.

r̃ = (pnorm - tnorm)/pnorm + dτ*shift
r = r̃ - <r̃>
w .*= 1 + level*r

Rimu.DontUpdate — Type

DontUpdate() <: ShiftStrategy Don't update the shift.

Rimu.DoubleLogProjected — Type

DoubleLogProjected(; target, projector, ζ = 0.08, ξ = ζ^2/4) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ after projecting onto projector.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{P⋅Ψ^{(n+1)}}{P⋅Ψ^{(n)}}\right)-\frac{ξ}{dτ}\ln\left(\frac{P⋅Ψ^{(n+1)}}{\text{target}}\right)\]

Rimu.DoubleLogUpdate — Type

DoubleLogUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)-\frac{ξ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^\text{target}}\right)\]

When ξ = ζ^2/4 this corresponds to critical damping with a damping time scale T = 2/ζ.

Rimu.DoubleLogUpdateAfterTargetWalkers — Type

LogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08, ξ = 0.0016) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and ξ. See DoubleLogUpdate.

Rimu.DoubleLogUpdateAfterTargetWalkersSwitch — Type

LogUpdateAfterTargetWalkersSwitch(targetwalkers, ζ = 0.08, ξ = 0.0016) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and ξ. After a steps the strategy switches to LogUpdate. See DoubleLogUpdate.

Rimu.EveryKthStep — Type

EveryKthStep(;k = 10, projector = missing)

Report every kth step. Include projection onto projector. See ReportingStrategy for details.

Rimu.EveryTimeStep — Type

EveryTimeStep(;projector = missing)

Report every time step. Include projection onto projector. See ReportingStrategy for details.

Rimu.FciqmcRunStrategy — Type

Abstract type representing the strategy for running and terminating fciqmc!(). Implemented strategies:

RunTillLastStep

Rimu.HistoryLogUpdate — Type

HistoryLogUpdate(df::DataFrame; d = 100, k=1, ζ= 0.08)

Strategy for updating the shift according to log formula but with walker numbers accumulated from k samples of the history with delay d. A recent history has to be passed with the data frame df for initialisation.

\[N_w^{n} = \sum_{i=0}^{k-1} \|Ψ\|_1^{n-i} \\ S^{n+a} = S^n -\frac{ζ}{a dτ}\ln\left(\frac{N_w^{n+1}}{N_w^n}\right)\]

Rimu.IsSemistochastic — Type

IsSemistochastic(threshold::Float16, d_space)

Trait for generalised vector of configurations indicating semistochastic propagation. Set with setSemistochastic!.

> StochasticStyle(V) = IsSemistochastic(threshold, d_space)

where d_space is a vector of addresses defining the the stochastic subspace.

Rimu.IsStochasticWithThreshold — Type

IsStochasticWithThreshold(threshold::Float16)

Trait for generalised vector of configurations indicating stochastic propagation with real walker numbers and cutoff threshold.

> StochasticStyle(V) = IsStochasticWithThreshold(threshold)

During stochastic propagation, walker numbers small than threshold will be stochastically projected to either zero or threshold.

The trait can be conveniently defined on an instance of a generalised vector with the macro @setThreshold. Example:

julia> dv = DVec(Dict(nearUniform(BoseFS{3,3})=>3.0))
julia> @setThreshold dv 0.6
julia> StochasticStyle(dv)
IsStochasticWithThreshold(0.6f0)

Rimu.LogUpdate — Type

LogUpdate(ζ = 0.08) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)\]

Rimu.LogUpdateAfterTargetWalkers — Type

LogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ. See LogUpdate.

Rimu.MPIData — Type

Simple wrapper used for signaling that this data is part of a distributed data structure and communication should happen with MPI.

Rimu.MPIDefault — Type

MPIDefault(nprocs, my_rank, comm)

Strategy for point-to-point MPI communication.

Rimu.MPINoWalkerExchange — Type

MPINoWalkerExchange(nprocs, my_rank, comm)

Strategy for for not exchanging walkers between ranks. Consequently there will be no cross-rank annihilations.

Rimu.MPIOSWin — Type

MPIOSWin(nprocs, myrank, comm, ::Type{T}, capacity)

Communication buffer for use with MPI one-sided communication (remote memory access). Up to capacity elements of type T can be exchanged between MPI ranks via put. It is important that isbitstype(T) == true. Objects of type MPIOSWin have to be freed manually with a (blocking) call to free().

Rimu.MemoryStrategy — Type

Abstract type for defining the strategy for injectimg memory noise. Implemented strategies:

NoMemory

DeltaMemory

ShiftMemory

Rimu.NoMemory — Type

NoMemory <: MemoryStrategy

Default strategy for MemoryStrategy indicating that no memory noise will be used.

Rimu.NoProjection — Type

Do not project the walker amplitudes. See norm_project.

Rimu.NoProjectionTwoNorm — Type

Do not project the walker amplitudes. Use two-norm to calculate walker numbers. This affects reported "norm" but also the shift update procedures. See norm_project.

Rimu.ProjectStrategy — Type

Abstract type for defining the strategy of projection for fciqmc with floating point walker number with norm_project. Implemented strategies:

NoProjection

NoProjectionTwoNorm

ThresholdProject

ScaledThresholdProject

Rimu.ProjectedMemory — Type

ProjectedMemory(Δ::Int, projector, pp::Number) <: MemoryStrategy
ProjectedMemory(Δ::Int, projector, v::AbstractDVec)

Before updating the shift, apply memory noise to minimize the fluctuations of the overlap of the coefficient vector with projector. Averaging over Δ time steps is applied, where Δ = 1 means no memory noise is applied. Use pp to initialise the value of the projection or pass v in order to initialise the projection with pp = projector.v.

r̃ = (projector⋅v - projector⋅w)/projector⋅v + dτ*shift
r = r̃ - <r̃>

where v is the coefficient vector before and w after applying a regular FCIQMC step.

Rimu.ReportDFAndInfo — Type

ReportDFAndInfo(; k=10, i=100, io=stdout, writeinfo=true, projector = missing)

Report every kth step in DataFrame and write info message to io every ith step (unless writeinfo == false). The flag writeinfo is useful for controlling info messages in MPI codes. Include projection onto projector. See ReportingStrategy for details.

Rimu.ReportingStrategy — Type

ReportingStrategy

Abstract type for strategies for reporting data in a DataFrame with report!(). It also affects the calculation and reporting of projected quantities in the DataFrame.

Implemented strategies:

EveryTimeStep

EveryKthStep

ReportDFAndInfo

Every strategy accepts the keyword argument projector according to which a projection of the instantaneous coefficient vector projector⋅v and Hamiltonian dot(projector, H, v) are reported to the DataFrame in the fields df.vproj and df.hproj, respectively. Possible values for projector are

missing - no projections are computed (default)

dv::AbstractDVec - compute projection onto coefficient vector dv

UniformProjector() - projection onto vector of all ones

NormProjector() - compute 1-norm instead of projection (only df.vproj)

Norm2Projector() - compute 2-norm instead of projection (only df.vproj)

Examples

r_strat = EveryTimeStep(projector = copy(svec))

Record the projection of instananeous coefficient vector and Hamiltonian onto the starting vector, and report every time step.

r_strat = EveryKthStep(k=10, projector = UniformProjector())

Record the projection of instananeous coefficient vector and Hamiltonian onto the uniform vector of all 1s, and report every kth time step.

Rimu.RunTillLastStep — Type

RunTillLastStep(step::Int = 0 # number of current/starting timestep
             laststep::Int = 100 # number of final timestep
             shiftMode::Bool = false # whether to adjust shift
             shift::Float64 = 0.0 # starting/current value of shift
             dτ::Float64 = 0.01 # current value of time step
) <: FciqmcRunStrategy

Parameters for running fciqmc!() for a fixed number of time steps. For alternative strategies, see FciqmcRunStrategy.

Rimu.ScaledThresholdProject — Type

ScaledThresholdProject(threshold = 1.0) <: ProjectStrategy

Project stochastically for walker amplitudes below threshold and scale configuration array as to keep the norm constant. As a consequence, the final configuration amplitudes may be smaller than threshold. See norm_project.

Rimu.ShiftMemory — Type

ShiftMemory(Δ::Int) <: MemoryStrategy

Effectively replaces the fluctuating shift update procedure for the coefficient vector by an averaged shift over Δ timesteps, where Δ = 1 means no averaging.

Rimu.ShiftStrategy — Type

Abstract type for defining the strategy for updating the shift with update_shift(). Implemented strategies:

DoubleLogUpdate - default in lomc!()

DontUpdate

LogUpdate

DelayedLogUpdate

LogUpdateAfterTargetWalkers - FCIQMC standard

DelayedLogUpdateAfterTargetWalkers

DoubleLogUpdateAfterTargetWalkers

DoubleLogUpdateAfterTargetWalkersSwitch

HistoryLogUpdate

DoubleLogProjected

Rimu.StochasticStyle — Type

StochasticStyle(v)
StochasticStyle(typeof(v))

StochasticStyle specifies the native style of the generalised vector v that determines how simulations are to proceed. This can be fully stochastic (with IsStochastic), fully deterministic (with IsDeterministic), or semistochastic (with IsSemistochastic).

Rimu.ThresholdProject — Type

ThresholdProject(threshold = 1.0) <: ProjectStrategy

Project stochastically for walker amplitudes below threshold. See norm_project.

Rimu.TimeStepStrategy — Type

Abstract type for strategies for updating the time step with update_dτ(). Implemented strategies:

ConstantTimeStep

Base.length — Method

length(md::MPIData)

Compute the length of the distributed data on every MPI rank with MPI.Allreduce.

LinearAlgebra.norm — Function

norm(md::MPIData, p=2)

Compute the norm of the distributed data on every MPI rank with MPI.Allreduce.

Rimu.applyMemoryNoise! — Method

r = applyMemoryNoise!(w, v, shift, dτ, pnorm, m_strat::MemoryStrategy)

Apply memory noise to w, i.e. w .+= r.*v, computing the noise r according to m_strat. Note that m_strat needs to be compatible with StochasticStyle(w). Otherwise, an error exception is thrown. See MemoryStrategy.

w is the walker array after fciqmc step, v the previous one, pnorm the norm of v, and r the instantaneously applied noise.

Rimu.clearDFlags! — Method

clearDFlags!(dv)

Clear all flags in dv of the deterministic bit (rightmost bit).

Rimu.compute_proj_observables — Method

compute_proj_observables(v, ham, r::ReportingStrategy)

Compute the projection of r.projector⋅v and r.projector⋅ham*v according to the ReportingStrategy r.

Rimu.fciqmc! — Function

fciqmc!(v, pa::FciqmcRunStrategy, [df,]
         ham, s_strat::ShiftStrategy,
         [r_strat::ReportingStrategy, τ_strat::TimeStepStrategy, w])
-> df

Perform the FCIQMC algorithm for determining the lowest eigenvalue of ham. v can be a single starting vector of type :<AbstractDVec or a vector of such structures. In the latter case, independent replicas are constructed. Returns a DataFrame df with statistics about the run, or a tuple of DataFrames for a replica run. Strategies can be given for updating the shift (see ShiftStrategy) and (optionally), for reporting (see ReportingStrategy), and for updating the time step dτ (see TimeStepStrategy).

A pre-allocated data structure w for working memory can be passed as argument, and controls multi-threading behaviour. By default multi-threading is turned on. To turn multi-threading off, pass similar(localpart(v)) for w.

This function mutates v, the parameter struct pa as well as df, and w.

NOTE: The function fciqmc!() may be deprecated soon. Change all scripts to call lomc!() instead!

Rimu.fciqmc_col! — Method

fciqmc_col!(w, ham, add, num, shift, dτ)
fciqmc_col!(::Type{T}, args...)
-> spawns, deaths, clones, antiparticles, annihilations

Spawning and diagonal step of FCIQMC for single column of ham. In essence it computes

w .+= (1 .+ dτ.*(shift .- ham[:,add])).*num.

Depending on T ==StochasticStyle(w), a stochastic or deterministic algorithm will be chosen. The possible values for T are:

IsDeterministic() deteministic algorithm

IsStochastic() stochastic version where the changes added to w are purely integer, according to the FCIQMC algorithm

IsStochasticNonlinear(c) stochastic algorithm with nonlinear diagonal

IsSemistochastic() semistochastic version: TODO

Rimu.fciqmc_step! — Method

fciqmc_step!(Ĥ, v, shift, dτ, pnorm, w;
                      m_strat::MemoryStrategy = NoMemory()) -> ṽ, w̃, stats

Perform a single matrix(/operator)-vector multiplication:

\[\tilde{v} = [1 - dτ(\hat{H} - S)]⋅v ,\]

where Ĥ == ham and S == shift. Whether the operation is performed in stochastic, semistochastic, or determistic way is controlled by the trait StochasticStyle(w). See StochasticStyle. w is a local data structure with the same size and type as v and used for working. Both v and w are modified.

Returns the result ṽ, a (possibly changed) reference to working memory w̃, and the array stats = [spawns, deaths, clones, antiparticles, annihilations]. Stats will contain zeros when running in deterministic mode.

Rimu.free — Method

free(obj::MPIOSWin)

De-reference the object, call finalizer and the garbage collector immediately. This is a syncronizing MPI call. Make sure that the object is not used later.

Rimu.greet — Method

brief greeting

Rimu.lomc! — Method

lomc!(ham, v; kwargs...)
-> nt::NamedTuple

Linear operator Monte Carlo: Perform the FCIQMC algorithm for determining the lowest eigenvalue of ham. v can be a single starting vector of (wrapped) type :<AbstractDVec, or a vector of such structures for a replica simulation. Returns a named tuple containg all information required for continuation runs. In particular, nt.df is a DataFrame with statistics about the run, or a tuple of DataFrames for a replica run.

Keyword arguments, defaults, and precedence:

laststep - can be used to override information otherwise contained in params

threading = :auto - can be used to control the use of multithreading (overridden by wm)

:auto - use multithreading if s_strat.targetwalkers ≥ 500

true - use multithreading if available (set shell variable JULIA_NUM_THREADS!)

false - run on single thread

wm - working memory; if set, it controls the use of multithreading and overrides threading; is mutated

params::FciqmcRunStrategy = RunTillLastStep(laststep = 100) - contains basic parameters of simulation state, see FciqmcRunStrategy; is mutated

s_strat::ShiftStrategy = DoubleLogUpdate(targetwalkers = 1000) - see ShiftStrategy

r_strat::ReportingStrategy = EveryTimeStep() - see ReportingStrategy

τ_strat::TimeStepStrategy = ConstantTimeStep() - see TimeStepStrategy

m_strat::MemoryStrategy = NoMemory() - see MemoryStrategy

p_strat::ProjectStrategy = NoProjection() - see ProjectStrategy

Return values

nt = lomc!(args...)

The named tuple nt contains the following fields:

nt = (
    ham = ham, # the linear operator, from input
    v = v, # the current coefficient vector, mutated from input
    params = params, # struct with state parameters, mutated from input
    df = df, # DataFrame with statistics per time step
    wm = wm, # working memory, mutated from input
    s_strat = s_strat, # from input
    r_strat = r_strat, # from input
    τ_strat = τ_strat, # from input
    m_strat = m_strat, # from input
    p_strat = p_strat, # from input
)

Rimu.lomc! — Method

lomc!(nt::NamedTuple, laststep::Int = nt.params.laststep)

Linear operator Monte Carlo: Call signature for a continuation run.

nt should have the same structure as the return value of lomc!(). The optional argument laststep can be used to set a new last step. If laststep > nt.params.step, additional time steps will be computed and the statistics in the DataFrame nt.df will be appended.

Rimu.mpi_default — Function

mpi_default(data, comm = MPI.COMM_WORLD, root = 0)

Declare data as mpi-distributed and set communication strategy to default. Sets up the MPIData structure with MPIDefault strategy.

Rimu.mpi_no_exchange — Function

mpi_no_exchange(data, comm = MPI.COMM_WORLD, root = 0)

Declare data as mpi-distributed and set communication strategy to MPINoWalkerExchange. Sets up the MPIData structure with MPINoWalkerExchange strategy.

Rimu.mpi_one_sided — Function

mpi_one_sided(data, comm = MPI.COMM_WORLD, root = 0)

Declare data as mpi-distributed and set communication strategy to one-sided with remote memory access (RMA). Sets up the MPIData structure with MPIOSWin strategy.

Rimu.next_mpiID — Function

next_mpiID()

Produce a new ID number for MPI distributed objects. Uses an internal counter.

Rimu.norm_project! — Method

norm_project!(w, p_strat::ProjectStrategy) -> norm

Computes the 1-norm of w. Project all elements of w to s.threshold preserving the sign if StochasticStyle(w) requires projection according to p_strat. See ProjectStrategy.

Rimu.put — Method

put(buf::Vector{T}, [len,] targetrank, s::MPIOSWin{T})
put(obj::T, targetrank, s::MPIOSWin{T})

Deposit a single obj or vector buf into the MPI window s on rank targetrank. If len is given, only the first len elements are transmitted.

Rimu.report! — Method

report!(df::DataFrame, t::Tuple, s<:ReportingStrategy)

Record results in df and write informational messages according to strategy s. See ReportingStrategy.

Rimu.setSemistochastic! — Method

setSemistochastic!(dv, threshold::Float16, d_space)

Set the deterministic space for dv with threshold threshold, where d_space is a vector of addresses defining the the stochastic subspace.

Rimu.setThreshold — Method

setThreshold(dv, threshold)

Set a threshold for non-integer walker number FCIQMC. Technically, the function sets the trait StochasticStyle of the generalised vector dv to IsStochasticWithThreshold(threshold), where dv must be a type that supports floating point walker numbers. Also available as macro, see @setThreshold.

Example usage:

julia> dv = DVec(Dict(nearUniform(BoseFS{3,3})=>3.0))
julia> setThreshold(dv, 0.6)
IsStochasticWithThreshold(0.6f0)

Rimu.sort_into_targets! — Method

sort_into_targets!(target, source, stats) -> agg, wm, agg_stats

Aggregate coefficients from source to agg and from stats to agg_stats according to thread- or MPI-level parallelism. wm passes back a reference to working memory.

Rimu.sort_into_targets! — Method

sort_into_targets!(target::MPIData, source)

Distribute the entries of source to the target data structure such that all entries in the target dictionaries are on the process with the correct rank as controlled by targetrank().

Rimu.targetrank — Function

targetrank(key, np, hash = hash(key))

Compute the rank where the key belongs.

Rimu.update_dτ — Method

update_dτ(s<:TimeStepStrategy, dτ, args...) -> new dτ

Update the time step according to the strategy s.

Rimu.update_shift — Method

update_shift(s <: ShiftStrategy, shift, shiftMode, tnorm, pnorm, dτ, step, df, v_new, v_old)

Update the shift according to strategy s. See ShiftStrategy.

Rimu.@setDeterministic — Macro

@setDeterministic dv

A macro to undo the effect of [@setThreshold] and set the trait StochasticStyle of the generalised vector dv to IsDeterministic().

Rimu.@setThreshold — Macro

@setThreshold dv threshold

A macro to set a threshold for non-integer walker number FCIQMC. Technically, the macro sets the trait StochasticStyle of the generalised vector dv to IsStochasticWithThreshold(threshold), where dv must be a type that supports floating point walker numbers. Also available as function, see setThreshold.

Example usage:

julia> dv = DVec(Dict(nearUniform(BoseFS{3,3})=>3.0))
julia> @setThreshold dv 0.6
IsStochasticWithThreshold(0.6f0)

Reexported Submodules

Hamiltonians

Link to Module Hamiltionians.jl

BitStringAddresses

Rimu.BitStringAddresses.BSAdd128 — Type

BSAdd128 <: BitStringAddressType

Address type that encodes a bistring address in a UInt128.

Rimu.BitStringAddresses.BSAdd64 — Type

BSAdd64 <: BitStringAddressType

Address type that encodes a bistring address in a UInt64.

Rimu.BitStringAddresses.BStringAdd — Type

BStringAdd <: BitStringAddressType

Address type that encodes a bistring address in a one-dim. BitArray.

Rimu.BitStringAddresses.BitAdd — Type

BitAdd{I,B} <: BitStringAddressType
BitAdd(address::Integer, B)
BitAdd(chunks::T, B) where T<:Union{Tuple,SVector}
BitAdd{B}(address)

Address type that encodes a bistring address with B bits. The bits are stored efficiently as SVector of I chunks of type UInt64. The two-argument constructor is preferred due to safety (consistency checks). In hot loops there maybe gain from the (unsafe) parametric constructor. If an integer address is passed, its bit representation is used to initialize BitAdd. For large bit numbers, BigInt is convenient. BitAdd[i] will return bit i (as Bool), counting from right to left.

BitAdd{B}() creates a BitAdd with all ones.

zero(BitAdd{B}) creates a BitAdd with all zeros.

Note that no checking for ghost bits occurs when constructing BitAdd from SVector or Tuple. See bitadd(), check_consistency(), and remove_ghost_bits() methods!

Rimu.BitStringAddresses.BitStringAddressType — Type

BitStringAddressType

Abstract type for configuration addresses with the nature of a bitstring. A number of methods need to be implemented, in particular Base.isless(a,b), numBits(), numChunks().

Rimu.BitStringAddresses.BoseFS — Type

BoseFS{N,M,A} <: BosonicFockStateAddress <: BitStringAddressType
BoseFS(bs::A) where A <: BitAdd
BoseFS(bs::A, b)

Address type that represents a Fock state of N spinless bosons in M orbitals by wrapping a bitstring of type A. Orbitals are stored in reverse order, i.e. the first orbital in a BoseFS is stored rightmost in the bitstring bs. If the number of significant bits b is not encoded in A it must be passed as an argument (e.g. for BSAdd64 and BSAdd128).

Rimu.BitStringAddresses.BoseFS — Method

BoseFS(onr::T) where T<:Union{AbstractVector,Tuple}
BoseFS{BST}(onr::T)

Create BoseFS address from an occupation number representation, specifying the occupation number of each orbital. If a type BST is given it will define the underlying bit string type. Otherwise, the bit string type is chosen to fit the onr.

Rimu.BitStringAddresses.BosonicFockStateAddress — Type

BosonicFockStateAddress <: BitStringAddressType

Supertype representing a bosonic Fock state. Implement numModes() and numParticles().

Base.:<< — Method

<<(b::BitAdd,n::Integer)

Bitshift b to the left by n bits and fill from the right with zeros.

Base.:>>> — Method

>>>(b::BitAdd,n::Integer)

Bitshift b to the right by n bits and fill from the left with zeros.

Rimu.BitStringAddresses.bitaddr — Method

bitaddr(onr, Type)

Calculate a bitstring address from an occupation number representation the type of the bitstring address is passed as the second argument.

Rimu.BitStringAddresses.lbshr — Method

lbshr(c,k)

Apply logical bit shift to the right by k bits to c.

Rimu.BitStringAddresses.nearUniform — Method

nearUniform(BoseFS{N,M})
nearUniform(BoseFS{N,M,A}) -> bfs::BoseFS{N,M,A}

Create bosonic Fock state with near uniform occupation number of M modes with a total of N particles. Specifying the bit address type A is optional.

Examples

julia> nearUniform(BoseFS{7,5,BitAdd})
BoseFS{BitAdd}((2,2,1,1,1))

julia> nearUniform(BoseFS{7,5})
BoseFS{BSAdd64}((2,2,1,1,1))

Rimu.BitStringAddresses.nearUniformONR — Method

nearUniformONR(N, M) -> onr::SVector{M,Int}

Create occupation number representation onr distributing N particles in M modes in a close-to-uniform fashion with each orbital filled with at least N ÷ M particles and at most with N ÷ M + 1 particles.

Rimu.BitStringAddresses.numBits — Method

numBits(a)

Number of bit chunks representing a.

Rimu.BitStringAddresses.numChunks — Method

numChunks(a)

Number of 64-bit chunks representing a.

Rimu.BitStringAddresses.numModes — Method

numModes(a)

Number of modes represented by a.

Rimu.BitStringAddresses.numParticles — Method

numParticles(a)

Number of particles represented by a.

Rimu.BitStringAddresses.onr — Method

onr(address, m)

Compute and return the occupation number representation as an array of Int corresponding to the given address.

Rimu.BitStringAddresses.onr — Method

onr(bs)

Compute and return the occupation number representation of the bit string address bs as an SVector{M,Int}, where M is the number of orbitals.

Rimu.BitStringAddresses.remove_ghost_bits — Method

remove_ghost_bits(bs)

Remove set bits outside data field if any are present.

DictVectors

Rimu.DictVectors — Module

Module that provides data structures that behave similar to sparse vectors, but are indexed by arbitrary types (could be non-integers) similarly to dictionaries. The idea is to do linear algebra with data structures that are neither subtyped to AbstractVector nor to AbstractDict and are suitable for use with KrylovKit.jl. For this, the abstract type and interface AbstractDVec is provided, with the following concrete subtypes:

DVec

DFVec

FastDVec

Rimu.DictVectors.ADVPairsIterator — Type

ADVPairsIterator

Iterator type for pairs from a AbstractDVec.

Rimu.DictVectors.AbstractDVec — Type

DictVectors.AbstractDVec{K,V}

Abstract type for sparse vectors with valtype() V based on dictionary-like structures. They behave like vectors except that indexing is performed with an arbitrary keytype() K (no order required). getindex() returns zero(V) for any unknown key and zeroed elements should be deleted with delete!(). Thus, iteration typically returns only non-zero entries. In variance to AbstractVector, length() will return the number of non-zero elements, while capacity() returns the holding capacity (assumed fixed) of the object. Even though AbstractDVec is similar to both AbstractDict and AbstractVector, it is explicitly not subtyped to either. The aim is to create data structures suitable for FCIQMC and compatible with KrylovKit.jl.

Interface

The interface is similar to the AbstractDict interface. Implement what would be needed for the AbstractDict interface (setindex!, getindex, delete!, length, haskey, empty!, isempty) and, in addition:

capacity(dv): holding capacity

similar(dv [,Type])

iterate(): should return values of type V

pairs(): should return an iterator over key::K => content pairs. If content ≠ value::V then provide values() iterator as well!

Rimu.DictVectors.DFVec — Type

DFVec{K,V,F}(capacity) <: AbstractDVec{K,V <: Number}
DFVec(key => (value, flag); capacity)
DFVec(args...; capacity)
DFVec(d::Dict [, capacity])
DFVec(v::Vector{V} [, capacity])

Dictionary-based vector-like data structure with minimum capacity capacity for storing values and flags with keys. The values have type eltype(dv) == V and the flags have flagtype(dv) == F. Indexing is done with an arbitrary (in general non-integer) key with keytype(dv) == K. If the keyword argument capacity is passed then args are parsed as for Dict. When constructed from a Vector{V}, the keys will be integers ∈ [1, length(v)] and the flag zero(UInt16). See AbstractDVec. The method capacity() is defined but not a strict upper limit as Dict objects can expand.

Rimu.DictVectors.DVec — Type

DVec{K,V}(capacity) <: AbstractDVec{K,V}
DVec(key => value; capacity)
DVec(args...; capacity)
DVec(d::Dict [, capacity])
DVec(v::Vector{V} [, capacity])

Dictionary-based vector-like data structure with minimum capacity capacity for storing values with keys. The type of the values is eltype(dv) == V. Indexing is done with an arbitrary (in general non-integer) key with keytype(dv) == K. If the keyword argument capacity is passed then args are parsed as for Dict. When constructed from a Vector, the keys will be integers ∈ [1, length(v)]. See AbstractDVec. The method capacity() is defined but not a strict upper limit as Dict objects can expand.

Rimu.DictVectors.FastDVec — Type

DictVectors.FastDVec{K,V}(capacity)

Create a dictionary-like array indexed by keys of type K and values of type V with a maximum capacity of capacity.

FastDVec(a::AbstractArray, [capacity = length(a)])
FastDVec(d::AbstractDict{K,V}, [capacity = length(d)])
FastDVec(d::AbstractDVec{K,V}, [capacity = length(d)])

Construct a FastDVec object from an existing array or dictionary.

FastDVec(args...; capacity)

If the keyword argument capacity is passed then args are parsed as for Dict.

Rimu.DictVectors.FlagsIterator — Type

FlagsIterator

Iterator type returned by flags().

Rimu.DictVectors.KVPairsIterator — Type

KVPairsIterator

Iterator type returned by kvpairs().

Rimu.DictVectors.Norm2Projector — Type

Norm2Projector()

Results in computing the two-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,2) # with type Float64

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.NormProjector — Type

NormProjector()

Results in computing the one-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,1) # with type valtype(x)

NormProjector() thus represents the vector sign.(x).

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.UniformProjector — Type

UniformProjector()

Represents a vector with all elements 1. To be used with dot(). Minimizes memory allocations.

UniformProjector()⋅v == sum(v)
dot(UniformProjector(), LO, v) == sum(LO*v)

See also ReportingStrategy for use of projectors in FCIQMC.

Base.fill! — Method

fill!(da::AbstractDVec, x)

Empties da if x==zero(valtype(da)) and throws an error otherwise.

Base.get — Method

get(dv::DFVec, key, deftup::Tuple)

Return tuple containing value and flag if key exists and otherwise deftup.

Base.get — Method

get(dv::DFVec, key, defnum::Number)

Return value only if key exists and otherwise defnum

Base.isequal — Method

isequal(l::AbstractDVec, r::AbstractDVec)

Returns true if all non-zero entries have the same value. Equality of flags is not tested unless both l and r support flags.

Base.isequal — Method

isequal(l::DFVec, r::DFVec)

Returns true if all non-zero entries have the same value and the same flag.

LinearAlgebra.axpy! — Method

axpy!(α::Number, X::AbstractDVec, Y::AbstractDVec)
axpy!(α::Number, X::AbstractDVec, Ys::Tuple, batchsize)

Overwrite Y with α*X + Y where α is scalar for AbstractDVecs. If a tuple Ys is passed with Threads.nthreads() AbstractDVecs, then perform the operation in parallel over threads with batchsize elements at a time.

LinearAlgebra.norm — Method

norm(x::AbstractDVec, p)

Computes the p-norm of the DVec x. Implemented for p ∈ {1, 2, Inf}. Returns zero if x is empty.

LinearAlgebra.norm — Method

norm(x::AbstractDVec{K,V})

Computes the 2-norm of the DVec x.

Rimu.DictVectors.add! — Method

add!(x::AbstractDVec,y::AbstactDVec)

Inplace add x+y and store result in x.

Rimu.DictVectors.capacity — Function

DictVectors.capacity(dv::AbstractDVec, [s = :effective])
capacity(dvs::Tuple, [s = :effective])

gives the effective holding capacity of dv. If a tuple of dvs is given it aggregates the capacities.

Optional argument s:

:effective the number of elements that

can be stored savely (default)

:allocated actual internal memory allocation

Rimu.DictVectors.flags — Method

flags(dv::DFVec)

An iterator that yields flags stored in the DFVec dv.

Rimu.DictVectors.gettuple — Method

gettuple(dv::DFVec, key)

Retrieve the tuple (v,f) stored at the given key. If the key is not found, return a tuple of zeros.

Rimu.DictVectors.kvpairs — Method

kvpairs(collection)

Return an iterator over key => value pairs ignoring any flags. If no flags are present, eg. for generic AbstractDVec, this falls back to Base.pairs.

Rimu.DictVectors.kvpairs — Method

kvpairs(dv::DFVec)

An iterator that yields key => value pairs stored in the DFVec dv ignoring any flags. In contrast, Base.pairs() will return pairs key => (value, flag).

Rimu.DictVectors.normInf — Method

normInf(x::AbstractDVec)

Infinity norm: largest absolute value of entries.

Rimu.DictVectors.norm_sqr — Method

norm_sqr(x::AbstractDVec)

Fast calculation of the square of the 2-norm of x.

Rimu.DictVectors.pairtype — Method

pairtype(dv)

Returns the type of stored data, as returned by the pairs() iterator.

Rimu.DictVectors.tuples — Method

tuples(dv::DFVec)

An iterator that yields tuples (v, f) with value v and flag f stored in the DFVec dv.

Rimu.DictVectors.zero! — Method

zero!(v::AbstractDVec)

Replace v by a zero vector as an inplace operation. For AbstractDVec types it means removing all non-zero elements.

Blocking

Rimu.Blocking — Module

Blocking

Module that contains functions performing the Flyvbjerg-Petersen (J. Chem. Phys. 91, 461 (1989)) blocking analysis for evaluating the standard error on a correlated data set. A "M-test" is also implemented based on Jonsson (Phys. Rev. E 98, 043304, (2018)).

Rimu.Blocking.autoblock — Method

autoblock(df::DataFrame; start = 1, stop = size(df)[1])
-> s̄, σs, ē, σe, k

Determine mean shift s̄ and projected energy ē with respective standard errors σs and σe by blocking analsis from the DataFrame df returned from fciqmc!(). The number k of blocking steps and decorrelation time 2^k are obtained from the M-test for the shift and also applied to the projected energy, assuming that the projected quantities decorrelate on the same time scale. Returns a named tuple.

Rimu.Blocking.autoblock — Method

autoblock(dftup::Tuple; start = 1, stop = size(dftup[1])[1])
-> s̄1, σs1, s̄2, σs2, ē1, σe1, ē2, σe2, ēH, σeH, k

Replica version. dftup is the tuple of DataFrames returned from replica fciqmc!(). Returns a named tuple with shifts and three variational energy estimators and respective errors obtained from blocking analysis. The larger of the k values from M-tests on the two shift time series is used.

Rimu.Blocking.autocovariance — Method

autocovariance(v::Vector,h::Int; corrected::Bool=true)

$\hat{\gamma}(h) =\frac{1}{n}\sum_{t=1}^{n-h}(v_{t+h}-\bar{v})(v_t-\bar{v})$ Calculate the autocovariance of dataset v with a delay h. If corrected is true (the default) then the sum is scaled with n-h, whereas the sum is scaled with n if corrected is false where n = length(v).

Rimu.Blocking.blockAndMTest — Method

v̄, σ, σσ, k, df = blockAndMTest(v::Vector)

Perform a blocking analysis and M-test on v returning the mean v̄, standard error σ, its error σσ, the number of blocking steps k, and the DataFrame df with blocking data.

Rimu.Blocking.blocker — Method

blocker(v::Vector) -> new_v::Vector

Reblock the data by successively taking the mean of two adjacent data points to form a new vector with a half of the length(v). The last data point will be discarded if length(v) is odd.

Rimu.Blocking.blocking — Method

blocking(x::Vector,y::Vector) -> df::DataFrame

Perform a blocking analysis for the quotient of means x̄/ȳ from two data sets. If corrected is true (the default) then the sums in both variance and covariance are scaled with n-1, whereas the sums are scaled with n if corrected is false where n = length(x) = length(y). Entries in returned dataframe:

blocks = number of blocks in current blocking step;

mean_x, SD_x, SE_x, SE_SE_x = the mean, standard deviation, standard error and error on standard error estimated for dataset x;

mean_y, SD_y, SE_y, SE_SE_y = ditto. for dataset y;

Covariance = the covariance between data in x and y;

mean_f = x̄/ȳ;

SE_f = standard error estimated for x̄/ȳ.

Rimu.Blocking.blocking — Method

blocking(v::Vector; corrected::Bool=true) -> df

Perform a blocking analysis according to Flyvberg and Peterson JCP (1989) for single data set and return a DataFrame with statistical data for each blocking step. M-test data according to Jonsson PRE (2018) is also provided. If corrected is true (the default) then the sum in var is scaled with n-1 and in autocovariance is scaled with n-h, whereas the sum is scaled with n for both if corrected is false where n = length(v).

Rimu.Blocking.blocking_old — Method

blocking(v::Vector; typos = nothing) -> df

Perform a blocking analysis according to Flyvberg and Peterson JCP (1989) for single data set and return a DataFrame with statistical data for each blocking step. M-test data according to Jonsson PRE (2018) is also provided.

Keyword argument typos

typos = nothing - correct all presumed typos.

typos = :FP - use Flyvberg and Peterson (correct) standard error and Jonsson formul for M.

typos = :Jonsson - calculate M and standard error as written in Jonsson.

Rimu.Blocking.combination_division — Method

combination_division(x::Vector,y::Vector; corrected::Bool=true)

Find the standard error on the quotient of means x̄/ȳ from two data sets, note that the standard errors are different on $(x̄/ȳ) \neq \bar{(\frac{x}{y})}$. If corrected is true (the default) then the sums in both variance and covariance are scaled with n-1, whereas the sums are scaled with n if corrected is false where n = length(x) = length(y).

Rimu.Blocking.covariance — Method

covariance(x::Vector,y::Vector; corrected::Bool=true)

Calculate the covariance between the two data sets x and y with equal length. If corrected is true (the default) then the sum is scaled with n-1, whereas the sum is scaled with n if corrected is false where n = length(x) = length(y).

Rimu.Blocking.growthWitness — Method

growthWitness(norm::AbstractArray, shift::AbstractArray, dt; b = 30, pad = :true) -> g
growthWitness(df::DataFrame; b = 30, pad = :true) -> g

Compute the growth witness

\[G_b^{(n)} = S̄^{(n)} - \frac{\log\vert\mathbf{c}^{(n+b)}\vert - \log\vert\mathbf{c}^{(n)}\vert}{b d\tau},\]

where S̄ is an average of the shift over b time steps and $\vert\mathbf{c}^{(n)}\vert ==$ norm[n]. The parameter b ≥ 1 averages the derivative quantity over b time steps and helps suppress noise.

If pad is set to :false then the returned array g has the length length(norm) - b. If set to :true then g will be padded up to the same length as norm and shift.

Rimu.Blocking.mtest — Method

mtest(df::DataFrame; warn = true) -> k

The "M test" based on Jonsson, M. Physical Review E, 98(4), 043304, (2018). Expects df to be output of a blocking analysis with column df.M containing relevant M_j values, which are compared to a χ^2 distribution. Returns the row number k where the M-test is passed. If the M-test has failed mtest() returns the value -1 and optionally prints a warning message.

Rimu.Blocking.se — Method

se(v::Vector;corrected::Bool=true)

Calculate the standard error of the dataset v. If corrected is true (the default) then the sum in std is scaled with n-1, whereas the sum is scaled with n if corrected is false where n = length(v).

Unexported Submodules

FastBufs

Rimu.FastBufs — Module

Provides the FastBuf data structure.

Rimu.FastBufs.FastBuf — Type

FastBuf{T}(n)

Create a first-in-last-out buffer of maximum capacity n. The element type is T. Fill a FastBuf with push!(), retrieve with pop!(). Accessing elements with getindex() and setindex!() as well as iteration are also supported. Check the capacity of a FastBuf with capacity().

Rimu.FastBufs.capacity — Method

Gives the capacity of a FastBuf object.

ConsistentRNG

Link to Module ConsistentRNG.jl

Index

Rimu.Blocking

Rimu.ConsistentRNG

Rimu.DictVectors

Rimu.FastBufs

Rimu.Hamiltonians

Rimu.Rimu

Rimu.BitStringAddresses.BSAdd128

Rimu.BitStringAddresses.BSAdd64

Rimu.BitStringAddresses.BStringAdd

Rimu.BitStringAddresses.BitAdd

Rimu.BitStringAddresses.BitStringAddressType

Rimu.BitStringAddresses.BoseFS

Rimu.BitStringAddresses.BoseFS

Rimu.BitStringAddresses.BosonicFockStateAddress

Rimu.ConsistentRNG.CRNG

Rimu.ConstantTimeStep

Rimu.DelayedDoubleLogUpdate

Rimu.DelayedDoubleLogUpdateAfterTW

Rimu.DelayedLogUpdate

Rimu.DelayedLogUpdateAfterTargetWalkers

Rimu.DeltaMemory

Rimu.DeltaMemory2

Rimu.DeltaMemory3

Rimu.DictVectors.ADVPairsIterator

Rimu.DictVectors.AbstractDVec

Rimu.DictVectors.DFVec

Rimu.DictVectors.DVec

Rimu.DictVectors.FastDVec

Rimu.DictVectors.FlagsIterator

Rimu.DictVectors.KVPairsIterator

Rimu.DictVectors.Norm2Projector

Rimu.DictVectors.NormProjector

Rimu.DictVectors.UniformProjector

Rimu.DontUpdate

Rimu.DoubleLogProjected

Rimu.DoubleLogUpdate

Rimu.DoubleLogUpdateAfterTargetWalkers

Rimu.DoubleLogUpdateAfterTargetWalkersSwitch

Rimu.EveryKthStep

Rimu.EveryTimeStep

Rimu.FastBufs.FastBuf

Rimu.FciqmcRunStrategy

Rimu.Hamiltonians.AbstractHamiltonian

Rimu.Hamiltonians.BoseHubbardMom1D

Rimu.Hamiltonians.BoseHubbardReal1D

Rimu.Hamiltonians.BosonicHamiltonian

Rimu.Hamiltonians.ExtendedBHReal1D

Rimu.Hamiltonians.Hops

Rimu.HistoryLogUpdate

Rimu.IsSemistochastic

Rimu.IsStochasticWithThreshold

Rimu.LogUpdate

Rimu.LogUpdateAfterTargetWalkers

Rimu.MPIData

Rimu.MPIDefault

Rimu.MPINoWalkerExchange

Rimu.MPIOSWin

Rimu.MemoryStrategy

Rimu.NoMemory

Rimu.NoProjection

Rimu.NoProjectionTwoNorm

Rimu.ProjectStrategy

Rimu.ProjectedMemory

Rimu.ReportDFAndInfo

Rimu.ReportingStrategy

Rimu.RunTillLastStep

Rimu.ScaledThresholdProject

Rimu.ShiftMemory

Rimu.ShiftStrategy

Rimu.StochasticStyle

Rimu.ThresholdProject

Rimu.TimeStepStrategy

Base.:<<

Base.:>>>

Base.fill!

Base.get

Base.get

Base.isequal

Base.isequal

Base.length

LinearAlgebra.axpy!

LinearAlgebra.norm

LinearAlgebra.norm

LinearAlgebra.norm

Rimu.BitStringAddresses.bitaddr

Rimu.BitStringAddresses.lbshr

Rimu.BitStringAddresses.nearUniform

Rimu.BitStringAddresses.nearUniform

Rimu.BitStringAddresses.nearUniformONR

Rimu.BitStringAddresses.numBits

Rimu.BitStringAddresses.numChunks

Rimu.BitStringAddresses.numModes

Rimu.BitStringAddresses.numParticles

Rimu.BitStringAddresses.onr

Rimu.BitStringAddresses.onr

Rimu.BitStringAddresses.remove_ghost_bits

Rimu.Blocking.autoblock

Rimu.Blocking.autoblock

Rimu.Blocking.autocovariance

Rimu.Blocking.blockAndMTest

Rimu.Blocking.blocker

Rimu.Blocking.blocking

Rimu.Blocking.blocking

Rimu.Blocking.blocking_old

Rimu.Blocking.combination_division

Rimu.Blocking.covariance

Rimu.Blocking.growthWitness

Rimu.Blocking.mtest

Rimu.Blocking.se

Rimu.ConsistentRNG.cRand

Rimu.ConsistentRNG.newChildRNG

Rimu.ConsistentRNG.seedCRNG!

Rimu.ConsistentRNG.trng

Rimu.DictVectors.add!

Rimu.DictVectors.capacity

Rimu.DictVectors.flags

Rimu.DictVectors.gettuple

Rimu.DictVectors.kvpairs

Rimu.DictVectors.kvpairs

Rimu.DictVectors.normInf

Rimu.DictVectors.norm_sqr

Rimu.DictVectors.pairtype

Rimu.DictVectors.tuples

Rimu.DictVectors.zero!

Rimu.FastBufs.capacity

Rimu.Hamiltonians.bit_String_Length

Rimu.Hamiltonians.diagME

Rimu.Hamiltonians.dimensionLO

Rimu.Hamiltonians.fDimensionLO

Rimu.Hamiltonians.generateRandHop

Rimu.Hamiltonians.hasIntDimension

Rimu.Hamiltonians.hop

Rimu.Hamiltonians.numOfHops

Rimu.applyMemoryNoise!

Rimu.clearDFlags!

Rimu.compute_proj_observables

Rimu.fciqmc!

Rimu.fciqmc_col!

Rimu.fciqmc_step!

Rimu.free

Rimu.greet

Rimu.lomc!

Rimu.lomc!

Rimu.mpi_default

Rimu.mpi_no_exchange

Rimu.mpi_one_sided

Rimu.next_mpiID

Rimu.norm_project!

Rimu.put

Rimu.report!

Rimu.setSemistochastic!

Rimu.setThreshold

Rimu.sort_into_targets!

Rimu.sort_into_targets!

Rimu.targetrank

Rimu.update_dτ

Rimu.update_shift

Rimu.@setDeterministic

Rimu.@setThreshold