API

Rimu

Rimu.Rimu — Module

Rimu

Random Integrator for Many-Body Quantum Systems

Rimu.AllOverlaps — Type

AllOverlaps(n=2, operator=nothing) <: ReplicaStrategy{n}

Run n replicas and report overlaps between all pairs of replica vectors. If operator is not nothing, the overlap dot(c1, operator, c2) is reported as well.

Column names in the report are of the form c{i}dotc{j} for vector-vector overlaps, and c{i}Opc{j} for operator overlaps.

See ReplicaStrategy and AbstractHamiltonian (for an interface for implementing operators).

Rimu.ConstantTimeStep — Type

Keep dτ constant.

Rimu.DelayedDoubleLogUpdate — Type

DelayedDoubleLogUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, A=10) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ and delay of A steps. See DoubleLogUpdate.

\[S^{n+A} = S^n -\frac{ζ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^n}\right)-\frac{ξ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^\text{target}}\right)\]

Rimu.DelayedDoubleLogUpdateAfterTW — Type

DelayedDoubleLogUpdateAfterTW(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, A=10) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ and delay of A steps after the number of target walkers is reached. See DoubleLogUpdate.

\[S^{n+A} = S^n -\frac{ζ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^n}\right)-\frac{ξ}{A dτ}\ln\left(\frac{\|Ψ\|_1^{n+A}}{\|Ψ\|_1^\text{target}}\right)\]

Rimu.DelayedLogUpdate — Type

DelayedLogUpdate(ζ = 0.08, a = 10) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and delay of a steps.

\[S^{n+a} = S^n -\frac{ζ}{a dτ}\ln\left(\frac{\|Ψ\|_1^{n+a}}{\|Ψ\|_1^n}\right)\]

Rimu.DelayedLogUpdateAfterTargetWalkers — Type

DelayedLogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08, a = 10) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and delay of a steps. See DelayedLogUpdate.

Rimu.DeltaMemory — Type

DeltaMemory(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = (pnorm - tnorm)/(dτ*pnorm) + shift
r = r̃ - <r̃>

Rimu.DeltaMemory2 — Type

DeltaMemory2(Δ::Int) <: MemoryStrategy

Before updating the shift, memory noise with a memory length of Δ is applied, where Δ = 1 means no memory noise.

r̃ = pnorm - tnorm + shift*dτ*pnorm
r = (r̃ - <r̃>)/(dτ*pnorm)

The long-term average of r is not guaranteed to be zero.

Rimu.DeltaMemory3 — Type

DeltaMemory3(Δ::Int, level::Float64) <: MemoryStrategy

Before updating the shift, apply multiplicative memory noise with a memory length of Δ at level level, where Δ = 1 means no memory noise.

r̃ = (pnorm - tnorm)/pnorm + dτ*shift
r = r̃ - <r̃>
w .*= 1 + level*r

Rimu.DontUpdate — Type

DontUpdate() <: ShiftStrategy Don't update the shift.

Rimu.DoubleLogProjected — Type

DoubleLogProjected(; target, projector, ζ = 0.08, ξ = ζ^2/4) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ after projecting onto projector.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{P⋅Ψ^{(n+1)}}{P⋅Ψ^{(n)}}\right)-\frac{ξ}{dτ}\ln\left(\frac{P⋅Ψ^{(n+1)}}{\text{target}}\right)\]

Rimu.DoubleLogSumUpdate — Type

DoubleLogSumUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, α = 1/2) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameters ζ and ξ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^n}\right) - \frac{ξ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^\text{target}}\right),\]

where $N_\mathrm{w} =$ (1-α)*walkernumber() + α*UniformProjector()⋅ψ computed with walkernumber() and UniformProjector(). When ξ = ζ^2/4 this corresponds to critical damping with a damping time scale T = 2/ζ.

Rimu.DoubleLogUpdate — Type

DoubleLogUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ and ξ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)-\frac{ξ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^\text{target}}\right)\]

When ξ = ζ^2/4 this corresponds to critical damping with a damping time scale T = 2/ζ.

Rimu.DoubleLogUpdateAfterTargetWalkers — Type

LogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08, ξ = 0.0016) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and ξ. See DoubleLogUpdate.

Rimu.DoubleLogUpdateAfterTargetWalkersSwitch — Type

LogUpdateAfterTargetWalkersSwitch(targetwalkers, ζ = 0.08, ξ = 0.0016) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ and ξ. After a steps the strategy switches to LogUpdate. See DoubleLogUpdate.

Rimu.EveryKthStep — Type

EveryKthStep(;k = 10, projector = nothing, hproj = :auto)

Report every kth step. Include projection onto projector. See ReportingStrategy for details.

Rimu.EveryTimeStep — Type

EveryTimeStep(;projector = nothing, hproj = :auto)

Report every time step. Include projection onto projector. See ReportingStrategy for details.

Rimu.FciqmcRunStrategy — Type

 FciqmcRunStrategy{T}

Abstract type representing the strategy for running and terminating fciqmc!(). The type parameter T is relevant for reporting the shift and the norm.

Implemented strategies:

RunTillLastStep

Rimu.HistoryLogUpdate — Type

HistoryLogUpdate(df::DataFrame; d = 100, k=1, ζ= 0.08)

Strategy for updating the shift according to log formula but with walker numbers accumulated from k samples of the history with delay d. A recent history has to be passed with the data frame df for initialisation.

\[N_w^{n} = \sum_{i=0}^{k-1} \|Ψ\|_1^{n-i} \\ S^{n+a} = S^n -\frac{ζ}{a dτ}\ln\left(\frac{N_w^{n+1}}{N_w^n}\right)\]

Rimu.LogUpdate — Type

LogUpdate(ζ = 0.08) <: ShiftStrategy

Strategy for updating the shift according to the log formula with damping parameter ζ.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{\|Ψ\|_1^{n+1}}{\|Ψ\|_1^n}\right)\]

Rimu.LogUpdateAfterTargetWalkers — Type

LogUpdateAfterTargetWalkers(targetwalkers, ζ = 0.08) <: ShiftStrategy

Strategy for updating the shift: After targetwalkers is reached, update the shift according to the log formula with damping parameter ζ. See LogUpdate.

Rimu.MemoryStrategy — Type

Abstract type for defining the strategy for injectimg memory noise. Implemented strategies:

NoMemory

DeltaMemory

ShiftMemory

Rimu.NoMemory — Type

NoMemory <: MemoryStrategy

Default strategy for MemoryStrategy indicating that no memory noise will be used.

Rimu.NoStats — Type

NoStats(N=1) <: ReplicaStrategy{N}

The default ReplicaStrategy. N replicas are run, but no statistics are collected.

Rimu.ProjectedMemory — Type

ProjectedMemory(Δ::Int, projector, pp::Number) <: MemoryStrategy
ProjectedMemory(Δ::Int, projector, v::AbstractDVec)

Before updating the shift, apply memory noise to minimize the fluctuations of the overlap of the coefficient vector with projector. Averaging over Δ time steps is applied, where Δ = 1 means no memory noise is applied. Use pp to initialise the value of the projection or pass v in order to initialise the projection with pp = projector.v.

r̃ = (projector⋅v - projector⋅w)/projector⋅v + dτ*shift
r = r̃ - <r̃>

where v is the coefficient vector before and w after applying a regular FCIQMC step.

Rimu.PurgeNegatives — Type

PurgeNegatives <: MemoryStrategy

Purge all negative sign walkers.

Rimu.QMCState — Type

QMCState

Holds all inforamtion needed to run FCIQMC, except the data frame. Holds a NTuple of ReplicaStates and various strategies that control the algorithm.

Rimu.ReplicaState — Type

ReplicaState(hamiltonian, v, w, pnorm, r_strat)

Struct that holds all information needed for an independent run of the algorithm.

Can be advanced a step forward with advance!.

Rimu.ReplicaStrategy — Type

ReplicaStrategy{N}

An abstract type that controles how lomc! uses replicas. A subtype of ReplicaStrategy{N} operates on N replicas and must implement the following function:

replica_stats(::ReplicaStrategy{N}, ::NTuple{N,ReplicaState}) - return a tuple of Strings or Symbols of replica statistic names and a tuple of the values. These will be reported to the DataFrame returned by lomc!

Concrete implementations:

NoStats: run (possibly one) replica(s), but don't report any additional info.

AllOverlaps: report overlaps between all pairs of replica vectors.

Rimu.Report — Type

Report

Internal structure that hold the temporary reported values. See report!.

Rimu.ReportDFAndInfo — Type

ReportDFAndInfo(; k=10, i=100, io=stdout, writeinfo=true, projector = nothing, hproj = :auto)

Report every kth step in DataFrame and write info message to io every ith step (unless writeinfo == false). The flag writeinfo is useful for controlling info messages in MPI codes. Include projection onto projector. See ReportingStrategy for details.

Rimu.ReportingStrategy — Type

ReportingStrategy

Abstract type for strategies for reporting data in a DataFrame with report!(). It also affects the calculation and reporting of projected quantities in the DataFrame.

Implemented strategies:

EveryTimeStep

EveryKthStep

ReportDFAndInfo

Every strategy accepts the keyword arguments projector and hproj according to which a projection of the instantaneous coefficient vector projector⋅v and hproj⋅v are reported to the DataFrame in the fields df.vproj and df.hproj, respectively. Possible values for projector are

nothing - no projections are computed (default)

dv::AbstractDVec - compute projection onto coefficient vector dv (set up with copy to conserve memory)

UniformProjector() - projection onto vector of all ones (i.e. sum of elements)

NormProjector() - compute 1-norm (instead of projection)

Norm1ProjectorPPop() - compute 1-norm per population

Norm2Projector() - compute 2-norm

In order to help set up the calculation of the projected energy, where df.hproj should report dot(projector, ham, v), the keyword hproj accepts the following values (for ReportingStrategys passed to lomc!()):

:auto - choose method depending on projector and ham (default)

:lazy - compute dot(projector, ham, v) every time (slow)

:eager - precompute hproj as ham'*v (fast, requires adjoint(ham))

:not - don't compute second projector (equivalent to nothing)

Interface

A ReportingStrategy must define the following:

report!

print_report (optional)

Examples

r_strat = EveryTimeStep(projector = copy(svec))

Record the projected energy components df.vproj = svec⋅v and df.hproj = dot(svec,ham,v) with respect to the starting vector (performs fast eager calculation if Hamiltonians.LOStructure(ham) ≠ Hamiltonians.AdjointUnknown()), and report every time step.

r_strat = EveryKthStep(k=10, projector = UniformProjector(), hproj = :lazy)

Record the projection of the instananeous coefficient vector v onto the uniform vector of all 1s into df.vproj and of ham⋅v into df.hproj, and report every kth time step.

Rimu.RunTillLastStep — Type

RunTillLastStep(step::Int = 0 # number of current/starting timestep
             laststep::Int = 100 # number of final timestep
             shiftMode::Bool = false # whether to adjust shift
             shift = 0.0 # starting/current value of shift
             dτ::Float64 = 0.01 # current value of time step
) <: FciqmcRunStrategy

Parameters for running fciqmc!() for a fixed number of time steps. For alternative strategies, see FciqmcRunStrategy.

Rimu.ShiftMemory — Type

ShiftMemory(Δ::Int) <: MemoryStrategy

Effectively replaces the fluctuating shift update procedure for the coefficient vector by an averaged shift over Δ timesteps, where Δ = 1 means no averaging.

Rimu.ShiftStrategy — Type

Abstract type for defining the strategy for updating the shift with update_shift(). Implemented strategies:

DoubleLogUpdate - default in lomc!()

DontUpdate

LogUpdate

DelayedLogUpdate

LogUpdateAfterTargetWalkers - FCIQMC standard

DelayedLogUpdateAfterTargetWalkers

DoubleLogUpdateAfterTargetWalkers

DoubleLogUpdateAfterTargetWalkersSwitch

HistoryLogUpdate

DoubleLogProjected

Rimu.TimeStepStrategy — Type

Abstract type for strategies for updating the time step with update_dτ(). Implemented strategies:

ConstantTimeStep

Rimu.TripleLogUpdate — Type

TripleLogUpdate(; targetwalkers = 1000, ζ = 0.08, ξ = ζ^2/4, η = 0.01) <: ShiftStrategy

Strategy for updating the shift according to the extended log formula with damping parameters ζ, ξ, and η.

\[S^{n+1} = S^n -\frac{ζ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^n}\right) - \frac{ξ}{dτ}\ln\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^\text{target}}\right) - \frac{η}{dτ}\ln\left(\frac{\|ℜ(Ψ^{n+1})\|_1^2 + \|ℑ(Ψ^{n+1})\|_1^2} {\|ℜ(Ψ^{n})\|_1^2 + \|ℑ(Ψ^{n})\|_1^2}\right),\]

where $N_\mathrm{w}$ is the walkernumber(). When ξ = ζ^2/4 this corresponds to critical damping with a damping time scale T = 2/ζ.

Rimu.advance! — Method

advance!(report::Report, state::QMCState, replica::ReplicaState)

Advance the replica by one step. The state is used only to access the various strategies involved. Steps, stats, and computed quantities are written to the report.

Returns true if the step was successful.

Rimu.apply_memory_noise! — Method

r = apply_memory_noise!(w, v, shift, dτ, pnorm, m_strat::MemoryStrategy)

Apply memory noise to w, i.e. w .+= r.*v, computing the noise r according to m_strat. Note that m_strat needs to be compatible with StochasticStyle(w). Otherwise, an error exception is thrown. See MemoryStrategy.

w is the walker array after fciqmc step, v the previous one, pnorm the norm of v, and r the instantaneously applied noise.

Rimu.compute_proj_observables — Method

compute_proj_observables(v, ham, r::ReportingStrategy)

Compute the projection of r.projector⋅v and r.hproj⋅v or r.projector⋅ham*v according to the ReportingStrategy r.

Rimu.fciqmc_col! — Method

fciqmc_col!(w, ham, add, num, shift, dτ)
fciqmc_col!(::Type{T}, args...)
-> spawns, deaths, clones, antiparticles, annihilations

Spawning and diagonal step of FCIQMC for single column of ham. In essence it computes

w .+= (1 .+ dτ.*(shift .- ham[:,add])).*num.

Depending on T ==StochasticStyle(w), a stochastic or deterministic algorithm will be chosen. The possible values for T are:

IsDeterministic() deteministic algorithm

IsStochasticInteger() stochastic version where the changes added to w are purely integer, according to the FCIQMC algorithm

IsStochasticWithThreshold(c) stochastic algorithm with floating point walkers.

Rimu.fciqmc_step! — Function

fciqmc_step!(Ĥ, v, shift, dτ, pnorm, w;
                      m_strat::MemoryStrategy = NoMemory()) -> ṽ, w̃, stats

Perform a single matrix(/operator)-vector multiplication:

\[\tilde{v} = [1 - dτ(\hat{H} - S)]⋅v ,\]

where Ĥ == ham and S == shift. Whether the operation is performed in stochastic, semistochastic, or determistic way is controlled by the trait StochasticStyle(w). See StochasticStyle. w is a local data structure with the same size and type as v and used for working. Both v and w are modified.

Returns the result ṽ, a (possibly changed) reference to working memory w̃, and the array stats = [spawns, deaths, clones, antiparticles, annihilations]. Stats will contain zeros when running in deterministic mode.

Rimu.lomc! — Method

lomc!(ham, v; kwargs...)

Linear operator Monte Carlo: Perform the FCIQMC algorithm for determining the lowest eigenvalue of ham. v can be a single starting vector of (wrapped) type :<AbstractDVec.

Returns a DataFrame with various statistics and a QMCState containing all information required for continuation runs.

Keyword arguments, defaults, and precedence:

laststep - can be used to override information otherwise contained in params

threading = :auto - can be used to control the use of multithreading (overridden by wm)

:auto - use multithreading if s_strat.targetwalkers ≥ 500

true - use multithreading if available (set shell variable JULIA_NUM_THREADS!)

false - run on single thread

wm - working memory; if set, it controls the use of multithreading and overrides threading; is mutated

params::FciqmcRunStrategy = RunTillLastStep(laststep = 100) - contains basic parameters of simulation state, see FciqmcRunStrategy; is mutated

s_strat::ShiftStrategy = DoubleLogUpdate(targetwalkers = 1000) - see ShiftStrategy

r_strat::ReportingStrategy = EveryTimeStep() - see ReportingStrategy

τ_strat::TimeStepStrategy = ConstantTimeStep() - see TimeStepStrategy

m_strat::MemoryStrategy = NoMemory() - see MemoryStrategy

replica::ReplicaStrategy = NoStats(1) - see ReplicaStrategy.

Return values

lomc! returns a named tuple with the following fields:

df: a DataFrame with all statistics being reported.

state: a QMCState that can be used for continuations.

Example

julia> add = BoseFS((1,2,3));

julia> H = HubbardReal1D(add);

julia> dv = DVec(add => 1);

julia> df1, state = lomc!(H, dv);

julia> df2, _ = lomc!(state, df1; laststep=200); # Contuniation run

julia> size(df1)
(100, 12)

julia> size(df2)
(200, 12)

Rimu.print_report — Method

print_report(::ReportingStrategy, step, report, state)

This function is called at the very end of a step. It can let the ReportingStrategy print some information to output.

Rimu.refine_r_strat — Method

Rimu.refine_r_strat(r_strat::ReportingStrategy, ham)

Refine the reporting strategy by replacing Symbols in the keyword argument hproj by the appropriate value. See ReportingStrategy

Rimu.replica_stats — Function

replica_stats(::ReplicaStrategy{N}, replicas::NTuple{N,ReplicaState}) -> (names, values)

Return the names and values of statistics reported by ReplicaStrategy. names should be a tuple of Symbols or Strings and values should be a tuple of the same length.

Rimu.report! — Method

report!(report, keys, values, id="")
report!(report, nt, id="")

Write keys, values pairs to report that will be converted to a DataFrame later. Alternatively, a named tuple can be passed instead of keys and values.

The value of id is appended to the name of the column, e.g. report!(report, :key, value, :_1) will report value to a column named :key_1.

Rimu.report! — Method

 report!(::ReportingStrategy, step, report::Report, keys, values, id="")
 report!(::ReportingStrategy, step, report::Report, nt, id="")

Report keys and values to report, which will be converted to a DataFrame before lomc! exits. Alternatively, a nt::NamedTuple can be passed in place of keys and values. If id is specified, it is appended to all keys. This is used to differentiate between values reported by different replicas.

To overload this function for a new ReportingStrategy, overload report!(::ReportingStrategy, step, args...) and apply the report by calling report!(args...).

Rimu.sort_into_targets! — Method

sort_into_targets!(target, source, stats) -> agg, wm, agg_stats

Aggregate coefficients from source to agg and from stats to agg_stats according to thread- or MPI-level parallelism. wm passes back a reference to working memory.

Rimu.step_stats — Method

step_stats(::StochasticStyle)

Return a tuple of names (Symbol or String) and a zeros of values of the same length. These will be reported as columns in the DataFrame returned by lomc!.

Rimu.threshold_projected_deposit! — Method

threshold_projected_deposit!

This function performs threshold projection before spawning, but only for IsDynamicSemistochastic with the project_later parameter set to false.

Rimu.update_dvec! — Method

update_dvec!([::StochasticStyle,] dvec) -> dvec, nt

Perform an arbitrary transformation on dvec after the spawning step is completed and report statistics to the DataFrame.

Returns the new dvec and a NamedTuple nt of statistics to be reported.

When extending this function for a custom StochasticStyle, define a method for the two-argument call signature!

Rimu.update_dτ — Method

update_dτ(s<:TimeStepStrategy, dτ, args...) -> new dτ

Update the time step according to the strategy s.

Rimu.update_shift — Method

update_shift(s <: ShiftStrategy, shift, shiftMode, tnorm, pnorm, dτ, step, df, v_new, v_old)

Update the shift according to strategy s. See ShiftStrategy.

Reexported Submodules

Hamiltonians

Link to Module Hamiltionians.jl

BitStringAddresses

Rimu.BitStringAddresses.AbstractFockAddress — Type

AbstractFockAddress

Supertype representing a Fock state.

Rimu.BitStringAddresses.BitString — Type

BitString{B,N} <: AbstractBitString

Type for storing bitstrings of static size. Holds B bits in N chunks, where each chunk is an UInt64

N is chosen automatically to accommodate B bits as efficiently as possible.

Constructors

BitString{B,N}(::SVector{N,T}): unsafe constructor. Does not check for ghost bits.

BitString{B,N}(i::UInt64): as above, but sets i as the rightmost chunk.

BitString{B}(::Integer): Convert integer to BitString. Integer is truncated to the correct number of bits.

Rimu.BitStringAddresses.BoseFS — Type

BoseFS{N,M,S} <: AbstractFockAddress
BoseFS(bs::S) where S <: BitAdd
BoseFS(bs::S, b)

Address type that represents a Fock state of N spinless bosons in M orbitals by wrapping a bitstring of type S. Orbitals are stored in reverse order, i.e. the first orbital in a BoseFS is stored rightmost in the bitstring bs. If the number of significant bits b is not encoded in S it must be passed as an argument (e.g. for BSAdd64 and BSAdd128).

Constructors

BoseFS{N,M}(::BitString): Unsafe constructor. Does not check whether the number of ones in a is equal to N.

BoseFS(::BitString): Automatically determine N and M. This constructor is not type stable!

BoseFS{[N,M,S]}(onr): Create BoseFS{N,M} from onr representation. This is efficient as long as at least N is provided.

Rimu.BitStringAddresses.BoseFS2C — Type

BoseFS2C{NA,NB,M,AA,AB} <: AbstractFockAddress

Address type that constructed with two BoseFS{N,M,S}. It represents a Fock state with two components, e.g. two different species of bosons with particle number NA from species S and particle number NB from species B. The number of orbitals M is expected to be the same for both components.

Rimu.BitStringAddresses.OccupiedOrbitalIterator — Type

occupied_orbitals(b)

Iterate over occupied orbitals in BoseFS address. Returns tuples of (boson_number, orbital_number, bit_offset).

Note that the bit_offset is zero-based!

Example

```jldoctest julia> b = BoseFS((1,5,0,4)) julia> for (n, i) in occupied_orbitals(b) @show n, i end (n, i) = (1, 1) (n, i) = (5, 2) (n, i) = (4, 4)

Rimu.BitStringAddresses.check_bitstring_typeparams — Method

check_bitstring_typeparams(::Val{B}, ::Val{N})

Check if number of bits B is consistent with number of chunks N. Throw an error if not.

Rimu.BitStringAddresses.chunk_bits — Method

chunks_bits(s, i)

Number of bits in the i-th chunk of s.

Rimu.BitStringAddresses.chunks — Method

chunks(::Type{<:BitString})
chunks(s)

SVector that stores the chunks of s.

Rimu.BitStringAddresses.has_ghost_bits — Method

has_ghost_bits(s)

Check for bits outside data field.

See also: remove_ghost_bits.

Rimu.BitStringAddresses.m_onr — Method

m_onr(bs)

Compute and return the occupation number representation of the bit string address bs as an MVector{M,Int32}, where M is the number of orbitals.

Rimu.BitStringAddresses.nearUniform — Method

nearUniform(BoseFS{N,M})
nearUniform(BoseFS{N,M,S}) -> bfs::BoseFS{N,M,S}

Create bosonic Fock state with near uniform occupation number of M modes with a total of N particles. Specifying the bit address type S is optional.

Examples

julia> nearUniform(BoseFS{7,5,BitString{14}})
BoseFS((2,2,1,1,1))

julia> nearUniform(BoseFS{7,5})
BoseFS((2,2,1,1,1))

Rimu.BitStringAddresses.nearUniformONR — Method

nearUniformONR(N, M) -> onr::SVector{M,Int}

Create occupation number representation onr distributing N particles in M modes in a close-to-uniform fashion with each orbital filled with at least N ÷ M particles and at most with N ÷ M + 1 particles.

Rimu.BitStringAddresses.num_bits — Method

num_bits(::Type{<:BitString})
num_bits(s::BitString)

Total number of bits stored in bitstring.

Rimu.BitStringAddresses.num_chunks — Method

num_chunks(::Type{<:BitString})
num_chunks(s::BitString)

Number of chunks in bitstring. Equivalent to length(chunks(s)).

Rimu.BitStringAddresses.num_chunks — Method

num_chunks(::Val{B})

Determine the number of chunks needed to store B bits.

Rimu.BitStringAddresses.num_modes — Method

num_modes(::Type{<:AbstractFockAddress})

Number of modes represented by address.

Rimu.BitStringAddresses.num_particles — Method

num_particles(::Type{<:AbstractFockAddress})

Number of particles represented by address.

Rimu.BitStringAddresses.onr — Method

onr(bs)

Compute and return the occupation number representation of the bit string address bs as an SVector{M,Int32}, where M is the number of orbitals.

Rimu.BitStringAddresses.remove_ghost_bits — Method

remove_ghost_bits(s)

Remove set bits outside data field if any are present.

See also: has_ghost_bits.

Rimu.BitStringAddresses.top_chunk_bits — Method

top_chunk_bits(::Type{<:BitString})
top_chunk_bits(s::BitString)

Number of bits stored in top chunk. Equivalent to chunk_bits(s, 1).

DictVectors

Rimu.DictVectors — Module

Module that provides data structures that behave similar to sparse vectors, but are indexed by arbitrary types (could be non-integers) similarly to dictionaries. The idea is to do linear algebra with data structures that are neither subtyped to AbstractVector nor to AbstractDict and are suitable for use with KrylovKit.jl. For this, the abstract type and interface AbstractDVec is provided, with the concrete implementation of DVec

Rimu.DictVectors.AbstractDVec — Type

DictVectors.AbstractDVec{K,V}

Abstract type for sparse vectors with valtype V based on dictionary-like structures. The vectors are designed to work well with FCIQMC and KrylovKit.

They lie somewhere between AbstractDicts and sparse AbstractVectors, generally behaving like a dictionary, while supportting various linear algebra functionality. Indexing with a value not stored in the dictionary returns zero(V). Setting a stored value to 0 or below eps(V::AbstractFloat) removes the value from the dictionary. Their length signals the number of stored elements, not the size of the vector space.

They have a StochasticStyle which selects the spawning algorithm in FCIQMC.

To iterate over an AbstractDVec, use pairs or values.

Interface

The interface is similar to the AbstractDict interface.

Implement what would be needed for the AbstractDict interface (pairs, keys, values, setindex!, getindex, delete!, length, haskey, empty!, isempty) and, in addition:

StochasticStyle

storage(dv) returns an AbstractDict storing the raw data with possibly different valtype than V.

Rimu.DictVectors.AbstractProjector — Type

Abstract supertype for projectors to be used in in lieu of DVecs or Vectors.

Rimu.DictVectors.CoherentInitiator — Type

CoherentInitiator(threshold) <: InitiatorRule

Initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators can spawn to initiators.

Multiple non-initiators can spawn to a single non-initiator if their contributions add up to a value greater than the initiator threshold.

See InitiatorRule.

Rimu.DictVectors.DVec — Type

DVec{K,V,D<:AbstractDict{K,V},S}

Dictionary-based vector-like data structure for use with FCIQMC and KrylovKit. While mostly behaving like a Dict, it supports various linear algebra operations such as norm and dot. It has a StochasticStyle that is used to select an appropriate spawning strategy in the FCIQMC algorithm.

See also: AbstractDVec.

Constructors

DVec(dict::AbstractDict[; style, capacity]): create a DVec with dict for storage. Note that the data may or may not be copied.

DVec(args...[; style, capacity]): args... are passed to the Dict constructor. The Dict is used for storage.

DVec{K,V}([; style, capacity]): create an empty DVec{K,V}.

DVec(dv::AbstractDVec[; style, capacity]): create a DVec with the same contents as adv. The style is inherited from dv by default.

The default style is selected based on the DVec's valtype (see default_style). If a style is given and the valtype does not match the style's eltype, the values are converted to an appropriate type.

The capacity argument is optional and sets the initial size of the DVec via sizehint!.

Examples

julia> dv = DVec(:a => 1)
DVec{Symbol,Int64} with 1 entries, style = IsStochasticInteger{Int64}()
  :a => 1

julia> dv = DVec(:a => 2, :b => 3; style=IsDynamicSemistochastic())
DVec{Symbol,Float32} with 2 entries, style = IsDynamicSemistochastic{Float64, true}(1.0, Inf, 1.0)
  :a => 2.0
  :b => 3.0

Rimu.DictVectors.Initiator — Type

Initiator(threshold) <: InitiatorRule

Initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators can spawn to initiators.

See InitiatorRule.

Rimu.DictVectors.InitiatorDVec — Type

InitiatorDVec{K,V} <: AbstractDVec{K,V}

Dictionary-based vector-like data structure for use with lomc! and KrylovKit.jl. See AbstractDVec. Functionally identical to DVec, but contains InitiatorValues internally in order to facilitate initiator methods. How the initiators are handled is controlled by the initiator keyword argument (see below).

Constructors

InitiatorDVec(dict::AbstractDict[; style, initiator, capacity]): create an InitiatorDVec with dict for storage. Note that the data may or may not be copied.

InitiatorDVec(args...[; style, initiator, capacity]): args... are passed to the Dict constructor. The Dict is used for storage.

InitiatorDVec{K,V}([; style, initiator, capacity]): create an empty InitiatorDVec{K,V}.

InitiatorDVec(dv::AbstractDVec[; style, initiator, capacity]): create an InitiatorDVec with the same contents as dv. The style is inherited from dv by default.

Keyword arguments

style: A valid StochasticStyle. The default is selected based on the InitiatorDVec's valtype (see default_style). If a style is given and the valtype does not match the style's eltype, the values are converted to an appropriate type.

initiator = Initiator(1): A valid InitiatorRule. See Initiator.

capacity: Indicative size as Int. Optional. Sets the initial size of the InitiatorDVec via sizehint!.

Rimu.DictVectors.InitiatorIterator — Type

InitiatorIterator

Iterator over pairs or values of an InitiatorDVec. Supports the SplittablesBase interface.

Rimu.DictVectors.InitiatorRule — Type

InitiatorRule{V}

Abstract type for defining initiator rules for InitiatorDVec. Concrete implementations:

Initiator

SimpleInitiator

CoherentInitiator

When defining a new InitiatorRule, also define a corresponding method for value!

Rimu.DictVectors.InitiatorValue — Type

InitiatorValue{V}(; safe::V, unsafe::V, initiator::V) where V

Composite "walker" with three fields. For use with InitiatorDVecs.

Rimu.DictVectors.IsDeterministic — Type

IsDeterministic{T=Float64}() <: StochasticStyle{T}

Trait for generalised vector of configuration indicating deterministic propagation of walkers.

See also StochasticStyle.

Rimu.DictVectors.IsDynamicSemistochastic — Type

IsDynamicSemistochastic{T=Float64}(rel_threshold=1, abs_threshold=Inf, proj_threshold=1) <: StochasticStyle{T}

QMC propagation with non-integer walker numbers and reduced noise. All possible spawns are performed deterministically when number of walkers in a configuration is high. Stochastic vector compression with threshold proj_threshold is applied after spawning and diagonal death steps.

Unlike with IsStochasticWithThreshold, when late_projection is set to true, walker annihilation is done before the stochastic vector compression.

Parameters:

late_projection = true: If set to true, threshold projection is done after all spawns are collected, otherwise, values are projected as they are being spawned.

rel_threshold = 1.0: If the walker number on a configuration times this threshold is greater than the number of offdiagonals, spawning is done deterministically. Should be set to 1 or more for best performance.

abs_threshold = Inf: If the walker number on a configuration is greater than this value, spawning is done deterministically. Can be set to e.g abs_threshold = 0.1 * target_walkers.

proj_threshold = 1.0: Values below this number are stochastically projected to this value or zero. See also IsStochasticWithThreshold.

See also StochasticStyle.

Rimu.DictVectors.IsStochastic2Pop — Type

IsStochastic2Pop{T=Complex{Int}}() <: StochasticStyle{T}

Trait for generalised vector of configurations indicating stochastic propagation with complex walker numbers representing two populations of integer walkers.

See also StochasticStyle.

Rimu.DictVectors.IsStochasticInteger — Type

IsStochasticInteger{T=Int}() <: StochasticStyle{T}

Trait for generalised vector of configurations indicating stochastic propagation as seen in the original FCIQMC algorithm.

See also StochasticStyle.

Rimu.DictVectors.IsStochasticWithThreshold — Type

IsStochasticWithThreshold(threshold=1.0) <: StochasticStyle

Trait for generalised vector of configurations indicating stochastic propagation with real walker numbers and cutoff threshold.

During stochastic propagation, walker numbers small than threshold will be stochastically projected to either zero or threshold.

See also StochasticStyle.

Rimu.DictVectors.Norm1ProjectorPPop — Type

Norm1ProjectorPPop()

Results in computing the one-norm per population when used in dot(). E.g.

dot(Norm1ProjectorPPop(),x)
-> norm(real.(x),1) + im*norm(imag.(x),1)

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.Norm2Projector — Type

Norm2Projector()

Results in computing the two-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,2) # with type Float64

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.NormProjector — Type

NormProjector()

Results in computing the one-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,1)

NormProjector() thus represents the vector sign.(x).

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.PopsProjector — Type

PopsProjector()

Results in computing the projection of one population on the other when used in dot(). E.g.

dot(PopsProjector(),x)
-> real(x) ⋅ imag(x)

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.SimpleInitiator — Type

SimpleInitiator(threshold) <: InitiatorRule

Simplified initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators cannot spawn.

See InitiatorRule.

Rimu.DictVectors.StochasticStyle — Type

StochasticStyle(v)

StochasticStyle specifies the native style of the generalised vector v that determines how simulations are to proceed. This can be fully stochastic (with IsStochasticInteger), fully deterministic (with IsDeterministic), or stochastic with floating point walker numbers and threshold (with IsStochasticWithThreshold).

When defining a new StochasticStyle, subtype it as MyStyle<:StochasticStyle{T} where T is the concrete value type the style is designed to work with.

For it to work with FCIQMC, a StochasticStyle must define the following:

fciqmc_col!(::StochasticStyle, w, H, address, value, shift, dτ)

step_stats(::StochasticStyle)

Optionally, it can also define update_dvec!, which can be used to perform arbitrary transformations on the dvec after the spawning step is complete.

Rimu.DictVectors.StyleUnknown — Type

StyleUnknown{T}() <: StochasticStyle

Trait for value types not (currently) compatible with FCIQMC. This style makes it possible to construct dict vectors with unsupported valtypes.

See also StochasticStyle.

Rimu.DictVectors.UniformProjector — Type

UniformProjector()

Represents a vector with all elements 1. To be used with dot(). Minimizes memory allocations.

UniformProjector()⋅v == sum(v)
dot(UniformProjector(), LO, v) == sum(LO*v)

See also ReportingStrategy for use of projectors in FCIQMC.

Rimu.DictVectors.add! — Method

add!(x::AbstractDVec,y::AbstactDVec)

Inplace add x+y and store result in x.

Rimu.DictVectors.default_style — Method

default_style(::Type)

Pick a StochasticStyle based on the value type. Throws an error if no known default style is known.

Rimu.DictVectors.deposit! — Method

deposit!(w::AbstractDVec, add, val, parent::Pair)

Add val into w at address add, taking into account initiator rules if applicable. parent contains the address => value pair from which the pair add => val was created. InitiatorDVec can intercept this and add its own functionality.

Rimu.DictVectors.deposit! — Method

deposit!(w::InitiatorDVec, add, val, p_add=>p_val)

Add val into w at address add as an InitiatorValue.

Rimu.DictVectors.localpart — Method

localpart(dv) -> AbstractDVec

Get the part of dv that is located on this MPI rank. Returns dv itself for DictVectors.

Rimu.DictVectors.storage — Function

storage(dvec) -> AbstractDict

Return the raw storage associated with dvec as an AbstractDict. Used in MPI communication.

Rimu.DictVectors.value — Function

value(i::InitiatorRule, v::InitiatorValue)

Convert the InitiatorValue v into a scalar value according to the InitiatorRule i.

Internal function that implements functionality of InitiatorDVec.

Rimu.DictVectors.walkernumber — Method

walkernumber(w)

Compute the number of walkers in w. It is used for updating the shift. Overload this function for modifying population control.

In most cases walkernumber(w) is identical to norm(w,1). For AbstractDVecs with complex coefficients it reports the one norm separately for the real and the imaginary part as a ComplexF64. See Norm1ProjectorPPop.

Rimu.DictVectors.zero! — Method

zero!(v)

Replace v by a zero vector as an inplace operation. For AbstractDVec types it means removing all non-zero elements. For AbstractArrays, it sets all of the values to zero.

StatsTools

Modules = [StatsTools]

Blocking

Modules = [Blocking]

ConsistentRNG

Link to Module ConsistentRNG.jl

Index

Rimu.ConsistentRNG

Rimu.DictVectors

Rimu.Hamiltonians

Rimu.Rimu

Rimu.ConsistentRNG.CRNGs

Rimu.AllOverlaps

Rimu.BitStringAddresses.AbstractFockAddress

Rimu.BitStringAddresses.BitString

Rimu.BitStringAddresses.BoseFS

Rimu.BitStringAddresses.BoseFS2C

Rimu.BitStringAddresses.OccupiedOrbitalIterator

Rimu.ConsistentRNG.CRNG

Rimu.ConstantTimeStep

Rimu.DelayedDoubleLogUpdate

Rimu.DelayedDoubleLogUpdateAfterTW

Rimu.DelayedLogUpdate

Rimu.DelayedLogUpdateAfterTargetWalkers

Rimu.DeltaMemory

Rimu.DeltaMemory2

Rimu.DeltaMemory3

Rimu.DictVectors.AbstractDVec

Rimu.DictVectors.AbstractProjector

Rimu.DictVectors.CoherentInitiator

Rimu.DictVectors.DVec

Rimu.DictVectors.Initiator

Rimu.DictVectors.InitiatorDVec

Rimu.DictVectors.InitiatorIterator

Rimu.DictVectors.InitiatorRule

Rimu.DictVectors.InitiatorValue

Rimu.DictVectors.IsDeterministic

Rimu.DictVectors.IsDynamicSemistochastic

Rimu.DictVectors.IsStochastic2Pop

Rimu.DictVectors.IsStochasticInteger

Rimu.DictVectors.IsStochasticWithThreshold

Rimu.DictVectors.Norm1ProjectorPPop

Rimu.DictVectors.Norm2Projector

Rimu.DictVectors.NormProjector

Rimu.DictVectors.PopsProjector

Rimu.DictVectors.SimpleInitiator

Rimu.DictVectors.StochasticStyle

Rimu.DictVectors.StyleUnknown

Rimu.DictVectors.UniformProjector

Rimu.DontUpdate

Rimu.DoubleLogProjected

Rimu.DoubleLogSumUpdate

Rimu.DoubleLogUpdate

Rimu.DoubleLogUpdateAfterTargetWalkers

Rimu.DoubleLogUpdateAfterTargetWalkersSwitch

Rimu.EveryKthStep

Rimu.EveryTimeStep

Rimu.FciqmcRunStrategy

Rimu.Hamiltonians.AbstractHamiltonian

Rimu.Hamiltonians.BoseHubbardMom1D

Rimu.Hamiltonians.BoseHubbardReal1D

Rimu.Hamiltonians.ExtendedBHReal1D

Rimu.HistoryLogUpdate

Rimu.LogUpdate

Rimu.LogUpdateAfterTargetWalkers

Rimu.MemoryStrategy

Rimu.NoMemory

Rimu.NoStats

Rimu.ProjectedMemory

Rimu.PurgeNegatives

Rimu.QMCState

Rimu.ReplicaState

Rimu.ReplicaStrategy

Rimu.Report

Rimu.ReportDFAndInfo

Rimu.ReportingStrategy

Rimu.RunTillLastStep

Rimu.ShiftMemory

Rimu.ShiftStrategy

Rimu.TimeStepStrategy

Rimu.TripleLogUpdate

Rimu.BitStringAddresses.check_bitstring_typeparams

Rimu.BitStringAddresses.chunk_bits

Rimu.BitStringAddresses.chunks

Rimu.BitStringAddresses.has_ghost_bits

Rimu.BitStringAddresses.m_onr

Rimu.BitStringAddresses.nearUniform

Rimu.BitStringAddresses.nearUniform

Rimu.BitStringAddresses.nearUniformONR

Rimu.BitStringAddresses.num_bits

Rimu.BitStringAddresses.num_chunks

Rimu.BitStringAddresses.num_chunks

Rimu.BitStringAddresses.num_modes

Rimu.BitStringAddresses.num_particles

Rimu.BitStringAddresses.onr

Rimu.BitStringAddresses.remove_ghost_bits

Rimu.BitStringAddresses.top_chunk_bits

Rimu.ConsistentRNG.cRand

Rimu.ConsistentRNG.cRandn

Rimu.ConsistentRNG.check_crng_independence

Rimu.ConsistentRNG.newChildRNG

Rimu.ConsistentRNG.seedCRNG!

Rimu.ConsistentRNG.sync_cRandn

Rimu.ConsistentRNG.trng

Rimu.DictVectors.add!

Rimu.DictVectors.default_style

Rimu.DictVectors.deposit!

Rimu.DictVectors.deposit!

Rimu.DictVectors.localpart

Rimu.DictVectors.storage

Rimu.DictVectors.value

Rimu.DictVectors.walkernumber

Rimu.DictVectors.zero!

Rimu.advance!

Rimu.apply_memory_noise!

Rimu.compute_proj_observables

Rimu.fciqmc_col!

Rimu.fciqmc_step!

Rimu.lomc!

Rimu.print_report

Rimu.refine_r_strat

Rimu.replica_stats

Rimu.report!

Rimu.report!

Rimu.sort_into_targets!

Rimu.step_stats

Rimu.threshold_projected_deposit!

Rimu.update_dvec!

Rimu.update_dτ

Rimu.update_shift