Module RMPI

Rimu.RMPI — Module

Module for providing MPI functionality for Rimu. This module is unexported. To use it, run

using Rimu.RMPI

MPIData

Rimu.RMPI.MPIData — Type

MPIData(data; kwargs...)

Wrapper used for signaling that this data is part of a distributed data structure and communication should happen with MPI. MPIData can generally be used where an AbstractDVec would be used otherwise. Unlike AbstractDVecs, MPIData does not support indexing, or iteration over keys, values, and pairs.

Keyword arguments:

setup = mpi_point_to_point - controls the communication stratgy

mpi_one_sided uses one-sided communication with remote memory access (RMA), sets MPIOneSided strategy.

mpi_point_to_point uses MPIPointTOPoint strategy.

mpi_all_to_all uses MPIAllToAll strategy.

mpi_no_exchange sets MPINoWalkerExchange strategy. Experimental. Use with caution!

comm = mpi_comm()

root = mpi_root

The rest of the keyword arguments are passed to setup.

The following distribute strategies are available. The functions are unexported.

Rimu.RMPI.mpi_point_to_point — Function

mpi_point_to_point(data, comm = mpi_comm(), root = mpi_root)

Declare data as mpi-distributed and set communication strategy to point-to-point.

Sets up the MPIData structure with MPIPointToPoint strategy.

Rimu.RMPI.mpi_one_sided — Function

mpi_one_sided(data, comm = mpi_comm(), root = mpi_root; capacity)

Declare data as mpi-distributed and set communication strategy to one-sided with remote memory access (RMA). capacity sets the capacity of the RMA windows.

Sets up the MPIData structure with MPIOneSided strategy.

Rimu.RMPI.mpi_all_to_all — Function

mpi_all_to_all(data, comm = mpi_comm(), root = mpi_root)

Declare data as mpi-distributed and set communication strategy to all-to-all.

Sets up the MPIData structure with MPIAllToAll strategy.

Rimu.RMPI.mpi_no_exchange — Function

mpi_no_exchange(data, comm = mpi_comm(), root = mpi_root)

Declare data as mpi-distributed and set communication strategy to MPINoWalkerExchange. Sets up the MPIData structure with MPINoWalkerExchange strategy.

MPI convenience functions

Rimu.RMPI.mpi_rank — Function

mpi_rank(comm = mpi_comm())

Return the current MPI rank.

Rimu.RMPI.is_mpi_root — Function

is_mpi_root(root = mpi_root)

Returns true if called from the root rank

Rimu.RMPI.@mpi_root — Macro

@mpi_root expr

Evaluate expression only on the root rank. Extra care needs to be taken as expr must not contain any code that involves syncronising MPI operations, i.e. actions that would require syncronous action of all MPI ranks.

Example:

wn = walkernumber(dv)   # an MPI syncronising function call that gathers
                        # information from all MPI ranks
@mpi_root @info "The current walker number is" wn # print info message on root only

Rimu.RMPI.mpi_barrier — Function

mpi_barrier(comm = mpi_comm())

The MPI barrier with optional argument. MPI syncronizing.

Rimu.RMPI.mpi_comm — Function

Default MPI communicator for RMPI.

Rimu.RMPI.mpi_root — Constant

Default MPI root for RMPI.

Rimu.RMPI.mpi_seed_CRNGs! — Function

mpi_seed_CRNGs!(seed = rand(Random.RandomDevice(), UInt))

Re-seed the random number generators in an MPI-safe way. If seed is provided, the random numbers from cRand() will follow a deterministic sequence.

Independence of the random number generators on different MPI ranks is achieved by adding hash(mpi_rank()) to seed.

Rimu.RMPI.mpi_allprintln — Function

mpi_allprintln(args...)

Print a message to stdout from each rank separately, in order. MPI synchronizing.