Module BitStringAddresses

This module contains the implementations of BitString and various Fock addresses. The addresses serve as a basis for a Hamiltonian.

While there are not restrictions on the type of address a Hamiltonian uses, Rimu provides implementations for Bosonic, Fermionic, and mixed Fock States.

When implementing a new address type, care must be taken to make them space-efficient and stack-allocated - avoid using arrays to represent your addresses at all costs!

Fock addresses

Rimu provides a variety of address implementations (see below) that should make it straightforward to implement efficient Hamiltonians.

Fock address API

Rimu.BitStringAddresses.AbstractFockAddress — Type

AbstractFockAddress

Supertype representing a Fock state.

Rimu.BitStringAddresses.SingleComponentFockAddress — Type

SingleComponentFockAddress{M}

A type representing a single component Fock state with M modes.

Implemented subtypes: BoseFS, FermiFS.

Supported functionality

find_mode

find_occupied_mode

is_occupied

num_occupied_modes

occupied_modes: Lazy iterator.

OccupiedModeMap: AbstractVector with eager construction.

excitation: Create a new address.

BoseFSIndex and FermiFSIndex for indexing.

Rimu.BitStringAddresses.BoseFS — Type

BoseFS{N,M,S} <: SingleComponentFockAddress

Address type that represents a Fock state of N spinless bosons in M modes by wrapping a bitstring of type S <: BitString.

Constructors

BoseFS{N,M}(bs::BitString): Unsafe constructor. Does not check whether the number of particles in bs is equal to N.

BoseFS(::BitString): Automatically determine N and M. This constructor is not type stable!

BoseFS{[N,M,S]}(onr): Create BoseFS{N,M} from onr representation. This is efficient as long as at least N is provided.

See also: SingleComponentFockAddress, FermiFS, BitString.

Rimu.BitStringAddresses.FermiFS — Type

FermiFS{N,M,S} <: SingleComponentFockAddress

Address type that represents a Fock state of N fermions of the same spin in M modes by wrapping a bitstring of type S <: BitString.

Constructors

FermiFS{N,M}(bs::BitString): Unsafe constructor. Does not check whether the number of particles in bs is equal to N.

FermiFS(::BitString): Automatically determine N and M. This constructor is not type stable!

FermiFS{[N,M,S]}(onr): Create FermiFS{N,M} from onr representation. This is efficient as long as at least N is provided.

See also: SingleComponentFockAddress, BoseFS, BitString.

Rimu.BitStringAddresses.BoseFS2C — Type

BoseFS2C{NA,NB,M,AA,AB} <: AbstractFockAddress
BoseFS2C(onr_a, onr_b)

Address type that constructed with two BoseFS{N,M,S}. It represents a Fock state with two components, e.g. two different species of bosons with particle number NA from species S and particle number NB from species B. The number of modes M is expected to be the same for both components.

Rimu.BitStringAddresses.CompositeFS — Type

CompositeFS(addresses::SingleComponentFockAddress...) <: AbstractFockAddress

Used to encode addresses for multi-component models. All component addresses are expected have the same number of modes.

See also: BoseFS, FermiFS, SingleComponentFockAddress, num_modes, FermiFS2C.

Rimu.BitStringAddresses.num_particles — Function

num_particles(::Type{<:AbstractFockAddress})
num_particles(::AbstractFockAddress)

Number of particles represented by address.

Rimu.BitStringAddresses.num_modes — Function

num_modes(::Type{<:AbstractFockAddress})
num_modes(::AbstractFockAddress)

Number of modes represented by address.

Rimu.BitStringAddresses.num_components — Function

num_components(::Type{<:AbstractFockAddress})
num_components(::AbstractFockAddress)

Number of components in address.

Rimu.BitStringAddresses.onr — Function

onr(add::AbstractFockAddress, geom::LatticeGeometry)

Returns the occupation number representation of a Fock state address as an SArray with the shape of the lattice geometry geom. For composite addresses, a tuple of onrs is returned.

onr(bs)

Compute and return the occupation number representation of the bit string address bs as an SVector{M,Int32}, where M is the number of modes.

Rimu.BitStringAddresses.near_uniform — Function

near_uniform(BoseFS{N,M})
near_uniform(BoseFS{N,M,S}) -> bfs::BoseFS{N,M,S}

Create bosonic Fock state with near uniform occupation number of M modes with a total of N particles. Specifying the bit address type S is optional.

Examples

julia> near_uniform(BoseFS{7,5,BitString{14}})
BoseFS{7,5}((2, 2, 1, 1, 1))

julia> near_uniform(BoseFS{7,5})
BoseFS{7,5}((2, 2, 1, 1, 1))

Rimu.BitStringAddresses.occupied_modes — Function

occupied_modes(::SingleComponentFockAddress)

Return a lazy iterator over all occupied modes in an address. Iterates over BoseFSIndexs for BoseFS, and over FermiFSIndexs for FermiFS. See OccupiedModeMap for an eager version.

Example

julia> b = BoseFS((1,5,0,4));

julia> foreach(println, occupied_modes(b))
BoseFSIndex(occnum=1, mode=1, offset=0)
BoseFSIndex(occnum=5, mode=2, offset=2)
BoseFSIndex(occnum=4, mode=4, offset=9)

julia> f = FermiFS((1,1,0,1,0,0,1));

julia> foreach(println, occupied_modes(f))
FermiFSIndex(occnum=1, mode=1)
FermiFSIndex(occnum=1, mode=2)
FermiFSIndex(occnum=1, mode=4)
FermiFSIndex(occnum=1, mode=7)

See also find_occupied_mode, SingleComponentFockAddress.

Rimu.BitStringAddresses.is_occupied — Function

is_occupied(::SingleComponentFockAddress, i)

Return true if index i points to an occupied mode.

See SingleComponentFockAddress.

Rimu.BitStringAddresses.num_occupied_modes — Function

num_occupied_modes(::SingleComponentFockAddress)

Get the number of occupied modes in address. Equivalent to length(occupied_modes(address)), or the number of non-zeros in its ONR representation.

Example

julia> num_occupied_modes(BoseFS((1, 0, 2)))
2
julia> num_occupied_modes(FermiFS((1, 1, 1, 0)))
3

See SingleComponentFockAddress.

Rimu.BitStringAddresses.find_occupied_mode — Function

find_occupied_mode(::SingleComponentFockAddress, k)
find_occupied_mode(::BoseFS, k, [n])

Find the k-th occupied mode in address (with at least n particles). Returns BoseFSIndex for BoseFS, and FermiFSIndex for FermiFS. When unsuccessful it returns a zero index.

Example

julia> find_occupied_mode(FermiFS((1, 1, 1, 0)), 2)
FermiFSIndex(occnum=1, mode=2)

julia> find_occupied_mode(BoseFS((1, 0, 2)), 1)
BoseFSIndex(occnum=1, mode=1, offset=0)

julia> find_occupied_mode(BoseFS((1, 0, 2)), 1, 2)
BoseFSIndex(occnum=2, mode=3, offset=3)

See also occupied_modes, OccupiedModeMap, SingleComponentFockAddress.

Rimu.BitStringAddresses.find_mode — Function

find_mode(::SingleComponentFockAddress, i)

Find the i-th mode in address. Returns BoseFSIndex for BoseFS, and FermiFSIndex for FermiFS. Can work on a tuple of modes. Does not check bounds.

julia> find_mode(BoseFS((1, 0, 2)), 2)
BoseFSIndex(occnum=0, mode=2, offset=2)

julia> find_mode(FermiFS((1, 1, 1, 0)), (2,3))
(FermiFSIndex(occnum=1, mode=2), FermiFSIndex(occnum=1, mode=3))

See SingleComponentFockAddress.

Rimu.BitStringAddresses.excitation — Function

excitation(a::SingleComponentFockAddress, creations::NTuple{N}, destructions::NTuple{N})

Generate an excitation on address a by applying creations and destructions, which are tuples of the appropriate address indices (i.e. BoseFSIndex for bosons, or FermiFSIndex for fermions).

\[a^†_{c_1} a^†_{c_2} \ldots a_{d_1} a_{d_2} \ldots |\mathrm{a}\rangle \to α|\mathrm{nadd}\rangle\]

Returns the new address nadd and the value α. If the excitation is illegal, returns an arbitrary address and the value 0.0.

Example

julia> f = FermiFS((1,1,0,0,1,1,1,1))
FermiFS{6,8}((1, 1, 0, 0, 1, 1, 1, 1))

julia> i, j, k, l = find_mode(f, (3,4,2,5))
(FermiFSIndex(occnum=0, mode=3), FermiFSIndex(occnum=0, mode=4), FermiFSIndex(occnum=1, mode=2), FermiFSIndex(occnum=1, mode=5))

julia> excitation(f, (i,j), (k,l))
(FermiFS{6,8}((1, 0, 1, 1, 0, 1, 1, 1)), -1.0)

See SingleComponentFockAddress.

BitStrings

The atomic addresses, BoseFS and FermiFS, are implemented as bitstrings. Using this approach over an occupation number representation makes the addresses much more space-efficient. The API for BitStrings is as follows.

BitString API

Rimu.BitStringAddresses.BitString — Type

BitString{B,N,T<:Unsigned}

Type for storing bitstrings of static size. Holds B bits in N chunks, where each chunk is of type T.

N is chosen automatically to accommodate B bits as efficiently as possible.

Constructors

BitString{B,N,T}(::SVector{N,T}): unsafe constructor. Does not check for ghost bits.

BitString{B,N,T}(i::T): as above, but sets i as the rightmost chunk.

BitString{B}(::Integer): Convert integer to BitString. Integer is truncated to the correct number of bits.

Rimu.BitStringAddresses.num_bits — Function

num_bits(::Type{<:BitString})
num_bits(s::BitString)

Total number of bits stored in bitstring.

Rimu.BitStringAddresses.num_chunks — Function

num_chunks(::Val{B})

Determine the number and type of chunks needed to store B bits.

num_chunks(::Type{<:BitString})
num_chunks(s::BitString)

Number of chunks in bitstring. Equivalent to length(chunks(s)).

Rimu.BitStringAddresses.chunk_type — Function

chunk_type(::Type{<:BitString})
chunk_type(s::BitString)

Type of unsigned integer used to store the chunks.

Rimu.BitStringAddresses.chunk_bits — Function

chunks_bits(::Type{<:BitString}, i)
chunks_bits(s, i)

Number of bits in the i-th chunk of s.

Rimu.BitStringAddresses.top_chunk_bits — Function

top_chunk_bits(::Type{<:BitString})
top_chunk_bits(s::BitString)

Number of bits stored in top chunk. Equivalent to chunk_bits(s, 1).