Module BitStringAddresses

This module contains the implementations of the underlying data structures to efficiently represent Fock states. In Rimu.jl, Fock states are used as the basis for linear (Hilbert) space. The concrete implementations of Fock states are used as addresses to identify matrix elements of operators, or elements of state vectors.

Fock addresses

Many Hamiltonians in Rimu.jl are implemented in a generic way and crucial information that defines the size of Hilbert space and the quantum statistics of particle are defined by passing a Fock state address. Rimu.jl provides a variety of address types that enable the efficient implementation of many-body Hamiltonians. All implementations of Fock states are subtype to the AbstractFockAddress abstract type. For Fock states representing a single component of indistinguishable quantum particles there is the more specialised type SingleComponentFockAddress.

Examples of Fock addresses are:

BoseFS: Bosonic Fock state with fixed mode number and fixed or variable particle number.

FermiFS: Fermionic Fock state with fixed mode number and fixed or variable particle number.

HardcoreBoseFS: Fock state for hardcore bosons with fixed mode number and fixed or variable particle number.

CompositeFS Multi-component Fock state composed of the above types.

The various address types make use efficient underlying data storage types like BitString and SortedParticleList.

Fock address API

Rimu.Interfaces.AbstractFockAddress — Type

AbstractFockAddress

Abstract type representing a Fock state, i.e. a state with a definite particle number. Instances of subtypes of AbstractFockAddress represent basis states that are used as the computational basis for representing vectors and operators.

Implementations

SingleComponentFockAddress is a supertype for single-component Fock states.

CompositeFS is a container type for multi-component Fock states,

Interface

The following functions can be called on the type or an instance of a subtype of `AbstractFockAddress:

num_particles: Returns the number of particles in the Fock state. May return missing if called on a type that allows for a variable number of particles.

num_components: Returns the number of components in the Fock state.

num_modes: Returns the number of modes in the Fock state. This returns a tuple for multi-component Fock states, and an integer for single-component Fock states.

num_modes_are_equal: Returns true if all components of a multi-component Fock state have the same number of modes, and false otherwise. Returns true for single-component Fock states.

num_modes_check_equal: Returns the number of modes in a Fock state as a single integer, and throws an ArgumentError if the components of a multi-component Fock state have different numbers of modes.

maximum_mode_occupation: Returns the maximum number of particles that can occupy a single mode in the Fock state. Returns an integer for single-component Fock states, and a tuple for multi-component Fock states.

See also SingleComponentFockAddress, CompositeFS, BoseFS, FermiFS, num_particles, num_modes, num_modes_check_equal, num_components, maximum_mode_occupation.

Rimu.Interfaces.num_particles — Function

num_particles(::Type{<:AbstractFockAddress})
num_particles(::AbstractFockAddress)

Number of particles represented by address.

See also num_modes, num_modes_check_equal, num_components, maximum_mode_occupation, CompositeFS.

Rimu.Interfaces.num_components — Function

num_components(::Type{<:AbstractFockAddress})
num_components(::AbstractFockAddress)

Number of components in address.

See also num_modes, num_modes_check_equal, num_particles, maximum_mode_occupation, CompositeFS.

Rimu.Interfaces.num_modes — Function

num_modes(::Type{<:AbstractFockAddress})
num_modes(::AbstractFockAddress)

Number of modes represented by address. Returns a tuple for multi-component addresses, and an integer for single-component addresses.

See also num_modes_check_equal, num_modes_are_equal, num_particles, num_components, maximum_mode_occupation, CompositeFS.

Rimu.Interfaces.num_modes_check_equal — Function

num_modes_check_equal(::Type{<:AbstractFockAddress})::Int
num_modes_check_equal(address::AbstractFockAddress)::Int
num_modes_check_equal(addrs::Vararg{AbstractFockAddress})::Int

Check that all components of a multi-component address have the same number of modes, and return the number of modes. Throws an ArgumentError if the components have different numbers of modes. For a single-component address, simply returns the number of modes. When called with multiple addresses, checks that all addresses have the same number of modes, and throws an ArgumentError if they do not.

See also num_modes_are_equal, num_modes, num_components, maximum_mode_occupation, CompositeFS, SingleComponentFockAddress.

Rimu.Interfaces.num_modes_are_equal — Function

num_modes_are_equal(::Type{<:AbstractFockAddress)::Bool
num_modes_are_equal(address::AbstractFockAddress)::Bool

Check if all components of a multi-component address have the same number of modes. Returns true for single-component addresses. For multi-component addresses, returns true if all components have the same number of modes, and false otherwise.

See also num_modes_check_equal, num_modes, num_particles, num_components, maximum_mode_occupation, CompositeFS.

Rimu.BitStringAddresses — Module

BitStringAddresses

Module with types and methods pertaining to bitstring addresses.

Rimu.BitStringAddresses.BoseFSIndex — Type

BoseFSIndex

Struct used for indexing and performing excitations on a BoseFS. BoseFSIndex is returned by find_mode and find_occupied_mode.

Fields:

occnum: the occupation number.

mode: the index of the mode.

offset: the position of the mode in the address. This is the bit offset of the mode when

the address is represented by a bitstring, and the position in the list when it is represented by SortedParticleList.

See also FermiFSIndex, find_mode, find_occupied_mode, excitation.

Rimu.BitStringAddresses.FermiFSIndex — Type

FermiFSIndex

Struct used for indexing and performing excitations on a FermiFS and HardcoreBoseFS. FermiFSIndex is returned by find_mode and find_occupied_mode.

Fields:

occnum: the occupation number.

mode: the index of the mode.

offset: the position of the mode in the address. This is mode - 1 when the address is represented by a bitstring, and the position in the list when using SortedParticleList.

See also BoseFSIndex, find_mode, find_occupied_mode, excitation.

Rimu.BitStringAddresses.OccupiedPairsMap — Type

OccupiedPairsMap(addr::SingleComponentFockAddress) <: AbstractVector

Get a map of all distinct pairs of indices in addr. Pairs involving multiply-occupied modes are counted once, (including self-pairing). This is useful for cases where identifying pairs of particles for eg. interactions is not well-defined or efficient to do on the fly. This is an eager iterator whose elements are a tuple of particle indices that can be given to excitation

Example

julia> addr = BoseFS(10, 0, 0, 0, 2, 0, 1)
BoseFS(10, 0, 0, 0, 2, 0, 1)

julia> pairs = OccupiedPairsMap(addr)
5-element OccupiedPairsMap{78, Tuple{BoseFSIndex, BoseFSIndex}}:
 (BoseFSIndex(occnum=10, mode=1, offset=0), BoseFSIndex(occnum=10, mode=1, offset=0))
 (BoseFSIndex(occnum=2, mode=5, offset=14), BoseFSIndex(occnum=2, mode=5, offset=14))
 (BoseFSIndex(occnum=2, mode=5, offset=14), BoseFSIndex(occnum=10, mode=1, offset=0))
 (BoseFSIndex(occnum=1, mode=7, offset=18), BoseFSIndex(occnum=10, mode=1, offset=0))
 (BoseFSIndex(occnum=1, mode=7, offset=18), BoseFSIndex(occnum=2, mode=5, offset=14))

julia> excitation(addr, pairs[2], pairs[4])
(BoseFS(9, 0, 0, 0, 4, 0, 0), 10.954451150103322)

See also occupied_mode_map.

Rimu.BitStringAddresses.SingleComponentFockAddress — Type

SingleComponentFockAddress{N,M} <: AbstractFockAddress

A type representing a single component Fock state with N particles and M modes. The particle number is part of the type but can be set to missing to allow for a variable particle number. The number of modes is always fixed.

Implemented subtypes:

BoseFS: Bosonic Fock state.

FermiFS: Fermionic Fock state.

HardcoreBoseFS: Fock state for hardcore bosons or spin or qubit systems.

Supported functionality

find_mode

find_occupied_mode

num_occupied_modes

occupied_modes: Lazy iterator.

occupied_mode_map: AbstractVector with eager construction.

excitation: Create a new address.

BoseFSIndex and FermiFSIndex for indexing.

maximum_mode_occupation: Maximum number of particles that can occupy a single mode.

See also CompositeFS, AbstractFockAddress.

Rimu.BitStringAddresses.each_mode — Method

each_mode(address::SingleComponentFockAddress)

Return an iterator that iterates over each mode in an address. Iterates value of either BoseFSIndex or FermiFSIndex.

Rimu.BitStringAddresses.excitation — Function

excitation(addr::SingleComponentFockAddress, creations::NTuple, destructions::NTuple)

Generate an excitation on address addr by applying creations and destructions, which are tuples of the appropriate address indices (i.e. BoseFSIndex for BoseFS, or FermiFSIndex for FermiFS and HardcoreBoseFS; BoseFS{missing} also supports integer indices).

\[a^†_{c_1} a^†_{c_2} … a_{d_1} a_{d_2} … |f⟩ → α|f'⟩\]

Returns the new address f' and the factor α. The value of α is given by the square root of the product of mode occupations before destruction and after creation. If the excitation is illegal, returns an arbitrary address and the value 0.0. Note that the number of particles may change if the number of creation and destruction operators is not equal and num_particles(typeof(addr)) is missing. See examples below.

Example

julia> f = FermiFS(1,1,0,0,1,1,1,1)
FermiFS(1, 1, 0, 0, 1, 1, 1, 1)

julia> i, j, k, l = find_mode(f, (3,4,2,5))
(FermiFSIndex(occnum=0, mode=3, offset=2), FermiFSIndex(occnum=0, mode=4, offset=3), FermiFSIndex(occnum=1, mode=2, offset=1), FermiFSIndex(occnum=1, mode=5, offset=4))

julia> excitation(f, (i,j), (k,l)) # number conserving excitation
(FermiFS(1, 0, 1, 1, 0, 1, 1, 1), -1.0)

julia> fm = FermiFS{missing}(1,1,0,0,1,1,1,1) # number non-conserving address
FermiFS{missing}(1, 1, 0, 0, 1, 1, 1, 1)

julia> excitation(fm, (i,), (k,l)) # particle number changes, same address type
(FermiFS{missing}(1, 0, 1, 0, 0, 1, 1, 1), 1.0)

julia> s = fs"|1 2 3⟩{}"
BoseFS{missing}(1, 2, 3)

julia> num_particles(s)
6

julia> es, α = excitation(s, (1,1), (3,))
(BoseFS{missing}(3, 2, 2), 4.242640687119285)

julia> num_particles(es)
7

See SingleComponentFockAddress.

Rimu.BitStringAddresses.find_mode — Function

find_mode(::SingleComponentFockAddress, i[, occ])

Find the i-th mode in address. Returns BoseFSIndex for BoseFS, and FermiFSIndex for FermiFS. Can work on a tuple of modes. Does not check bounds. Setting occ to the result of occupied_mode_map can increase the performance for BoseFS.

julia> find_mode(BoseFS(1, 0, 2), 2)
BoseFSIndex(occnum=0, mode=2, offset=2)

julia> find_mode(FermiFS(1, 1, 1, 0), (2,3))
(FermiFSIndex(occnum=1, mode=2, offset=1), FermiFSIndex(occnum=1, mode=3, offset=2))

See SingleComponentFockAddress.

Rimu.BitStringAddresses.find_occupied_mode — Function

find_occupied_mode(::SingleComponentFockAddress, k)
find_occupied_mode(::BoseFS, k, [n])

Find the k-th occupied mode in address (with at least n particles). Returns BoseFSIndex for BoseFS, and FermiFSIndex for FermiFS. When unsuccessful it returns a zero index.

Example

julia> find_occupied_mode(FermiFS(1, 1, 1, 0), 2)
FermiFSIndex(occnum=1, mode=2, offset=1)

julia> find_occupied_mode(BoseFS(1, 0, 2), 1)
BoseFSIndex(occnum=1, mode=1, offset=0)

julia> find_occupied_mode(BoseFS(1, 0, 2), 1, 2)
BoseFSIndex(occnum=2, mode=3, offset=3)

See also occupied_modes, occupied_mode_map, SingleComponentFockAddress.

Rimu.BitStringAddresses.num_occupied_modes — Function

num_occupied_modes(::SingleComponentFockAddress)

Get the number of occupied modes in address. Equivalent to length(occupied_modes(address)), or the number of non-zeros in its ONR representation.

Example

julia> num_occupied_modes(BoseFS(1, 0, 2))
2
julia> num_occupied_modes(FermiFS(1, 1, 1, 0))
3

See SingleComponentFockAddress.

Rimu.BitStringAddresses.occupied_mode_map — Method

occupied_mode_map(addr) <: AbstractVector

Get a map of occupied modes in address as an AbstractVector of indices compatible with excitation - BoseFSIndex or FermiFSIndex.

occupied_mode_map(addr)[i] contains the index for the i-th occupied mode. This is useful because repeatedly looking for occupied modes with find_occupied_mode can be time-consuming. occupied_mode_map(addr) is an eager version of the iterator returned by occupied_modes. It is similar to onr but contains more information.

Example

julia> b = BoseFS(10, 0, 0, 0, 2, 0, 1)
BoseFS(10, 0, 0, 0, 2, 0, 1)

julia> mb = occupied_mode_map(b)
3-element Rimu.BitStringAddresses.ModeMap{7, BoseFSIndex}:
 BoseFSIndex(occnum=10, mode=1, offset=0)
 BoseFSIndex(occnum=2, mode=5, offset=14)
 BoseFSIndex(occnum=1, mode=7, offset=18)

julia> f = FermiFS(1,1,1,1,0,0,1,0,0)
FermiFS(1, 1, 1, 1, 0, 0, 1, 0, 0)

julia> mf = occupied_mode_map(f)
5-element Rimu.BitStringAddresses.ModeMap{5, FermiFSIndex}:
 FermiFSIndex(occnum=1, mode=1, offset=0)
 FermiFSIndex(occnum=1, mode=2, offset=1)
 FermiFSIndex(occnum=1, mode=3, offset=2)
 FermiFSIndex(occnum=1, mode=4, offset=3)
 FermiFSIndex(occnum=1, mode=7, offset=6)

julia> mf == collect(occupied_modes(f))
true

julia> dot(mf, mb)
11

julia> dot(mf, 1:20)
17

See also dot, unoccupied_mode_map, SingleComponentFockAddress.

Rimu.BitStringAddresses.occupied_modes — Function

occupied_modes(::SingleComponentFockAddress)

Return a lazy iterator over all occupied modes in an address. Iterates over BoseFSIndexs for BoseFS, and over FermiFSIndexs for FermiFS. See occupied_mode_map for an eager version.

Example

julia> b = BoseFS(1,5,0,4);

julia> foreach(println, occupied_modes(b))
BoseFSIndex(occnum=1, mode=1, offset=0)
BoseFSIndex(occnum=5, mode=2, offset=2)
BoseFSIndex(occnum=4, mode=4, offset=9)

julia> f = FermiFS(1,1,0,1,0,0,1);

julia> foreach(println, occupied_modes(f))
FermiFSIndex(occnum=1, mode=1, offset=0)
FermiFSIndex(occnum=1, mode=2, offset=1)
FermiFSIndex(occnum=1, mode=4, offset=3)
FermiFSIndex(occnum=1, mode=7, offset=6)

See also find_occupied_mode, SingleComponentFockAddress.

Rimu.BitStringAddresses.onr — Function

occupation_number_representation(fs::SingleComponentFockAddress)
onr(fs::SingleComponentFockAddress)

Compute and return the occupation number representation of the Fock state fs as an SVector{M}, where M is the number of modes.

Rimu.BitStringAddresses.unoccupied_modes — Function

unoccupied_modes(::FermiFS)

Return a lazy iterator over all unoccupied modes in an Fermi type address. Iterates over over FermiFSIndexs for FermiFS. See unoccupied_mode_map for an eager version.

Example

julia> f = FermiFS(1,1,0,1,0,0,1);

julia> foreach(println, unoccupied_modes(f))
FermiFSIndex(occnum=0, mode=3, offset=2)
FermiFSIndex(occnum=0, mode=5, offset=4)
FermiFSIndex(occnum=0, mode=6, offset=5)

See also find_occupied_mode, occupied_modes.

Rimu.BitStringAddresses.@fs_str — Macro

fs"$(string)"

Parse the compact representation of a Fock state. Useful for copying the printout from a vector to the REPL.

Example

julia> DVec(BoseFS{3,4}(0, 1, 2, 0) => 1)
DVec{BoseFS{3, 4, BitString{6, 1, UInt8}},Int64} with 1 entry, style = IsStochasticInteger{Int64}()
  fs"|0 1 2 0⟩" => 1

julia> fs"|0 1 2 0⟩" => 1 # Copied from above printout; \rangle(tab) -> ⟩
BoseFS(0, 1, 2, 0) => 1

julia> fs"|1 2 3⟩⊗|0 1 0⟩" # composite bosonic Fock state; \otimes(tab) -> ⊗
CompositeFS(
  BoseFS(1, 2, 3),
  BoseFS(0, 1, 0),
)

julia> fs"|↑↓↑⟩" # 2-component fermions; \uparrow(tab) -> ↑, \downarrow(tab) -> ↓
CompositeFS(
  FermiFS(1, 0, 1),
  FermiFS(0, 1, 0),
)

julia> fs"|↑↓↑⇅⟩{}" # spinor fermions; \dblarrowupdown(tab) -> ⇅
CompositeFS(
  FermiFS{missing}(1, 0, 1, 1),
  FermiFS{missing}(0, 1, 0, 1),
)

julia> s = fs"|0 1 2 0⟩{}" # constructing BoseFS{missing} with default UInt8 container
BoseFS{missing}(0, 1, 2, 0)

julia> [s] # prints out with the container type specified in braces (empty for default UInt8)
1-element Vector{BoseFS{missing, 4, SVector{4, UInt8}}}:
 fs"|0 1 2 0⟩{}"

julia> fs"|●∘●⟩" # \mdlgblkcircle(tab) -> ●, \circ(tab) -> ∘
HardcoreBoseFS(1, 0, 1)

See also FermiFS, BoseFS, CompositeFS, FermiFS2C, HardcoreBoseFS.

Rimu.BitStringAddresses.BoseFS — Type

BoseFS{N,M,S} <: SingleComponentFockAddress

Address type that represents a Fock state of N spinless bosons in M modes. The particle number N can be set to missing. In the latter case the number of particles is not known at compile time and can be changed by excitations.

Constructors

BoseFS{[N,M]}(val::Integer...): Create BoseFS{N,M} from occupation numbers. This is type-stable if the number of modes M and the number of particles N are provided. Otherwise, M and N are inferred from the arguments.

BoseFS{missing}(arg; type=nothing): Create BoseFS{missing,M} from occupation numbers. The number of particles is not known at compile time and can be changed by excitations. The keyword argument type can be used to specify the type of the occupation numbers. It must be an unsigned integer type. If unspecified, the smallest unsigned integer type that can hold the maximum occupation number is chosen automatically.

BoseFS{[N,M]}(onr): Create BoseFS{N,M} from occupation number representation, see onr. This is efficient if N and M are provided, and onr is a statically-sized collection, such as a Tuple or SVector.

BoseFS{[N,M]}([M, ]pairs...): Provide the number of modes M and mode => occupation_number pairs. If M is provided as a type parameter, it should not be provided as the first argument. Useful for creating sparse addresses. pairs can be multiple arguments or an iterator of pairs.

BoseFS{N,M,S}(bs::S): Unsafe constructor. Does not check whether the number of particles in bs is equal to N.

@fs_str: Addresses are sometimes printed in a compact manner. This representation can also be used as a constructor. See the examples below.

Examples

julia> address = BoseFS(0, 1, 2, 3, 0)
BoseFS(0, 1, 2, 3, 0)

julia> num_modes(address), num_particles(address)
(5, 6)

julia> BoseFS(abs(i - 3) ≤ 1 ? i - 1 : 0 for i in 1:5)
BoseFS(0, 1, 2, 3, 0)

julia> BoseFS(5, 2 => 1, 3 => 2, 4 => 3) # sparse constructor
BoseFS(0, 1, 2, 3, 0)

julia> BoseFS{6,5}(i => i - 1 for i in 2:4)
BoseFS(0, 1, 2, 3, 0)

julia> fs"|0 1 2 3 0⟩" # \rangle(tab) -> ⟩
BoseFS(0, 1, 2, 3, 0)

julia> fs"|b 5: 2 3 3 4 4 4⟩" # compact sparse constructor
BoseFS(0, 1, 2, 3, 0)

julia> BoseFS{missing}(0, 1, 2, 3, 0) === fs"|0 1 2 3 0⟩{}" # missing particle number UInt8
true

julia> BoseFS{missing}(0, 1, 2, 3, 0; type=UInt16) === fs"|0 1 2 3 0⟩{UInt16}"
true

See also: SingleComponentFockAddress, FermiFS, CompositeFS, FermiFS2C, @fs_str.

Extended Help

The particle number N is a type parameter of the address and will be inferred by default, unless it is specifically set to missing. Having the particle number as a type parameter allows for type-stable code and optimizations. It should be used by default for number-conserving Hamiltonians.

Internally, there are three different storage types for BoseFS. The type S is chosen automatically based on the properties of the address. The storage types BitString and SortedParticleList are used for dense and sparse representations, respectively, when the number of particles is a type parameter and thus known at compile time.

When the number of particles is set to missing and not known at compile time, the occupation numbers are stored in a statically-sized vector of type SVector{M,T} where T is an unsigned integer type. For a type-stable constructor with missing particle number, use

BoseFS{missing}(v::SVector{M,T}) where {M,T<:Unsigned}

Rimu.BitStringAddresses.HardcoreBoseFS — Type

HardcoreBoseFS{N,M,S} <: SingleComponentFockAddress

Address type that represents a Fock state of N hardcore bosons in M modes, with occupancies restricted to 0 or 1 per mode, by wrapping a BitString or a SortedParticleList. Which is wrapped is chosen automatically based on the properties of the address. Set N to missing if the number of particles is not known at compile time and can be changed by excitations.

Constructors

HardcoreBoseFS{[N,M]}(val::Integer...): Create HardcoreBoseFS{N,M} from occupation numbers. This is type-stable if the number of modes M and the number of particles N are provided. Otherwise, M and N are inferred from the arguments.

HardcoreBoseFS{[N,M]}(onr): Create HardcoreBoseFS{N,M} from occupation number representation, see onr. This is efficient if N and M are provided, and onr is a statically-sized collection, such as a Tuple{M} or SVector{M}.

HardcoreBoseFS{[N,M]}([M, ]pairs...): Provide the number of modes M and pairs of the form mode => 1. If M is provided as a type parameter, it should not be provided as the first argument. Useful for creating sparse addresses. pairs can be multiple arguments or an iterator of pairs.

HardcoreBoseFS{N,M,S}(bs::S): Unsafe constructor. Does not check whether the number of particles in bs is equal to N, or whether each mode only contains one particle.

@fs_str: Addresses are sometimes printed in a compact manner. This representation can also be used as a constructor. See the examples below.

See the examples below.

Examples

julia> HardcoreBoseFS(1, 1, 1, 0, 0)
HardcoreBoseFS(1, 1, 1, 0, 0)

julia> HardcoreBoseFS([abs(i - 2) ≤ 1 for i in 1:5])
HardcoreBoseFS(1, 1, 1, 0, 0)

julia> HardcoreBoseFS(5, 1 => 1, 2 => 1, 3 => 1)
HardcoreBoseFS(1, 1, 1, 0, 0)

julia> HardcoreBoseFS{3,5}(i => 1 for i in 1:3)
HardcoreBoseFS(1, 1, 1, 0, 0)

julia> fs"|●●●∘∘⟩" # \mdlgblkcircle(tab) -> ●, \circ(tab) -> ∘, \rangle(tab) -> ⟩
HardcoreBoseFS(1, 1, 1, 0, 0)

julia> HardcoreBoseFS{missing}(1, 1, 1, 0, 0) == fs"|●●●∘∘⟩{}" # missing particle number
true

julia> fs"|h 5: 1 2 3⟩"
HardcoreBoseFS(1, 1, 1, 0, 0)

See also: SingleComponentFockAddress, BoseFS, CompositeFS, FermiFS, BitString, @fs_str.

Rimu.BitStringAddresses.bose_hubbard_interaction — Method

bose_hubbard_interaction(address)

Return $Σ_i n_i (n_i-1)$ for computing the Bose-Hubbard on-site interaction (without the $U$ prefactor.)

Example

julia> Hamiltonians.bose_hubbard_interaction(BoseFS{4,4}((2,1,1,0)))
2
julia> Hamiltonians.bose_hubbard_interaction(BoseFS{4,4}((3,0,1,0)))
6

Rimu.BitStringAddresses.hopnextneighbour — Method

new_address, value = hopnextneighbour(add, chosen, boundary_condition)

Compute the new address of a hopping event for the Hubbard model. Returns the new address and the square root of product of occupation numbers of the involved modes multiplied by a term consistent with boundary condition as the value. The following boundary conditions are supported:

:periodic: hopping over the boundary gives does not change the value.

:twisted: hopping over the boundary flips the sign of the value.

:hard_wall: hopping over the boundary gives a value of zero.

θ <: Number: hopping over the boundary gives a value multiplied by $\exp(iθ)$ or $\exp(−iθ)$ depending on the direction of hopping.

The off-diagonals are indexed as follows:

(chosen + 1) ÷ 2 selects the hopping site.

Even chosen indicates a hop to the left.

Odd chosen indicates a hop to the right.

Example

julia> using Rimu.Hamiltonians: hopnextneighbour

julia> hopnextneighbour(BoseFS(1, 0, 1), 3)
(BoseFS(2, 0, 0), 1.4142135623730951)

julia> hopnextneighbour(BoseFS(1, 0, 1), 4)
(BoseFS(1, 1, 0), 1.0)

julia> hopnextneighbour(BoseFS(1, 0, 1), 3, :twisted)
(BoseFS(2, 0, 0), -1.4142135623730951)

julia> hopnextneighbour(BoseFS(1, 0, 1), 3, :hard_wall)
(BoseFS(2, 0, 0), 0.0)

julia> hopnextneighbour(BoseFS(1, 0, 1), 3, π/4)
(BoseFS(2, 0, 0), 1.0000000000000002 + 1.0im)

Rimu.BitStringAddresses.near_uniform — Method

near_uniform(T::Type{<:SingleComponentFockAddress{N,M}}) → address::T
near_uniform(T::Type{<:SingleComponentFockAddress}, N::Integer, M::Integer) → address::T
near_uniform(address::SingleComponentFockAddress) → address::typeof(address)

Create a single component Fock state with M modes and N particles with near uniform occupation numbers.

Examples

julia> near_uniform(BoseFS{7,5})
BoseFS(2, 2, 1, 1, 1)

julia> near_uniform(FermiFS{3,5})
FermiFS(1, 1, 1, 0, 0)

julia> near_uniform(HardcoreBoseFS{missing}, 3, 5)
HardcoreBoseFS{missing}(1, 1, 1, 0, 0)

julia> near_uniform(BoseFS(10,0,0,0))
BoseFS(3, 3, 2, 2)

Rimu.BitStringAddresses.FermiFS — Type

FermiFS{N,M,S} <: SingleComponentFockAddress

Address type that represents a Fock state of N fermions of the same spin in M modes by wrapping a BitString, or a SortedParticleList. Which is wrapped is chosen automatically based on the properties of the address. Set N to missing if the number of particles is not known at compile time and can be changed by excitations.

Constructors

FermiFS{[N,M]}(val::Integer...): Create FermiFS{N,M} from occupation numbers. This is type-stable if the number of modes M and the number of particles N are provided. Otherwise, M and N are inferred from the arguments.

FermiFS{[N,M]}(onr): Create FermiFS{N,M} from occupation number representation, see onr. This is efficient if N and M are provided, and onr is a statically-sized collection, such as a Tuple{M} or SVector{M}.

FermiFS{[N,M]}([M, ]pairs...): Provide the number of modes M and pairs of the form mode => 1. If M is provided as a type parameter, it should not be provided as the first argument. Useful for creating sparse addresses. pairs can be multiple arguments or an iterator of pairs.

FermiFS{N,M,S}(bs::S): Unsafe constructor. Does not check whether the number of particles in bs is equal to N, or whether each mode only contains one particle.

@fs_str: Addresses are sometimes printed in a compact manner. This representation can also be used as a constructor. See the examples below.

Examples

julia> FermiFS(0, 1, 1, 1, 0)
FermiFS(0, 1, 1, 1, 0)

julia> FermiFS([abs(i - 3) ≤ 1 for i in 1:5])
FermiFS(0, 1, 1, 1, 0)

julia> FermiFS(5, 2 => 1, 3 => 1, 4 => 1)
FermiFS(0, 1, 1, 1, 0)

julia> FermiFS{3,5}(i => 1 for i in 2:4)
FermiFS(0, 1, 1, 1, 0)

julia> fs"|⋅↑↑↑⋅⟩" # \uparrow(tab) -> ↑, \cdot(tab) -> ⋅, \rangle(tab) -> ⟩
FermiFS(0, 1, 1, 1, 0)

julia> FermiFS{missing}(0, 1, 1, 1, 0) == fs"|⋅↑↑↑⋅⟩{}" # missing particle number
true

julia> fs"|f 5: 2 3 4⟩"
FermiFS(0, 1, 1, 1, 0)

See also: SingleComponentFockAddress, BoseFS, CompositeFS, FermiFS2C, BitString, @fs_str.

Rimu.BitStringAddresses.unoccupied_mode_map — Method

unoccupied_mode_map(address::Union{FermiFS, HardcoreBoseFS}) <: AbstractVector

Get a map of unoccupied modes in address as an AbstractVector of indices compatible with excitation.

unoccupied_mode_map(address)[i] contains the index for the i-th unoccupied mode. This is useful because unoccupied modes is required in some cases. unoccupied_mode_map(address) is an eager version of the iterator returned by unoccupied_modes. It is similar to onr but contains more information.

Note that this function is only implemented for addresses of type FermiFS and HardcoreBoseFS.

Example

julia> f = FermiFS(1,1,0,0)
FermiFS(1, 1, 0, 0)

julia> mf = unoccupied_mode_map(f)
2-element Rimu.BitStringAddresses.ModeMap{2, FermiFSIndex}:
 FermiFSIndex(occnum=0, mode=3, offset=2)
 FermiFSIndex(occnum=0, mode=4, offset=3)

julia> mf == collect(unoccupied_modes(f))
true

See also occupied_mode_map.

Rimu.BitStringAddresses.CompositeFS — Type

CompositeFS(addresses::SingleComponentFockAddress...) <: AbstractFockAddress

Address type used to encode addresses for multi-component models or spinor quantum particles.

Examples

julia> c1 = CompositeFS(FermiFS(1,0,0), FermiFS(1,1,0), BoseFS{missing}(0,0))
CompositeFS(
  FermiFS(1, 0, 0),
  FermiFS(1, 1, 0),
  BoseFS{missing}(0, 0),
)

julia> num_components(c1)
3

julia> num_modes(c1)
(3, 3, 2)

julia> fs"|↑⋅⋅⟩ ⊗ |↑↑⋅⟩ ⊗ |0 0⟩{}" == c1 # compact representation for printing
true

julia> FermiFS2C{missing}((1,1,0), (1,0,1)) # special case for spin-1/2 fermions
CompositeFS(
  FermiFS{missing}(1, 1, 0),
  FermiFS{missing}(1, 0, 1),
)

See also: FermiFS2C, BoseFS, FermiFS, num_components, num_modes, num_modes_are_equal, num_modes_check_equal, num_particles, maximum_mode_occupation, @fs_str, SingleComponentFockAddress, AbstractFockAddress.

Rimu.BitStringAddresses.FermiFS2C — Type

FermiFS2C <: AbstractFockAddress
FermiFS2C(onr_a, onr_b)
FermiFS2C{missing}(onr_a, onr_b)

Fock state address with two fermionic (spin) components. Alias for CompositeFS with two FermiFS components. If the type parameter missing is specified, the number of particles is not known at compile time. This is useful for spinful fermionic systems where the number of particles in each spin channel may vary.

Construct by specifying either two compatible FermiFSs, two onrs, or the number of modes followed by mode => occupation_number pairs, where occupation_number = 1 will put a particle in the first component and occupation_number = -1 will put a particle in the second component. See examples below.

Examples

julia> FermiFS2C(FermiFS(1,0,0), FermiFS(0,1,1))
CompositeFS(
  FermiFS(1, 0, 0),
  FermiFS(0, 1, 1),
)

julia> FermiFS2C((1,0,0), (0,1,1))
CompositeFS(
  FermiFS(1, 0, 0),
  FermiFS(0, 1, 1),
)

julia> FermiFS2C{missing}((1,0,0), (0,1,1)) # number non-conserving, spin flips allowed
CompositeFS(
  FermiFS{missing}(1, 0, 0),
  FermiFS{missing}(0, 1, 1),
)

julia> FermiFS2C{missing}(3, 1 => 1, 2 => -1, 3 => -1)
CompositeFS(
  FermiFS{missing}(1, 0, 0),
  FermiFS{missing}(0, 1, 1),
)

julia> fs"|↑↓↓⟩" # \uparrow(tab) -> ↑, \downarrow(tab) -> ↓, \rangle(tab) -> ⟩
CompositeFS(
  FermiFS(1, 0, 0),
  FermiFS(0, 1, 1),
)

julia> fs"|↑↓↓⇅⟩{}" # \dblarrowupdown(tab) -> ⇅
CompositeFS(
  FermiFS{missing}(1, 0, 0, 1),
  FermiFS{missing}(0, 1, 1, 1),
)

See also: CompositeFS, FermiFS, @fs_str.

Rimu.BitStringAddresses.time_reverse — Method

time_reverse(addr)

Apply the time-reversal operation on a two-component Fock address that flips all the spins.

Requires each component address to have the same type.

Internal representations

The addresses types BoseFS, FermiFS and HardcoreBoseFS, where the particle number is either fixed (for fixed particle number BoseFS) or severely constrained, are implemented as either bitstrings through BitString, or sorted lists of particles with SortedParticleList. This allows for a space efficient representation.

Therewhile, BoseFS{missing}, which allows for flexible and unconstrained particle number internally uses the occupation number representation, which allows it to handle excitation operations that change the particle number. This is fast but requires more storage space.

Internal APIs

Rimu.BitStringAddresses.BitString — Type

BitString{B,N,T<:Unsigned}

Type for storing bitstrings of static size. Holds B bits in N chunks, where each chunk is of type T.

N is chosen automatically to accommodate B bits as efficiently as possible.

Constructors

BitString{B,N,T}(::SVector{N,T}): unsafe constructor. Does not check for ghost bits.

BitString{B,N,T}(i::T): as above, but sets i as the rightmost chunk.

BitString{B}(::Integer): Convert integer to BitString. Integer is truncated to the correct number of bits.

Rimu.BitStringAddresses.SortedParticleList — Type

SortedParticleList{N,M,T<:Unsigned}

Type for storing sparse Fock states. Stores the mode number of each particle as an entry with only its mode stored. The entries are always kept sorted.

Iterating over SortedParticleLists yields occupied modes as a tuple of occupation number, mode number, and position in list.

Constructors

SortedParticleList{N,M,T}(::SVector{N,T}): unsafe constructor. Does not sort input.

SortedParticleList(arr::AbstractVector): convert ONR to SortedParticleList

The following APIs are used by Module Hamiltonians.

Rimu.BitStringAddresses.ModeMap — Type

ModeMap <: AbstractVector

A unified storage structure for indices of SingleComponentFockAddress. It stores the FSIndex of corresponding address as an AbstractVector compatible with excitation - BoseFSIndex or FermiFSIndex.

This struct is not intended to be constructed directly. Use occupied_mode_map or unoccupied_mode_map to obtain an instance.

See also SingleComponentFockAddress.

Rimu.BitStringAddresses.FermiFS2CModes — Type

FermiFS2CModes

This struct stores the occupied and unoccupied mode maps associated with an address of type FermiFS2C. It should be constructed using the full_mode_maps function.

The struct has two fields, occupied and unoccupied, each containing a ModeMap represented as a two-element Tuple:

Index 1 corresponds to the α spin channel

Index 2 corresponds to the β spin channel

This convention follows the spin-channel indexing defined in FermiFS2C.

See also FermiFS2C, ModeMap, occupied_mode_map, and unoccupied_mode_map.

Rimu.BitStringAddresses.full_mode_maps — Function

full_mode_maps(addr::FermiFS2C)

The constructor function of FermiFS2CModes.

Index

Rimu.BitStringAddresses

Rimu.BitStringAddresses.BitString

Rimu.BitStringAddresses.BoseFS

Rimu.BitStringAddresses.BoseFSIndex

Rimu.BitStringAddresses.CompositeFS

Rimu.BitStringAddresses.FermiFS

Rimu.BitStringAddresses.FermiFS2C

Rimu.BitStringAddresses.FermiFS2CModes

Rimu.BitStringAddresses.FermiFSIndex

Rimu.BitStringAddresses.HardcoreBoseFS

Rimu.BitStringAddresses.ModeMap

Rimu.BitStringAddresses.OccupiedPairsMap

Rimu.BitStringAddresses.SingleComponentFockAddress

Rimu.BitStringAddresses.SortedParticleList

Rimu.Interfaces.AbstractFockAddress

Rimu.BitStringAddresses.bose_hubbard_interaction

Rimu.BitStringAddresses.each_mode

Rimu.BitStringAddresses.excitation

Rimu.BitStringAddresses.find_mode

Rimu.BitStringAddresses.find_occupied_mode

Rimu.BitStringAddresses.full_mode_maps

Rimu.BitStringAddresses.hopnextneighbour

Rimu.BitStringAddresses.near_uniform

Rimu.BitStringAddresses.num_occupied_modes

Rimu.BitStringAddresses.occupied_mode_map

Rimu.BitStringAddresses.occupied_modes

Rimu.BitStringAddresses.onr

Rimu.BitStringAddresses.time_reverse

Rimu.BitStringAddresses.unoccupied_mode_map

Rimu.BitStringAddresses.unoccupied_modes

Rimu.Interfaces.num_components

Rimu.Interfaces.num_modes

Rimu.Interfaces.num_modes_are_equal

Rimu.Interfaces.num_modes_check_equal

Rimu.Interfaces.num_particles

Rimu.BitStringAddresses.@fs_str