Module DictVectors

Rimu.DictVectors — Module

Module that provides concrete implementations of the AbstractDVec interface.

DVec: basic AbstractDVec

InitiatorDVec: allows storing information about initiator status

See Interfaces.

Rimu.Interfaces.AbstractDVec — Type

AbstractDVec{K,V}

Abstract type for data structures that behave similar to sparse vectors, but are indexed by an arbitrary type V (could be non-integers) similarly to dictionaries. AbstractDVecs are designed to work well with lomc! and KrylovKit.

Concrete implementations are available as DVec and InitiatorDVec.

AbstractDvecs lie somewhere between AbstractDicts and sparse AbstractVectors, while being subtyped to neither. Generally they behave like a dictionary, while supportting various linear algebra functionality. Indexing with a value not stored in the dictionary returns zero(V). Setting a stored value to 0 or below eps(V::AbstractFloat) removes the value from the dictionary. Their length signals the number of stored elements, not the size of the vector space.

They have a StochasticStyle which selects the spawning algorithm in FCIQMC.

To iterate over an AbstractDVec, use keys, pairs, or values.

Interface

The interface is similar to the AbstractDict interface, but with the changed behaviour as noted above. Implement what would be needed for the AbstractDict interface (pairs, keys, values, setindex!, getindex, delete!, length, haskey, empty!, isempty) and, in addition:

StochasticStyle

storage returns an AbstractDict storing the raw data with possibly different valtype than V.

deposit!

See also DictVectors, Interfaces.

Concrete implementations

Rimu.DictVectors.DVec — Type

DVec{K,V,D<:AbstractDict{K,V},S}

Dictionary-based vector-like data structure for use with FCIQMC and KrylovKit. While mostly behaving like a Dict, it supports various linear algebra operations such as norm and dot. It has a StochasticStyle that is used to select an appropriate spawning strategy in the FCIQMC algorithm.

See also: AbstractDVec.

Constructors

DVec(dict::AbstractDict[; style, capacity]): create a DVec with dict for storage. Note that the data may or may not be copied.

DVec(args...[; style, capacity]): args... are passed to the Dict constructor. The Dict is used for storage.

DVec{K,V}([; style, capacity]): create an empty DVec{K,V}.

DVec(dv::AbstractDVec[; style, capacity]): create a DVec with the same contents as adv. The style is inherited from dv by default.

The default style is selected based on the DVec's valtype (see default_style). If a style is given and the valtype does not match the style's eltype, the values are converted to an appropriate type.

The capacity argument is optional and sets the initial size of the DVec via Base.sizehint!.

Examples

julia> dv = DVec(:a => 1)
DVec{Symbol,Int64} with 1 entry, style = IsStochasticInteger{Int64}()
  :a => 1

julia> dv = DVec(:a => 2, :b => 3; style=IsDeterministic())
DVec{Symbol,Float64} with 2 entries, style = IsDeterministic{Float64}()
  :a => 2.0
  :b => 3.0

Rimu.DictVectors.InitiatorDVec — Type

InitiatorDVec{K,V} <: AbstractDVec{K,V}

Dictionary-based vector-like data structure for use with lomc! and KrylovKit.jl. See AbstractDVec. Functionally identical to DVec, but contains InitiatorValues internally in order to facilitate initiator methods. How the initiators are handled is controlled by the initiator keyword argument (see below).

Constructors

InitiatorDVec(dict::AbstractDict[; style, initiator, capacity]): create an InitiatorDVec with dict for storage. Note that the data may or may not be copied.

InitiatorDVec(args...[; style, initiator, capacity]): args... are passed to the Dict constructor. The Dict is used for storage.

InitiatorDVec{K,V}([; style, initiator, capacity]): create an empty InitiatorDVec{K,V}.

InitiatorDVec(dv::AbstractDVec[; style, initiator, capacity]): create an InitiatorDVec with the same contents as dv. The style is inherited from dv by default.

Keyword arguments

style: A valid StochasticStyle. The default is selected based on the InitiatorDVec's valtype (see default_style). If a style is given and the valtype does not match the style's eltype, the values are converted to an appropriate type.

initiator = Initiator(1): A valid InitiatorRule. See Initiator.

capacity: Indicative size as Int. Optional. Sets the initial size of the InitiatorDVec via Base.sizehint!.

Interface functions

Rimu.Interfaces.deposit! — Function

deposit!(w::AbstractDVec, add, val, parent::Pair)

Add val into w at address add, taking into account initiator rules if applicable. parent contains the address => value pair from which the pair add => val was created. InitiatorDVec can intercept this and add its own functionality.

deposit!(w::InitiatorDVec, add, val, p_add=>p_val)

Add val into w at address add as an InitiatorValue.

Rimu.Interfaces.storage — Function

storage(dvec) -> AbstractDict

Return the raw storage associated with dvec as an AbstractDict. Used in MPI communication.

OrderedCollections.freeze — Function

freeze(dv)

Create a "frozen" version of dv which can no longer be modified or used in the conventional manner, but supports faster dot products.

If dv is an MPIData, synchronize its contents among the ranks first.

Rimu.Interfaces.localpart — Function

localpart(dv) -> AbstractDVec

Get the part of dv that is located on this MPI rank. Returns dv itself for AbstractDVecs.

Supported operations

AbstractDVecs generally support most operations that are defined on Vectors and Dicts. This includes common linear algebra operations such as dot or norm.

A significant difference between AbstractDVecs, Vectors, and Dicts, is that iteration on them is disabled by default. Iteration must be explicitly performed on keys, calues, or pairs, however, it is highly recommended you use mapreduce, reduce, or similar functions when performing reductions, as that will make the operations compatible with MPI.

In addition, Rimu defines the following functions.

Rimu.DictVectors.zero! — Function

zero!(v)

Replace v by a zero vector as an inplace operation. For AbstractDVec types it means removing all non-zero elements. For AbstractArrays, it sets all of the values to zero.

Rimu.DictVectors.add! — Function

add!(x::AbstractDVec,y::AbstactDVec)

Inplace add x+y and store result in x.

Rimu.DictVectors.walkernumber — Function

walkernumber(md::MPIData)

Compute the walkernumber of the distributed data on every MPI rank with MPI.Allreduce. MPI syncronizing.

walkernumber(w)

Compute the number of walkers in w. It is used for updating the shift. Overload this function for modifying population control.

In most cases walkernumber(w) is identical to norm(w,1). For AbstractDVecs with complex coefficients it reports the one norm separately for the real and the imaginary part as a ComplexF64. See Norm1ProjectorPPop.

Projectors

Rimu.DictVectors.AbstractProjector — Type

Abstract supertype for projectors to be used in in lieu of DVecs or Vectors in dot products. Implemented subtypes:

UniformProjector

NormProjector

Norm2Projector

Norm1ProjectorPPop

See also PostStepStrategy for use of projectors in lomc!.

Interface

Define a method for LinearAlgebra.dot(projector, v).

Rimu.DictVectors.NormProjector — Type

NormProjector() <: AbstractProjector

Results in computing the one-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,1)

NormProjector() thus represents the vector sign.(x).

See also PostStepStrategy, and AbstractProjector for use of projectors in lomc!.

Rimu.DictVectors.Norm2Projector — Type

Norm2Projector() <: AbstractProjector

Results in computing the two-norm when used in dot(). E.g.

dot(NormProjector(),x)
-> norm(x,2) # with type Float64

See also PostStepStrategy, and AbstractProjector for use of projectors in lomc!.

Rimu.DictVectors.UniformProjector — Type

UniformProjector() <: AbstractProjector

Represents a vector with all elements 1. To be used with dot(). Minimizes memory allocations.

UniformProjector()⋅v == sum(v)
dot(UniformProjector(), LO, v) == sum(LO*v)

See also PostStepStrategy, and AbstractProjector for use of projectors in lomc!.

Rimu.DictVectors.Norm1ProjectorPPop — Type

Norm1ProjectorPPop() <: AbstractProjector

Results in computing the one-norm per population when used in dot(). E.g.

dot(Norm1ProjectorPPop(),x)
-> norm(real.(x),1) + im*norm(imag.(x),1)

See also PostStepStrategy, and AbstractProjector for use of projectors in lomc!.

Initiator rules

Rimu.DictVectors.InitiatorRule — Type

InitiatorRule{V}

Abstract type for defining initiator rules for InitiatorDVec. Concrete implementations:

Initiator

SimpleInitiator

CoherentInitiator

When defining a new InitiatorRule, also define a corresponding method for value!

Rimu.DictVectors.InitiatorValue — Type

InitiatorValue{V}(; safe::V, unsafe::V, initiator::V) where V

Composite "walker" with three fields. For use with InitiatorDVecs.

Rimu.DictVectors.value — Function

value(i::InitiatorRule, v::InitiatorValue)

Convert the InitiatorValue v into a scalar value according to the InitiatorRule i.

Internal function that implements functionality of InitiatorDVec.

Rimu.DictVectors.Initiator — Type

Initiator(threshold) <: InitiatorRule

Initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators can spawn to initiators.

See InitiatorRule.

Rimu.DictVectors.SimpleInitiator — Type

SimpleInitiator(threshold) <: InitiatorRule

Simplified initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators cannot spawn.

See InitiatorRule.

Rimu.DictVectors.CoherentInitiator — Type

CoherentInitiator(threshold) <: InitiatorRule

Initiator rule to be passed to InitiatorDVec. An initiator is a configuration add with a coefficient with magnitude abs(v[add]) > threshold. Rules:

Initiators can spawn anywhere.

Non-initiators can spawn to initiators.

Multiple non-initiators can spawn to a single non-initiator if their contributions add up to a value greater than the initiator threshold.

See InitiatorRule.

Index

Rimu.DictVectors

Rimu.DictVectors.AbstractProjector

Rimu.DictVectors.CoherentInitiator

Rimu.DictVectors.DVec

Rimu.DictVectors.Initiator

Rimu.DictVectors.InitiatorDVec

Rimu.DictVectors.InitiatorRule

Rimu.DictVectors.InitiatorValue

Rimu.DictVectors.Norm1ProjectorPPop

Rimu.DictVectors.Norm2Projector

Rimu.DictVectors.NormProjector

Rimu.DictVectors.SimpleInitiator

Rimu.DictVectors.UniformProjector

Rimu.Interfaces.AbstractDVec

OrderedCollections.freeze

Rimu.DictVectors.add!

Rimu.DictVectors.value

Rimu.DictVectors.walkernumber

Rimu.DictVectors.zero!

Rimu.Interfaces.deposit!

Rimu.Interfaces.localpart

Rimu.Interfaces.storage