Developer Documentation

Module Hamiltionians.jl

This module contains definitions of Hamiltonians, in particular specific physical models of interest. These are organised by means of an interface around the abstract type AbstractHamiltonian, in the spirit of the AbstractArray interface as discussed in the Julia Documentation.

Rimu.Hamiltonians — Module

This module defines Hamiltonian types and standard methods. Model Hamiltonians should be subtyped to AbstractHamiltonian. Models implemented so far are:

BoseHubbardReal1D Bose-Hubbard chain, real space

ExtendedBHReal1D extended Bose-Hubbard model with on-site and nearest neighbour interactions, real space, one dimension

Relation to other parts of the Rimu code

In order to define a specific model Hamiltonian with relevant parameters for the model, instantiate the model like this in the input file:

ham = BoseHubbardReal1D(n=6, m=6, u=1.0, t=1.0, AT = BSAdd64)

In the rest of the Rimu code, access to properties and matrix elements of the model are then provided by the following methods:

ham[address1, address2]: indexing of matrix elements (slow - use with caution)

ham(dv::AbstractDVec) or mul!(dv1, ham, dv2): use as linear operator

diagME(ham, add): diagonal matrix element

numOfHops(ham, add): number of off-diagonals

hop(ham, add, chosen): access off-diagonal matrix element

Hops(ham, add): iterator over off-diagonal matrix elements

generateRandHop(hops::Hops): choose random off-diagonal

bit_String_Length(ham): number of bits in the configuration

ham(:dim) and ham(:fdim): dimension of linear space. See hasIntDimension(ham), dimensionLO(ham), fDimensionLO(ham)

nearUniform(ham): configuration with particles spread across modes

Model Hamiltonians

Here is a list of fully implemented model Hamiltonians. So far there are two variants implemented of the one-dimentional Bose-Hubbard model real space as well as a momentum-space Hubbard chain.

Rimu.Hamiltonians.BoseHubbardReal1D — Type

ham = BoseHubbardReal1D(;[n=6, m=6, u=1.0, t=1.0, AT = BSAdd64])

Implements a one-dimensional Bose Hubbard chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1)\]

Arguments

n::Int: the number of bosons

m::Int: the number of lattice sites

u::Float64: the interaction parameter

t::Float64: the hopping strength

AT::Type: the address type

Functor use:

w = ham(v)
ham(w, v)

Compute the matrix - vector product w = ham * v. The two-argument version is mutating for w.

ham(:dim)

Return the dimension of the linear space if representable as Int, otherwise return nothing.

ham(:fdim)

Return the approximate dimension of linear space as Float64.

Rimu.Hamiltonians.ExtendedBHReal1D — Type

ham = ExtendedBHReal1D(n=6, m=6, u=1.0, v=1.0, t=1.0, AT=BSAdd64)

Implements the extended Bose Hubbard model on a one-dimensional chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1) + v \sum_{\langle i,j\rangle} n_i n_j\]

Arguments

n::Int: number of bosons

m::Int: number of lattice sites

u::Float64: on-site interaction parameter

v::Float64: the next-neighbor interaction

t::Float64: the hopping strength

AT::Type: address type for identifying configuration

Rimu.Hamiltonians.BoseHubbardMom1D — Type

ham = BoseHubbardMom1D(;[n=6, m=6, u=1.0, t=1.0], add = add)
ham = BoseHubbardMom1D(add; u=1.0, t=1.0)

Implements a one-dimensional Bose Hubbard chain in momentum space.

\[\hat{H} = -t \sum_{k} ϵ_k n_k + \frac{u}{M}\sum_{kpqr} a^†_{r} a^†_{q} a_p a_k δ_{r+q,p+k}\\ ϵ_k = - 2 t \cos(k)\]

Arguments

n::Int: the number of bosons

m::Int: the number of lattice sites

u::Float64: the interaction parameter

t::Float64: the hopping strength

AT::Type: the address type

Functor use:

w = ham(v)
ham(w, v)

Compute the matrix - vector product w = ham * v. The two-argument version is mutating for w.

ham(:dim)

Return the dimension of the linear space if representable as Int, otherwise return nothing.

ham(:fdim)

Return the approximate dimension of linear space as Float64.

Hamiltonians interface

Behind the implementation of a particular model is a more abstract interface for defining hamiltonians. If you want to define a new model you should make use of this interface. The most general form of a model Hamiltonian should subtype to AbstractHamiltonian and implement the relevant methods.

Rimu.Hamiltonians.AbstractHamiltonian — Type

AbstractHamiltonian{T}

Supertype that provides an interface for linear operators over a linear space with scalar type T that are suitable for FCIQMC. Indexing is done with addresses (typically not integers) from an address space that may be large (and will not need to be completely generated).

AbstractHamiltonian instances operate on vectors of type AbstractDVec from the module DictVectors and work well with addresses of type BitStringAddressType from the module BitStringAddresses. The type works well with the external package KrylovKit.jl.

Provides:

Hops: iterator over reachable off-diagonal matrix elements

generateRandHop: function to generate random off-diagonal matrix element

hamiltonian[address1, address2]: indexing with getindex() - mostly for testing purposes

*(LO, v) deterministic matrix-vector multiply (== LO(v))

mul!(w, LO, v) mutating matrix-vector multiply

dot(x, LO, v) compute x⋅(LO*v) minimizing allocations

Methods that need to be implemented:

numOfHops(lo::AbstractHamiltonian, address)

hop(lo::AbstractHamiltonian, address, chosen::Integer)

diagME(lo::AbstractHamiltonian, address)

hasIntDimension(lo::AbstractHamiltonian)

dimensionLO(lo::AbstractHamiltonian), if applicable

fDimensionLO(lo::AbstractHamiltonian)

Optional:

Hamiltonians.LOStructure(::Type{typeof(lo)})

Rimu.Hamiltonians.Hops — Type

Hops(ham, add)

Iterator over new address and matrix element for reachable off-diagonal matrix elements of linear operator ham from address add. Represents an abstract vector containing the possibly non-zero off-diagonal matrix elements of the column of ham indexed by add.

Examples

new_address, matrix_element = Hops(ham, current_address)[i]
number_of_hops = length(Hops(ham, current_address))
for (add,elem) in Hops(ham, current_address)
   # do something with address and elem
end

Rimu.Hamiltonians.generateRandHop — Function

generateRandHop(ham, add)
generateRandHop(hops::Hops)

Generate a single random excitation, i.e. choose from one of the accessible off-diagonal elements in the column corresponding to address add of the Hamiltonian matrix represented by ham. Alternatively, pass as argument an iterator over the accessible matrix elements.

Core functions

The following functions are part of the core functionality of a Hamiltonian and need to be implemented efficiently and specifically for each model.

Rimu.Hamiltonians.numOfHops — Function

numOfHops(ham, add)

Compute the number of number of reachable configurations from address add.

Rimu.Hamiltonians.hop — Function

newadd, me = hop(ham, add, chosen)

Compute matrix element of hamiltonian and new address of a single hop from address add with integer index chosen.

Rimu.Hamiltonians.diagME — Function

diagME(ham, add)

Compute the diagonal matrix element of the linear operator ham at address add.

BosonicHamiltonian

For a many-body system consisting of spinless bosons, we already know more about the structure of the problem. BosonicHamiltonian is a pre-defined subtype of AbstractHamiltonian.

Rimu.Hamiltonians.BosonicHamiltonian — Type

BosonicHamiltonian{T} <: AbstractHamiltonian{T}

Abstract type for representing Hamiltonians in a Fock space of fixed number of scalar bosons. At least the following fields should be present:

n # number of particles

m # number of modes

AT # address type

Methods that need to be implemented:

numOfHops(lo::AbstractHamiltonian, address) - number of off-diagonal matrix elements

hop(lo::AbstractHamiltonian, address, chosen::Integer) - access an off-diagonal m.e. by index chosen

diagME(lo::AbstractHamiltonian, address) - diagonal matrix element

Optional:

Hamiltonians.LOStructure(::Type{typeof(lo)}) - can speed up deterministic calculations if HermitianLO

Provides:

hasIntDimension(lo::AbstractHamiltonian)

dimensionLO(lo::AbstractHamiltonian), might fail if linear space too large

fDimensionLO(lo::AbstractHamiltonian)

bit_String_Length

nearUniform, default version

Rimu.Hamiltonians.hasIntDimension — Function

hasIntDimension(ham)

Return true if dimension of the linear operator ham can be computed as an integer and false if not.

If true, dimensionLO(h) will be successful and return an Int. The method fDimensionLO(h) should be useful in other cases.

Rimu.Hamiltonians.dimensionLO — Function

dimensionLO(hamiltonian)

Compute dimension of linear operator as integer.

Rimu.Hamiltonians.fDimensionLO — Function

fDimensionLO(hamiltonian)

Returns the dimension of Hilbert space as Float64. The exact result is returned if the value is smaller than 2^53. Otherwise, an improved Stirling formula is used.

Rimu.Hamiltonians.bit_String_Length — Function

bit_String_Length(ham)

Number of bits needed to represent an address for the linear operator ham.

Rimu.BitStringAddresses.nearUniform — Function

nearUniform(BoseFS{N,M})
nearUniform(BoseFS{N,M,A}) -> bfs::BoseFS{N,M,A}

Create bosonic Fock state with near uniform occupation number of M modes with a total of N particles. Specifying the bit address type A is optional.

Examples

julia> nearUniform(BoseFS{7,5,BitAdd})
BoseFS{BitAdd}((2,2,1,1,1))

julia> nearUniform(BoseFS{7,5})
BoseFS{BSAdd64}((2,2,1,1,1))

nearUniform(ham)

Create bitstring address with near uniform distribution of particles across modes for the Hamiltonian ham.