Module Hamiltionians.jl

This module contains definitions of Hamiltonians, in particular specific physical models of interest. These are organised by means of an interface around the abstract type AbstractHamiltonian, in the spirit of the AbstractArray interface as discussed in the Julia Documentation.

Rimu.Hamiltonians — Module

module Hamiltonians

This module defines Hamiltonian types, interfaces, and functions for working with Hamiltonians.

Relation to other parts of the Rimu code

In order to define a specific model Hamiltonian with relevant parameters for the model, instantiate the model like this in the input file:

ham = HubbardReal1D(BoseFS((1,2,0,3)); u=1.0, t=1.0)

In the rest of the Rimu code, access to properties and matrix elements of the model are then provided by the following methods:

ham[address1, address2]: indexing of matrix elements (slow - use with caution)

ham * dv, ham(dv::AbstractDVec) or mul!(dv1, ham, dv2): use as linear operator

diagonal_element(ham, add): diagonal matrix element

num_offdiagonals(ham, add): number of off-diagonals

get_offdiagonal(ham, add, chosen): access off-diagonal matrix element

offdiagonals(ham, add): iterator over off-diagonal matrix elements

random_offdiagonal(hops): choose random off-diagonal

dimension(T, ham): dimension of linear space

nearUniform(ham): configuration with particles spread across modes

starting_address(ham): address for accessing one of the diagonal elements of ham

Model Hamiltonians

Here is a list of fully implemented model Hamiltonians. So far there are two variants implemented of the one-dimensional Bose-Hubbard model real space as well as a momentum-space Hubbard chain.

Rimu.Hamiltonians.HubbardReal1D — Type

HubbardReal1D(address; u=1.0, t=1.0)

Implements a one-dimensional Bose Hubbard chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1)\]

Arguments

address: the starting address, defines number of particles and sites.

u: the interaction parameter.

t: the hopping strength.

See also

HubbardMom1D

ExtendedHubbardReal1D

Rimu.Hamiltonians.ExtendedHubbardReal1D — Type

ExtendedHubbardReal1D(address; u=1.0, v=1.0, t=1.0)

Implements the extended Hubbard model on a one-dimensional chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1) + v \sum_{\langle i,j\rangle} n_i n_j\]

Arguments

address: the starting address.

u: on-site interaction parameter

v: the next-neighbor interaction

t: the hopping strength

Rimu.Hamiltonians.HubbardMom1D — Type

HubbardMom1D(address; u=1.0, t=1.0)

Implements a one-dimensional Bose Hubbard chain in momentum space.

\[\hat{H} = \sum_{k} ϵ_k n_k + \frac{u}{M}\sum_{kpqr} a^†_{r} a^†_{q} a_p a_k δ_{r+q,p+k}\\ ϵ_k = -2t \cos(k)\]

Arguments

address: the starting address, defines number of particles and sites.

u: the interaction parameter.

t: the hopping strength.

See also

HubbardReal1D

ExtendedHubbardReal1D

Rimu.Hamiltonians.MatrixHamiltonian — Type

MatrixHamiltonian(mat::AbstractMatrix{T}; starting_address=1) <: AbstractHamiltonian{T}

Wrap an abstract matrix mat as an AbstractHamiltonian object for use with regular Vectors indexed by integers. Works with stochatic methods of lomc!(). Optionally, a starting_address can be provided.

Specialised methods are implemented for sparse matrices of type AbstractSparseMatrixCSC.

Hamiltonians interface

Behind the implementation of a particular model is a more abstract interface for defining hamiltonians. If you want to define a new model you should make use of this interface. The most general form of a model Hamiltonian should subtype to AbstractHamiltonian and implement the relevant methods.

Rimu.Hamiltonians.AbstractHamiltonian — Type

AbstractHamiltonian{T}

Supertype that provides an interface for linear operators over a linear space with scalar type T that are suitable for FCIQMC. Indexing is done with addresses (typically not integers) from an address space that may be large (and will not need to be completely generated).

AbstractHamiltonian instances operate on vectors of type AbstractDVec from the module DictVectors and work well with addresses of type AbstractFockAddress from the module BitStringAddresses. The type works well with the external package KrylovKit.jl.

Methods

Provides:

offdiagonals: iterator over reachable off-diagonal matrix elements

random_offdiagonal: function to generate random off-diagonal matrix element

dimension: get the dimension of the address space.

H[address1, address2]: indexing with getindex() - mostly for testing purposes

*(H, v): deterministic matrix-vector multiply.

H(v): equivalent to H * v.

mul!(w, H, v): mutating matrix-vector multiply.

dot(x, H, v): compute x⋅(H*v) minimizing allocations.

Methods that need to be implemented:

num_offdiagonals(::AbstractHamiltonian, address)

get_offdiagonal(::AbstractHamiltonian, address, chosen::Integer)

diagonal_element(::AbstractHamiltonian, address)

starting_address(::AbstractHamiltonian)

Optional methods to implement:

Hamiltonians.LOStructure(::Type{typeof(lo)})

dimension(::Type{T}, ::AbstractHamiltonian)

offdiagonals(::AbstractHamiltonian, ::AbstractFockAddress)

momentum(::AbstractHamiltonian)

Rimu.Hamiltonians.offdiagonals — Function

offdiagonals(h::AbstractHamiltonian, address)

Return an iterator over reachable off-diagonal matrix elements of type <:AbstractOffdiagonals. Defaults to returning Offdiagonals(h, a)

See also

Offdiagonals

AbstractOffdiagonals

julia> addr = BoseFS((3,2,1));

julia> H = HubbardReal1D(addr);

julia> h = offdiagonals(H, addr)
6-element Rimu.Hamiltonians.Offdiagonals{Float64,BoseFS{6,3,BitString{8,1}},HubbardReal1D{Float64,BoseFS{6,3,BitString{8,1}},1.0,1.0}}:
 (BoseFS{6,3}((2, 3, 1)), -3.0)
 (BoseFS{6,3}((2, 2, 2)), -2.449489742783178)
 (BoseFS{6,3}((3, 1, 2)), -2.0)
 (BoseFS{6,3}((4, 1, 1)), -2.8284271247461903)
 (BoseFS{6,3}((4, 2, 0)), -2.0)
 (BoseFS{6,3}((3, 3, 0)), -1.7320508075688772)

Rimu.Hamiltonians.random_offdiagonal — Function

random_offdiagonal(offdiagonals::AbstractOffdiagonals)
random_offdiagonal(ham::AbstractHamiltonian, add)

Generate a single random excitation, i.e. choose from one of the accessible off-diagonal elements in the column corresponding to address add of the Hamiltonian matrix represented by ham. Alternatively, pass as argument an iterator over the accessible matrix elements.

Core functions

The following functions are part of the core functionality of a Hamiltonian and need to be implemented efficiently and specifically for each model.

Rimu.Hamiltonians.num_offdiagonals — Function

num_offdiagonals(ham, add)

Compute the number of number of reachable configurations from address add.

Example

julia> addr = BoseFS((3, 2, 1));

julia> H = HubbardMom1D(addr);

julia> num_offdiagonals(H, addr)
10

Rimu.Hamiltonians.get_offdiagonal — Function

newadd, me = get_offdiagonal(ham, add, chosen)

Compute value me and new address newadd of a single (off-diagonal) matrix element in a Hamiltonian ham. The off-diagonal element is in the same column as address add and is indexed by integer index chosen.

Example

julia> addr = BoseFS((3, 2, 1));

julia> H = HubbardMom1D(addr);

julia> get_offdiagonal(H, addr, 3)
(BoseFS{6,3}((2, 1, 3)), 1.0)

Rimu.Hamiltonians.diagonal_element — Function

diagonal_element(ham, add)

Compute the diagonal matrix element of the linear operator ham at address add.

Example

julia> addr = BoseFS((3, 2, 1));

julia> H = HubbardMom1D(addr);

julia> diagonal_element(H, addr)
8.666666666666664