Module Hamiltionians

This module contains definitions of Hamiltonians, in particular specific physical models of interest. These are organised by means of an interface around the abstract type AbstractHamiltonian, in the spirit of the AbstractArray interface as discussed in the Julia Documentation.

Rimu.Hamiltonians — Module

module Hamiltonians

This module defines Hamiltonian types and functions for working with Hamiltonians.

Exported concrete Hamiltonian types include

MatrixHamiltonian

HubbardReal1D

ExtendedHubbardReal1D

HubbardMom1D

BoseHubbardMom1D2C

BoseHubbardReal1D2C

Wrappers

GutzwillerSampling

GuidingVectorSampling

Other

G2Correlator

Interface for working with Hamiltonians

AbstractHamiltonian: defined in the module Interfaces

Relation to other parts of the Rimu code

In order to define a specific model Hamiltonian with relevant parameters for the model, instantiate the model like this in the input file:

ham = HubbardReal1D(BoseFS((1,2,0,3)); u=1.0, t=1.0)

In the rest of the Rimu code, access to properties and matrix elements of the model are then provided by the following methods:

ham[address1, address2]: indexing of matrix elements (slow - use with caution)

ham * dv, ham(dv::AbstractDVec) or mul!(dv1, ham, dv2): use as linear operator

diagonal_element(ham, add): diagonal matrix element

num_offdiagonals(ham, add): number of off-diagonals

get_offdiagonal(ham, add, chosen): access off-diagonal matrix element

offdiagonals(ham, add): iterator over off-diagonal matrix elements

random_offdiagonal(hops): choose random off-diagonal

dimension(T, ham): dimension of linear space

near_uniform(ham): configuration with particles spread across modes

starting_address(ham): address for accessing one of the diagonal elements of ham

Model Hamiltonians

Here is a list of fully implemented model Hamiltonians. So far there are two variants implemented of the one-dimensional Bose-Hubbard model real space as well as a momentum-space Hubbard chain.

Rimu.Hamiltonians.HubbardReal1D — Type

HubbardReal1D(address; u=1.0, t=1.0)

Implements a one-dimensional Bose Hubbard chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1)\]

Arguments

address: the starting address, defines number of particles and sites.

u: the interaction parameter.

t: the hopping strength.

See also

HubbardMom1D

ExtendedHubbardReal1D

Rimu.Hamiltonians.ExtendedHubbardReal1D — Type

ExtendedHubbardReal1D(address; u=1.0, v=1.0, t=1.0)

Implements the extended Hubbard model on a one-dimensional chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1) + v \sum_{\langle i,j\rangle} n_i n_j\]

Arguments

address: the starting address.

u: on-site interaction parameter

v: the next-neighbor interaction

t: the hopping strength

Rimu.Hamiltonians.HubbardMom1D — Type

HubbardMom1D(address; u=1.0, t=1.0, dispersion=hubbard_dispersion)

Implements a one-dimensional Bose Hubbard chain in momentum space.

\[\hat{H} = \sum_{k} ϵ_k n_k + \frac{u}{M}\sum_{kpqr} a^†_{r} a^†_{q} a_p a_k δ_{r+q,p+k}\]

Arguments

address: the starting address, defines number of particles and sites.

u: the interaction parameter.

t: the hopping strength.

dispersion: defines $ϵ_k =$t*dispersion(k)

hubbard_dispersion: $ϵ_k = -2t \cos(k)$

continuum_dispersion: $ϵ_k = tk^2$

See also

HubbardReal1D

ExtendedHubbardReal1D

Rimu.Hamiltonians.BoseHubbardReal1D2C — Type

BoseHubbardReal1D2C(address::BoseFS2C; ua=1.0, ub=1.0, ta=1.0, tb=1.0, v=1.0)

Implements a two-component one-dimensional Bose Hubbard chain in real space.

\[\hat{H} = \hat{H}_a + \hat{H}_b + V\sum_{i} n_{a_i}n_{b_i}\]

Arguments

address: the starting address, defines number of particles and sites.

ua: the on-site interaction parameter parameter for Hamiltonian a.

ub: the on-site interaction parameter parameter for Hamiltonian b.

ta: the hopping strength for Hamiltonian a.

tb: the hopping strength for Hamiltonian b.

v: the inter-species interaction parameter V.

See also

HubbardReal1D

BoseHubbardMom1D2C

Rimu.Hamiltonians.BoseHubbardMom1D2C — Type

BoseHubbardMom1D2C(add::BoseFS2C; ua=1.0, ub=1.0, ta=1.0, tb=1.0, v=1.0, kwargs...)

Implements a one-dimensional Bose Hubbard chain in momentum space with a two-component Bose gas.

\[\hat{H} = \hat{H}_a + \hat{H}_b + \frac{V}{M}\sum_{kpqr} b^†_{r} a^†_{q} b_p a_k δ_{r+q,p+k}\]

Arguments

add: the starting address.

ua: the u parameter for Hamiltonian a.

ub: the u parameter for Hamiltonian b.

ta: the t parameter for Hamiltonian a.

tb: the t parameter for Hamiltonian b.

v: the inter-species interaction parameter V.

Further keyword arguments are passed on to the constructor of HubbardMom1D.

See also

BoseFS2C

BoseHubbardReal1D2C

Rimu.Hamiltonians.HubbardRealSpace — Type

HubbardRealSpace(address; u=ones(C, C), t=ones(C), geometry=PeriodicBoundaries(M,))

Hubbard model in real space. Supports single or multi-component Fock state addresses (with C components) and various (rectangular) lattice geometries in arbitrary dimensions.

\[ \hat{H} = -\sum_{\langle i,j\rangle,σ} t_σ a^†_{iσ} a_{jσ} + \frac{1}{2}\sum_{i,σ} u_{σ,σ} n_{iσ} (n_{iσ} - 1) + \frac{1}{2}\sum_{i,σ≠τ}u_{σ,τ} n_{iσ} n_{iτ}\]

Address types

BoseFS: Single-component Bose-Hubbard model.

FermiFS: Single-component Fermi-Hubbard model. This address only provides a single species of (non-interacting) fermions. You probably want to use CompositeFS.

CompositeFS: For multi-component models.

Geometries

PeriodicBoundaries

HardwallBoundaries

LadderBoundaries

Other parameters

u: the on-site interaction parameters. Must be a symmetric matrix. u[i, j] corresponds to the interaction between the i-th and j-th component. u[i, i] corresponds to the interaction of a component with itself. Note that u[i,i] must be zero for fermionic components.

t: the hopping strengths. Must be a vector of length C. The i-th element of the vector corresponds to the hopping strength of the i-th component.

Rimu.Hamiltonians.MatrixHamiltonian — Type

MatrixHamiltonian(
    mat::AbstractMatrix{T};
    starting_address::Int = starting_address(mat)
) <: AbstractHamiltonian{T}

Wrap an abstract matrix mat as an AbstractHamiltonian object for use with regular Vectors indexed by integers. Works with stochastic methods of lomc!(). Optionally, a starting_address can be provided.

Specialised methods are implemented for sparse matrices of type AbstractSparseMatrixCSC.

Hamiltonians interface

Behind the implementation of a particular model is a more abstract interface for defining hamiltonians. If you want to define a new model you should make use of this interface. The most general form of a model Hamiltonian should subtype to AbstractHamiltonian and implement the relevant methods.

Rimu.Interfaces.AbstractHamiltonian — Type

AbstractHamiltonian{T}

Supertype that provides an interface for linear operators over a linear space with scalar type T that are suitable for FCIQMC (with lomc!). Indexing is done with addresses (typically not integers) from an address space that may be large (and will not need to be completely generated).

AbstractHamiltonian instances operate on vectors of type AbstractDVec from the module DictVectors and work well with addresses of type AbstractFockAddress from the module BitStringAddresses. The type works well with the external package KrylovKit.jl.

For available implementations see Hamiltonians.

Interface

Methods that need to be implemented:

num_offdiagonals(::AbstractHamiltonian, address)

get_offdiagonal(::AbstractHamiltonian, address, chosen::Integer)

diagonal_element(::AbstractHamiltonian, address)

starting_address(::AbstractHamiltonian)

Optional methods to implement:

LOStructure(::Type{typeof(lo)}): defaults to AdjointUnknown

dimension(::Type{T}, ::AbstractHamiltonian): defaults to dimension of address space

momentum(::AbstractHamiltonian): no default

Provides:

offdiagonals: iterator over reachable off-diagonal matrix elements

random_offdiagonal: function to generate random off-diagonal matrix element

*(H, v): deterministic matrix-vector multiply (allocating)

H(v): equivalent to H * v.

mul!(w, H, v): mutating matrix-vector multiply.

dot(x, H, v): compute x⋅(H*v) minimizing allocations.

H[address1, address2]: indexing with getindex() - mostly for testing purposes (slow!)

BasisSetRep: construct a basis set repesentation

sparse, Matrix: construct a (sparse) matrix representation

See also Hamiltonians, Interfaces.

Rimu.Interfaces.offdiagonals — Function

offdiagonals(h::AbstractHamiltonian, address)

Return an iterator over reachable off-diagonal matrix elements of type <:AbstractOffdiagonals. Defaults to returning Offdiagonals(h, a)

See also

Offdiagonals

AbstractOffdiagonals

julia> addr = BoseFS((3,2,1));

julia> H = HubbardReal1D(addr);

julia> h = offdiagonals(H, addr)
6-element Rimu.Hamiltonians.Offdiagonals{BoseFS{6, 3, BitString{8, 1, UInt8}}, Float64, HubbardReal1D{Float64, BoseFS{6, 3, BitString{8, 1, UInt8}}, 1.0, 1.0}}:
 (BoseFS{6,3}((2, 3, 1)), -3.0)
 (BoseFS{6,3}((2, 2, 2)), -2.449489742783178)
 (BoseFS{6,3}((3, 1, 2)), -2.0)
 (BoseFS{6,3}((4, 1, 1)), -2.8284271247461903)
 (BoseFS{6,3}((4, 2, 0)), -2.0)
 (BoseFS{6,3}((3, 3, 0)), -1.7320508075688772)

Rimu.Interfaces.diagonal_element — Function

diagonal_element(g::G2Correlator, add::BoseFS2C{NA,NB,M,AA,AB})

The diagonal element in G2Correlator, where (p-q)=0, hence it becomes

\[\frac{1}{M}\sum_{k,p=1}^M a^†_{k} b^†_{p} b_p a_k .\]

See also

G2Correlator

diagonal_element(ham, add)

Compute the diagonal matrix element of the linear operator ham at address add.

Example

julia> addr = BoseFS((3, 2, 1));

julia> H = HubbardMom1D(addr);

julia> diagonal_element(H, addr)
8.666666666666664

Rimu.Interfaces.starting_address — Function

starting_address(h)

Return the starting address for Hamiltonian h. Part of the AbstractHamiltonian interface. When called on an AbstractMatrix return the index of the lowest diagonal element.

Example

julia> addr = BoseFS((3, 2, 1));

julia> H = HubbardMom1D(addr);

julia> addr == starting_address(H)
true

The following functions may be implemented instead of offdiagonals.

Rimu.Interfaces.num_offdiagonals — Function

num_offdiagonals(ham, add)

Compute the number of number of reachable configurations from address add.

Example

julia> addr = BoseFS((3, 2, 1));

julia> H = HubbardMom1D(addr);

julia> num_offdiagonals(H, addr)
10

Rimu.Interfaces.get_offdiagonal — Function

newadd, me = get_offdiagonal(ham, add, chosen)

Compute value me and new address newadd of a single (off-diagonal) matrix element in a Hamiltonian ham. The off-diagonal element is in the same column as address add and is indexed by integer index chosen.

Example

julia> addr = BoseFS((3, 2, 1));

julia> H = HubbardMom1D(addr);

julia> get_offdiagonal(H, addr, 3)
(BoseFS{6,3}((2, 1, 3)), 1.0)

The following functions come with default implementations, but may be customized.

Rimu.Interfaces.random_offdiagonal — Function

random_offdiagonal(offdiagonals::AbstractOffdiagonals)
random_offdiagonal(ham::AbstractHamiltonian, add)

Generate a single random excitation, i.e. choose from one of the accessible off-diagonal elements in the column corresponding to address add of the Hamiltonian matrix represented by ham. Alternatively, pass as argument an iterator over the accessible matrix elements.

Rimu.Interfaces.LOStructure — Type

LOStructure(op::AbstractHamiltonian)
LOStructure(typeof(op))

LOStructure speficies properties of the linear operator op. If a special structure is known this can speed up calculations. Implemented structures are:

IsHermitian: The operator is complex and Hermitian or real and symmetric.

AdjointKnown: The operator is not Hermitian, but its adjoint is implemented.

AdjointUnknown: adjoint for this operator is not implemented.

In order to define this trait for a new linear operator type, define a method for LOStructure(::Type{<:MyNewLOType}) = ….

Rimu.Hamiltonians.dimension — Function

dimension(::Type{T}, h)

Return the dimension of Hilbert space as T. If the result does not fit into T, return nothing. If T<:AbstractFloat, an approximate value computed with the improved Stirling formula may be returned instead.

Examples

julia> dimension(HubbardMom1D(BoseFS((1,2,3))))
28
julia> dimension(HubbardMom1D(near_uniform(BoseFS{200,100})))

julia> dimension(Float64, HubbardMom1D(near_uniform(BoseFS{200,100})))
1.3862737677578234e81
julia> dimension(BigInt, HubbardMom1D(near_uniform(BoseFS{200,100})))
1386083821086188248261127842108801860093488668581216236221011219101585442774669540

Geometry

Rimu.Hamiltonians.LatticeGeometry — Type

abstract type LatticeGeometry

A LatticeGeometry controls which sites in an AbstractFockAddress are considered to be neighbours.

Currently only supported by HubbardRealSpace.

Available implementations

PeriodicBoundaries

HardwallBoundaries

LadderBoundaries

Interface to implement

Base.size: return the lattice size.

neighbour_site(::LatticeGeometry, ::Int, ::Int)

num_neighbours(::LatticeGeometry)

Rimu.Hamiltonians.PeriodicBoundaries — Type

PeriodicBoundaries(size...) <: LatticeGeometry

Rectangular lattice with periodic boundary conditions of size size.

The dimension of the lattice is controlled by the number of arguments given to its constructor.

This is the default geometry used by HubbardRealSpace.

Example

julia> lattice = PeriodicBoundaries(5, 4) # 2D lattice of size 5 × 4
PeriodicBoundaries(5, 4)

julia> num_neighbours(lattice)
4

julia> neighbour_site(lattice, 1, 1)
2

julia> neighbour_site(lattice, 1, 2)
5

julia> neighbour_site(lattice, 1, 3)
6

julia> neighbour_site(lattice, 1, 4)
16

See also

LatticeGeometry

HardwallBoundaries

LadderBoundaries

num_neighbours

neighbour_site

Rimu.Hamiltonians.HardwallBoundaries — Type

HardwallBoundaries

Rectangular lattice with hard wall boundary conditions of size size. neighbour_site() will return 0 for some neighbours of boundary sites.

The dimension of the lattice is controlled by the number of arguments given to its constructor.

Example

julia> lattice = HardwallBoundaries(5) # 1D lattice of size 5
HardwallBoundaries(5)

julia> neighbour_site(lattice, 1, 1)
2

julia> neighbour_site(lattice, 1, 2)
0

See also

LatticeGeometry

PeriodicBoundaries

LadderBoundaries

Rimu.Hamiltonians.LadderBoundaries — Type

LadderBoundaries(size...; subgeometry=PeriodicBoundaries) <: LatticeGeometry

Lattice geometry where the first dimension is of size 2 and has hardwall boundary conditions. Using this geometry is more efficient than using HardwallBoundaries with a size of 2, as it does not generate rejected neighbours.

In other dimensions, it behaves like its subgeometry, which can be any LatticeGeometry.

Example

julia> lattice = LadderBoundaries(2, 3, 4) # 3D lattice of size 2 × 3 × 4
LadderBoundaries(2, 3, 4)

julia> num_neighbours(lattice)
5

julia> neighbour_site(lattice, 1, 1)
2

julia> neighbour_site(lattice, 1, 2)
3

julia> neighbour_site(lattice, 1, 3)
5

julia> neighbour_site(lattice, 1, 4)
7

julia> neighbour_site(lattice, 1, 5)
19

See also

LatticeGeometry

PeriodicBoundaries

HardwallBoundaries

Rimu.Hamiltonians.num_neighbours — Function

num_neighbours(geom::LatticeGeometry)

Return the number of neighbours each lattice site has in this geometry.

Note that for efficiency reasons, all sites are expected to have the same number of neighbours. If some of the neighbours are invalid, this is handled by having neighbour_site return 0.

Rimu.Hamiltonians.neighbour_site — Function

neighbour_site(geom::LatticeGeometry, site, i)

Find the i-th neighbour of site in the geometry. If the move is illegal, return 0.