Developer Documentation

Module Hamiltionians.jl

This module contains definitions of Hamiltonians, in particular specific physical models of interest. These are organised by means of an interface around the abstract type AbstractHamiltonian, in the spirit of the AbstractArray interface as discussed in the Julia Documentation.

Rimu.Hamiltonians — Module

module Hamiltonians

This module defines Hamiltonian types, interfaces, and functions for working with Hamiltonians.

Relation to other parts of the Rimu code

In order to define a specific model Hamiltonian with relevant parameters for the model, instantiate the model like this in the input file:

ham = BoseHubbardReal1D(n=6, m=6, u=1.0, t=1.0, AT = BSAdd64)

In the rest of the Rimu code, access to properties and matrix elements of the model are then provided by the following methods:

ham[address1, address2]: indexing of matrix elements (slow - use with caution)

ham(dv::AbstractDVec) or mul!(dv1, ham, dv2): use as linear operator

diagME(ham, add): diagonal matrix element

numOfHops(ham, add): number of off-diagonals

hop(ham, add, chosen): access off-diagonal matrix element

Hops(ham, add): iterator over off-diagonal matrix elements

generateRandHop(hops::Hops): choose random off-diagonal

bit_String_Length(ham): number of bits in the configuration

ham(:dim) and ham(:fdim): dimension of linear space. See hasIntDimension(ham), dimensionLO(ham), fDimensionLO(ham)

nearUniform(ham): configuration with particles spread across modes

Model Hamiltonians

Here is a list of fully implemented model Hamiltonians. So far there are two variants implemented of the one-dimentional Bose-Hubbard model real space as well as a momentum-space Hubbard chain.

Rimu.Hamiltonians.BoseHubbardReal1D — Type

ham = BoseHubbardReal1D(;[n=6, m=6, u=1.0, t=1.0, AT = BSAdd64])

Implements a one-dimensional Bose Hubbard chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1)\]

Warning

This Hamiltonian is deprecated. Please use HubbardReal1D instead.

Arguments

n::Int: the number of bosons

m::Int: the number of lattice sites

u::Float64: the interaction parameter

t::Float64: the hopping strength

AT::Type: the address type

Functor use:

w = ham(v)
ham(w, v)

Compute the matrix - vector product w = ham * v. The two-argument version is mutating for w.

Rimu.Hamiltonians.ExtendedBHReal1D — Type

ham = ExtendedBHReal1D(n=6, m=6, u=1.0, v=1.0, t=1.0, AT=BSAdd64)

Implements the extended Bose Hubbard model on a one-dimensional chain in real space.

\[\hat{H} = -t \sum_{\langle i,j\rangle} a_i^† a_j + \frac{u}{2}\sum_i n_i (n_i-1) + v \sum_{\langle i,j\rangle} n_i n_j\]

Arguments

n::Int: number of bosons

m::Int: number of lattice sites

u::Float64: on-site interaction parameter

v::Float64: the next-neighbor interaction

t::Float64: the hopping strength

AT::Type: address type for identifying configuration

Rimu.Hamiltonians.BoseHubbardMom1D — Type

ham = BoseHubbardMom1D(;[n=6, m=6, u=1.0, t=1.0], add = add)
ham = BoseHubbardMom1D(add; u=1.0, t=1.0)

Implements a one-dimensional Bose Hubbard chain in momentum space.

\[\hat{H} = \sum_{k} ϵ_k n_k + \frac{u}{M}\sum_{kpqr} a^†_{r} a^†_{q} a_p a_k δ_{r+q,p+k}\\ ϵ_k = -2t \cos(k)\]

Warning

This Hamiltonian is deprecated. Please use HubbardMom1D instead.

Arguments

n::Int: the number of bosons

m::Int: the number of lattice sites

u::Float64: the interaction parameter

t::Float64: the hopping strength

AT::Type: the address type

Hamiltonians interface

Behind the implementation of a particular model is a more abstract interface for defining hamiltonians. If you want to define a new model you should make use of this interface. The most general form of a model Hamiltonian should subtype to AbstractHamiltonian and implement the relevant methods.

Rimu.Hamiltonians.AbstractHamiltonian — Type

AbstractHamiltonian{T}

Supertype that provides an interface for linear operators over a linear space with scalar type T that are suitable for FCIQMC. Indexing is done with addresses (typically not integers) from an address space that may be large (and will not need to be completely generated).

AbstractHamiltonian instances operate on vectors of type AbstractDVec from the module DictVectors and work well with addresses of type AbstractFockAddress from the module BitStringAddresses. The type works well with the external package KrylovKit.jl.

Methods

Provides:

offdiagonals: iterator over reachable off-diagonal matrix elements

random_offdiagonal: function to generate random off-diagonal matrix element

dimension: get the dimension of the address space.

H[address1, address2]: indexing with getindex() - mostly for testing purposes

*(H, v): deterministic matrix-vector multiply.

H(v): equivalent to H * v.

mul!(w, H, v): mutating matrix-vector multiply.

dot(x, H, v): compute x⋅(H*v) minimizing allocations.

Methods that need to be implemented:

num_offdiagonals(::AbstractHamiltonian, address)

get_offdiagonal(::AbstractHamiltonian, address, chosen::Integer)

diagonal_element(::AbstractHamiltonian, address)

starting_address(::AbstractHamiltonian)

Optional methods to implement:

Hamiltonians.LOStructure(::Type{typeof(lo)})

dimension(::Type{T}, ::AbstractHamiltonian)

offdiagonals(::AbstractHamiltonian, ::AbstractFockAddress)

momentum(::AbstractHamiltonian)

Missing docstring.

Missing docstring for Hops. Check Documenter's build log for details.

Missing docstring.

Missing docstring for generateRandHop. Check Documenter's build log for details.

Core functions

The following functions are part of the core functionality of a Hamiltonian and need to be implemented efficiently and specifically for each model.

Missing docstring.

Missing docstring for numOfHops. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for diagME. Check Documenter's build log for details.

BosonicHamiltonian

For a many-body system consisting of spinless bosons, we already know more about the structure of the problem. BosonicHamiltonian is a pre-defined subtype of AbstractHamiltonian.

Missing docstring.

Missing docstring for BosonicHamiltonian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hasIntDimension. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dimensionLO. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fDimensionLO. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bit_String_Length. Check Documenter's build log for details.

Rimu.BitStringAddresses.nearUniform — Function

nearUniform(BoseFS{N,M})
nearUniform(BoseFS{N,M,S}) -> bfs::BoseFS{N,M,S}

Create bosonic Fock state with near uniform occupation number of M modes with a total of N particles. Specifying the bit address type S is optional.

Examples

julia> nearUniform(BoseFS{7,5,BitString{14}})
BoseFS((2,2,1,1,1))

julia> nearUniform(BoseFS{7,5})
BoseFS((2,2,1,1,1))