Module Interfaces

Rimu.Interfaces — Module

module Interfaces

This module contains interfaces that can be used to extend and modify the algorithms and behaviours of Rimu.

Interfaces

Follow the links for the definitions of the interfaces!

AbstractHamiltonian for defining Hamiltonians

AbstractDVec for defining data structures for Rimu as in DictVectors

StochasticStyle for controlling the stochastic algorithms used by lomc! as implemented in StochasticStyles

Additonal exports

Interface functions forAbstractHamiltonians:

diagonal_element

num_offdiagonals

get_offdiagonal

offdiagonals.

random_offdiagonal

starting_address

LOStructure

working with AbstractDVecs and StochasticStyle

deposit!

default_style

CompressionStrategy

Functions Rimu.jl uses to do FCIQMC:

fciqmc_col!

step_stats

update_dvec!