Module StochasticStyles

This module contains the implementations of StochasticStyles, which control how the stochastic matrix-vector multiplication is performed.

Rimu.Interfaces.StochasticStyle — Type

StochasticStyle(v)

Abstract type. When called as a function it returns the native style of the generalised vector v that determines how simulations are to proceed.

Implemented styles

IsStochasticInteger - integer walker FCIQMC

IsDeterministic - perform deterministic variant of power method

IsStochasticWithThreshold - floating point walker FCIQMC

IsDynamicSemistochastic - floating point walker FCIQMC with smarter spawning and vector compression

IsStochastic2Pop

IsExplosive

Usage

Concrete StochasticStyles can be used for the style keyword argument of lomc! and DVec.

Interface

When defining a new StochasticStyle, subtype it as MyStyle<:StochasticStyle{T} where T is the concrete value type the style is designed to work with.

For it to work with lomc!, a StochasticStyle must define the following:

fciqmc_col!(::StochasticStyle, w, H, address, value, shift, dτ)

step_stats(::StochasticStyle)

Optionally, it can also define update_dvec!, which can be used to perform arbitrary transformations on the generalised vector after the spawning step is complete.

Available StochasticStyles

Rimu.Interfaces.StyleUnknown — Type

StyleUnknown{T}() <: StochasticStyle

Trait for value types not (currently) compatible with FCIQMC. This style makes it possible to construct dict vectors with unsupported valtypes.

See also StochasticStyle.

Rimu.StochasticStyles.IsStochasticInteger — Type

IsStochasticInteger{T=Int}() <: StochasticStyle{T}

Trait for generalised vector of configurations indicating stochastic propagation as seen in the original FCIQMC algorithm.

See also StochasticStyle.

Rimu.StochasticStyles.IsDeterministic — Type

IsDeterministic{T=Float64}() <: StochasticStyle{T}

Trait for generalised vector of configuration indicating deterministic propagation of walkers. The optional compression argument can set a CompressionStrategy.

See also StochasticStyle.

Rimu.StochasticStyles.IsStochasticWithThreshold — Type

IsStochasticWithThreshold{T=Float64}(threshold=1.0) <: StochasticStyle{T}

Trait for generalised vector of configurations indicating stochastic propagation with real walker numbers and cutoff threshold.

During stochastic propagation, walker numbers small than threshold will be stochastically projected to either zero or threshold.

See also StochasticStyle.

Rimu.StochasticStyles.IsDynamicSemistochastic — Type

IsDynamicSemistochastic{T=Float64}(; kwargs...) <: StochasticStyle{T}

QMC propagation with non-integer walker numbers and reduced noise. All possible spawns are performed deterministically when number of walkers in a configuration is high. Unlike with IsStochasticInteger or IsStochasticWithThreshold, where spawns are projected on the fly, stochastic vector compression is applied after spawning and diagonal death steps.

Note: if you want IsDynamicSemistochastic to project spawns as they are being performed, set a threshold to spawning, and set compression to NoCompression.

Parameters:

rel_threshold = 1.0: If the walker number on a configuration times this threshold is greater than the number of offdiagonals, spawning is done deterministically. Should be set to 1 or more for best performance.

abs_threshold = Inf: If the walker number on a configuration is greater than this value, spawning is done deterministically. Can be set to e.g. abs_threshold = 0.1 * target_walkers.

proj_threshold = 1.0: Values below this number are stochastically projected to this value or zero. See also IsStochasticWithThreshold.

spawning = WithReplacement(): SpawningStrategy to use for the non-exact spawns.

compression = ThresholdCompression(proj_threshold): CompressionStartegy used to compress the vector after a step. Overrides proj_threshold.

See also StochasticStyle.

Rimu.StochasticStyles.IsStochastic2Pop — Type

IsStochastic2Pop{T=Complex{Int}}() <: StochasticStyle{T}

Trait for generalised vector of configurations indicating stochastic propagation with complex walker numbers representing two populations of integer walkers.

When using this style, make sure to set a complex target number walkers in the ShiftStrategy!

This style is experimental and unexported.

See also StochasticStyle.

Rimu.StochasticStyles.IsExplosive — Type

IsExplosive{T=Float64}(; splatter_factor, explosion_threshold, compression) <: StochasticStyle{T}

QMC propagation with explosive walkers. Walkers with small walker numbers do not perform the standard death/spawning steps. Instead, a walker will either die completely and spawn with a greater magnitude (refered to as explosion below), or stay unchanged and not spawn. The probabilty of exploding is controlled by the shift and dτ.

Walkers with high walker numbers spawn as if IsDynamicSemistochastic was used.

Like IsDynamicSemistochastic, the vector is compressed after all spawning is performed.

This style is experimental and unexported.

Parameters

splatter_factor = 1.0: The spawning strength to use with exploded walkers.

explosion_threshold = 1.0: Entries smaller or equal than this value will attempt to explode.

proj_threshold = 1.0: Threshold to use in vector compression.

compression = ThresholdCompression(proj_threshold): CompressionStrategy to use to compress the vector. Overrides proj_threshold.

See also StochasticStyle.

The StochasticStyle interface

Rimu.Interfaces.step_stats — Function

step_stats(::StochasticStyle)

Return a tuple of names (Symbol or String) and a tuple of zeros of values of the same length. These will be reported as columns in the DataFrame returned by lomc!.

Rimu.Interfaces.fciqmc_col! — Function

fciqmc_col!(w, ham, add, num, shift, dτ)
fciqmc_col!(::StochasticStyle, args...)

Spawning and diagonal step of FCIQMC for single column of ham. In essence it computes

w .+= (1 .+ dτ.*(shift .- ham[:,add])).*num.

The StochasticStyle(w), picks the algorithm used.

Rimu.Interfaces.update_dvec! — Function

update_dvec!([::StochasticStyle,] dvec) -> dvec, nt

Perform an arbitrary transformation on dvec after the spawning step is completed and report statistics to the DataFrame.

Returns the new dvec and a NamedTuple nt of statistics to be reported.

When extending this function for a custom StochasticStyle, define a method for the two-argument call signature!

The default implementation uses CompressionStrategy to compress the vector.

Note: update_dvec! may return a new vector.

Rimu.Interfaces.CompressionStrategy — Type

CompressionStrategy

The CompressionStrategy controls how a vector is compressed after a step. To use, define CompressionStrategy(::StochasticStyle). The default implementation returns NoCompression.

Rimu.Interfaces.compress! — Function

compress!(::CompressionStrategy, v)

Compress the vector v and return it.

Rimu.Interfaces.default_style — Function

default_style(::Type)

Pick a StochasticStyle based on the value type. Throws an error if no known default style is known.

Spawning strategies and convenience functions

The following functions and types are unexported, but are useful when defining new styles.

Rimu.StochasticStyles.diagonal_step! — Function

diagonal_step!(w, ham, add, val, dτ, shift, threshold=0, report_stats=false)

Perform diagonal step on a walker add => val. Optional argument threshold sets the projection threshold. If eltype(w) is an Integer, the val is rounded to the nearest integer stochastically.

Rimu.StochasticStyles.projected_deposit! — Function

projected_deposit!(w, add, val, parent, threshold=0)

Like deposit!, but performs threshold projection before spawning. If eltype(w) is an Integer, values are stochastically rounded.

Returns the value deposited and the number of annihilations.

Rimu.StochasticStyles.SpawningStrategy — Type

SpawningStrategy

A SpawningStrategy is used to control how (offdiagonal) spawns are peformed. It is used by passing it as the first argument to spawn!.

The following methods are implemented:

Exact

SingleSpawn

WithReplacement

WithoutReplacement

Bernoulli

DynamicSemistochastic

Rimu.StochasticStyles.Exact — Type

Exact(threshold=0.0) <: SpawningStrategy

Perform an exact spawning step.

Parameters

threshold sets the projection threshold. If set to zero, no projection is performed.

spawn! with this strategy returns the number of spawns and annihilations.

Rimu.StochasticStyles.SingleSpawn — Type

SingleSpawn(threshold=0.0) <: SpawningStrategy

Perform a single spawn. Useful as a building block for other stochastic styles.

Parameters

threshold sets the projection threshold. If set to zero, no projection is performed.

spawn! with this strategy returns the number of spawns and annihilations.

Rimu.StochasticStyles.WithReplacement — Type

WithReplacement(threshold=0.0, strength=1.0) <: SpawningStrategy

SpawningStrategy where spawn targets are sampled with replacement. This is the default spawning strategy for most of the StochasticStyles.

Parameters

threshold sets the projection threshold. If set to zero, no projection is performed.

strength sets the number of spawns to perform, e.g. if val=5 and strength=2, 10 spawns will be performed.

spawn! with this strategy returns the number of spawns and annihilations.

Rimu.StochasticStyles.WithoutReplacement — Type

WithoutReplacement(threshold=0.0, strength=1.0) <: SpawningStrategy

SpawningStrategy where spawn targets are sampled without replacement.

Only to be used with DynamicSemistochastic.

Parameters

threshold sets the projection threshold. If set to zero, no projection is performed.

strength sets the number of spawns to perform, e.g. if val=5 and strength=2, 10 spawns will be performed.

spawn! with this strategy returns the number of spawns and annihilations.

Rimu.StochasticStyles.Bernoulli — Type

Bernoulli(threshold=0.0, strength=1.0) <: SpawningStrategy

Perform Bernoulli sampling. A spawn is attempted on each offdiagonal element with a probability that results in an expected number of spawns equal to the number of walkers on the spawning configuration.

Only to be used with DynamicSemistochastic.

Parameters

threshold sets the projection threshold.

strength sets the number of spawns to perform, e.g. if val=5 and strength=2, 10 spawns will be performed on average.

spawn! with this strategy returns the number of spawns and annihilations.

Rimu.StochasticStyles.DynamicSemistochastic — Type

DynamicSemistochastic(; start, rel_threshold, abs_threshold) <: SpawningStrategy

SpawningStrategy that behaves like strat when the number of walkers is low, but performs exact steps when it is high. What "high" means is controlled by the two thresholds described below.

Parameters

strat = WithReplacement(): a SpawningStrategy to use when the multiplication is not performed exactly. If the strat has a threshold different from zero, all spawns will be projected to that threshold.

rel_threshold = 1.0: When deciding on whether to perform an exact spawn, this value is multiplied to the number of walkers. Should be set to 1 or more for best performance. This threshold is affected by strat.strength.

abs_threshold = Inf: When deciding on whether to perform an exact spawn, min(abs_threshold, num_offdiagonals) is used. This threshold is affected by strat.strength.

See e.g. WithoutReplacement for a description of the strat.threshold and strat.strength parameters.

spawn! with this strategy returns the numbers of exact and inexact spawns, and the number of spawns and annihilations.

Rimu.StochasticStyles.spawn! — Function

spawn!(s::SpawningStrategy, w, ham::AbstractHamiltonian, add, val, dτ)
spawn!(s::SpawningStrategy, w, offdiags::AbstractOffdiagonals, add, val, dτ)

Perform stochastic spawns to w from address add with val walkers. dτ is a factor multiplied to every spawns, while val also controls the number of spawns performed.

This function should be overloaded in the second form, with offdiags as an argument.